ACF {nlme}	R Documentation

Autocorrelation Function

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include: gls and lme.

Usage

ACF(object, maxLag, ...)

Arguments

object
any object from which an autocorrelation function can be obtained. Generally an object resulting from a model fit, from which residuals can be extracted.

maxLag
maximum lag for which the autocorrelation should be calculated.

...
some methods for this generic require additional arguments.

Value

will depend on the method function used; see the appropriate documentation.

Author(s)

José Pinheiro and Douglas Bates Bates@stat.wisc.edu

References

Box, G.E.P., Jenkins, G.M., and Reinsel G.C. (1994) "Time Series Analysis: Forecasting and Control", 3rd Edition, Holden-Day.

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

ACF.gls, ACF.lme, plot.ACF

Examples

## see the method function documentation

[Package nlme version 3.1-121 Index]