BIC.logLik {nlme}	R Documentation

BIC of a logLik Object

Description

This function calculates the Bayesian information criterion, also known as Schwarz's Bayesian criterion (SBC) for an object inheriting from class logLik, according to the formula log-likelihood + npar*log(nobs), where npar represents the number of parameters and nobs the number of observations in the fitted model. When comparing fitted objects, the smaller the BIC, the better the fit.

Usage

## S3 method for class 'logLik':
BIC(object, ...)

Arguments

object 	 an object inheriting from class logLik, usually resulting from applying a logLik method to a fitted model object.

... 	 some methods for this generic use optional arguments. None are used in this method.

Value

a numeric value with the corresponding BIC.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

References

Schwarz, G. (1978) "Estimating the Dimension of a Model", Annals of Statistics, 6, 461-464.

See Also

BIC, logLik, AIC.

Examples

fm1 <- lm(distance ~ age, data = Orthodont) 
BIC(logLik(fm1))

[Package nlme version 3.1-93 Index]