Covariate {nlme}	R Documentation

Assign Covariate Values

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include all "varFunc" classes.

Usage


covariate(object) <- value

Arguments

object
any object with a covariate component.

value
a value to be assigned to the covariate associated with object.

Value

will depend on the method function; see the appropriate documentation.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

getCovariate

Examples


## see the method function documentation

[Package nlme version 3.1-137 Index]