Dim.pdMat {nlme}	R Documentation

Dimensions of a pdMat Object

Description

This method function returns the dimensions of the matrix represented by object.

Usage

## S3 method for class 'pdMat'
Dim(object, ...)

Arguments

object
an object inheriting from class "pdMat", representing a positive-definite matrix.

...
some methods for this generic require additional arguments. None are used in this method.

Value

an integer vector with the number of rows and columns of the matrix represented by object.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

Dim

Examples

Dim(pdSymm(diag(3)))

[Package nlme version 3.1-121 Index]