Initialize.corStruct {nlme}	R Documentation

Initialize corStruct Object

Description

This method initializes object by evaluating its associated covariate(s) and grouping factor, if any is present, in data, calculating various dimensions and constants used by optimization algorithms involving corStruct objects (see the appropriate Dim method documentation), and assigning initial values for the coefficients in object, if none were present.

Usage

## S3 method for class 'corStruct'
Initialize(object, data, ...)

Arguments

object
an object inheriting from class "corStruct" representing a correlation structure.

data
a data frame in which to evaluate the variables defined in formula(object).

...
this argument is included to make this method compatible with the generic.

Value

an initialized object with the same class as object representing a correlation structure.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

Dim.corStruct

Examples

cs1 <- corAR1(form = ~ 1 | Subject)
cs1 <- Initialize(cs1, data = Orthodont)

[Package nlme version 3.1-127 Index]