Initialize.glsStruct {nlme}	R Documentation

Initialize a glsStruct Object

Description

The individual linear model components of the glsStruct list are initialized.

Usage

## S3 method for class 'glsStruct'
Initialize(object, data, control, ...)

Arguments

object
an object inheriting from class "glsStruct", representing a list of linear model components, such as corStruct and varFunc objects.

data
a data frame in which to evaluate the variables defined in formula(object).

control
an optional list with control parameters for the initialization and optimization algorithms used in gls. Defaults to list(singular.ok = FALSE), implying that linear dependencies are not allowed in the model.

...
some methods for this generic require additional arguments. None are used in this method.

Value

a glsStruct object similar to object, but with initialized model components.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

gls, Initialize.corStruct, Initialize.varFunc, Initialize

[Package nlme version 3.1-127 Index]