Initialize.lmeStruct {nlme}	R Documentation

Initialize an lmeStruct Object

Description

The individual linear mixed-effects model components of the lmeStruct list are initialized.

Usage

## S3 method for class 'lmeStruct'
Initialize(object, data, groups, conLin, control, ...)

Arguments

object
an object inheriting from class "lmeStruct", representing a list of linear mixed-effects model components, such as reStruct, corStruct, and varFunc objects.

data
a data frame in which to evaluate the variables defined in formula(object).

groups
a data frame with the grouping factors corresponding to the lme model associated with object as columns, sorted from innermost to outermost grouping level.

conLin
an optional condensed linear model object, consisting of a list with components "Xy", corresponding to a regression matrix (X) combined with a response vector (y), and "logLik", corresponding to the log-likelihood of the underlying lme model. Defaults to attr(object, "conLin").

control
an optional list with control parameters for the initialization and optimization algorithms used in lme. Defaults to list(niterEM=20, gradHess=TRUE), implying that 20 EM iterations are to be used in the derivation of initial estimates for the coefficients of the reStruct component of object and, if possible, numerical gradient vectors and Hessian matrices for the log-likelihood function are to be used in the optimization algorithm.

...
some methods for this generic require additional arguments. None are used in this method.

Value

an lmeStruct object similar to object, but with initialized model components.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

lme, Initialize.reStruct, Initialize.corStruct, Initialize.varFunc, Initialize

[Package nlme version 3.1-121 Index]