Generate system matrix for LDEs

Generate the system matrix for the linear differential equations determined by a compartment model.

Usage

LDEsysMat(pars, incidence)

Arguments

pars

a numeric vector of parameter values.

incidence

an integer matrix with columns named From, To, and Par. Values in the Par column must be in the range 1 to length(pars). Values in the From column must be between 1 and the number of compartments. Values in the To column must be between 0 and the number of compartments.

Details

A compartment model describes material transfer between k in a system of k compartments to a linear system of differential equations. Given a description of the system and a vector of parameter values this function returns the system matrix.

This function is intended for use in a general system for solving compartment models, as described in Bates and Watts (1988).

Value

A k by k numeric matrix.

References

Bates, D. M. and Watts, D. G. (1988), Nonlinear Regression Analysis and Its Applications, Wiley, New York.

Author

Douglas Bates bates@stat.wisc.edu

Examples

# incidence matrix for a two compartment open system
incidence <-
  matrix(c(1,1,2,2,2,1,3,2,0), ncol = 3, byrow = TRUE,
   dimnames = list(NULL, c("Par", "From", "To")))
incidence
#>      Par From To
#> [1,]   1    1  2
#> [2,]   2    2  1
#> [3,]   3    2  0
LDEsysMat(c(1.2, 0.3, 0.4), incidence)
#>      [,1] [,2]
#> [1,] -1.2  0.3
#> [2,]  1.2 -0.7