Mathematics achievement scores

The MathAchieve data frame has 7185 rows and 6 columns.

Format

This data frame contains the following columns:

School

an ordered factor identifying the school that the student attends

Minority

a factor with levels No Yes indicating if the student is a member of a minority racial group.

Sex

a factor with levels Male Female

SES

a numeric vector of socio-economic status.

MathAch

a numeric vector of mathematics achievement scores.

MEANSES

a numeric vector of the mean SES for the school.

Details

Each row in this data frame contains the data for one student.

Examples

summary(MathAchieve)
#>      School     Minority       Sex            SES               MathAch      
#>  2305   :  67   No :5211   Male  :3390   Min.   :-3.758000   Min.   :-2.832  
#>  5619   :  66   Yes:1974   Female:3795   1st Qu.:-0.538000   1st Qu.: 7.275  
#>  4292   :  65                            Median : 0.002000   Median :13.131  
#>  8857   :  64                            Mean   : 0.000143   Mean   :12.748  
#>  4042   :  64                            3rd Qu.: 0.602000   3rd Qu.:18.317  
#>  3610   :  64                            Max.   : 2.692000   Max.   :24.993  
#>  (Other):6795                                                                
#>     MEANSES         
#>  Min.   :-1.188000  
#>  1st Qu.:-0.317000  
#>  Median : 0.038000  
#>  Mean   : 0.006138  
#>  3rd Qu.: 0.333000  
#>  Max.   : 0.831000  
#>