Matrix {nlme}	R Documentation

Assign Matrix Values

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include pdMat, pdBlocked, and reStruct.

Usage

matrix(object) <- value

Arguments

object
any object to which as.matrix can be applied.

value
a matrix, or list of matrices, with the same dimensions as as.matrix(object) with the new values to be assigned to the matrix associated with object.

Value

will depend on the method function; see the appropriate documentation.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

as.matrix, also for examples, matrix<-.pdMat, matrix<-.reStruct.

[Package nlme version 3.1-127 Index]