Matrix.reStruct {nlme}	R Documentation

Assign reStruct Matrices

Description

The individual matrices in value are assigned to each pdMat component of object, in the order they are listed. The new matrices must have the same dimensions as the matrices they are meant to replace.

Usage

## S3 replacement method for class 'reStruct'
matrix(object) <-  value

Arguments

object
an object inheriting from class "reStruct", representing a random effects structure and consisting of a list of pdMat objects.

value
a matrix, or list of matrices, with the new values to be assigned to the matrices associated with the pdMat components of object.

Value

an reStruct object similar to object, but with the coefficients of the individual pdMat components modified to produce the matrices listed in value.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

reStruct, pdMat, "matrix<-"

Examples

rs1 <- reStruct(list(Dog = ~day, Side = ~1), data = Pixel)
matrix(rs1) <- list(diag(2), 3)

[Package nlme version 3.1-127 Index]