Muscle {nlme}	R Documentation

Contraction of heart muscle sections

Description

The Muscle data frame has 60 rows and 3 columns.

Format

This data frame contains the following columns:

Strip

an ordered factor indicating the strip of muscle being measured.

conc

a numeric vector giving the concentration of CaCl2

length

a numeric vector giving the shortening of the heart muscle strip.

Details

Baumann and Waldvogel (1963) describe data on the shortening of heart muscle strips dipped in a CaCl_2 solution. The muscle strips are taken from the left auricle of a rat's heart.

Source

Baumann, F. and Waldvogel, F. (1963), La restitution pastsystolique de la contraction de l'oreillette gauche du rat. Effets de divers ions et de l'acetylcholine, Helvetica Physiologica Acta, 21.

[Package nlme version 3.1-137 Index]