Names.pdMat {nlme}	R Documentation

Names of a pdMat Object

Description

This method function returns the fist element of the Dimnames attribute of object, which contains the column names of the matrix represented by object.

Usage

## S3 method for class 'pdMat'
Names(object, ...)
## S3 replacement method for class 'pdMat'
Names(object, ...) <- value

Arguments

object
an object inheriting from class "pdMat", representing a positive-definite matrix.

value
a character vector with the replacement values for the column and row names of the matrix represented by object. It must have length equal to the dimension of the matrix represented by object and, if names have been previously assigned to object, it must correspond to a permutation of the original names.

...
some methods for this generic require additional arguments. None are used in this method.

Value

if object has a Dimnames attribute then the first element of this attribute is returned; otherwise NULL.

SIDE EFFECTS

On the left side of an assignment, sets the Dimnames attribute of object to list(value, value).

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

Names, Names.pdBlocked

Examples

pd1 <- pdSymm(~age, data = Orthodont)
Names(pd1)

[Package nlme version 3.1-121 Index]