X-ray pixel intensities over time

The Pixel data frame has 102 rows and 4 columns of data on the pixel intensities of CT scans of dogs over time

Format

This data frame contains the following columns:

Dog

a factor with levels 1 to 10 designating the dog on which the scan was made

Side

a factor with levels L and R designating the side of the dog being scanned

day

a numeric vector giving the day post injection of the contrast on which the scan was made

pixel

a numeric vector of pixel intensities

Source

Pinheiro, J. C. and Bates, D. M. (2000) Mixed-effects Models in S and S-PLUS, Springer.

Examples

fm1 <- lme(pixel ~ day + I(day^2), data = Pixel,
           random = list(Dog = ~ day, Side = ~ 1))
summary(fm1)
#> Linear mixed-effects model fit by REML
#>   Data: Pixel 
#>        AIC      BIC    logLik
#>   841.2102 861.9712 -412.6051
#> 
#> Random effects:
#>  Formula: ~day | Dog
#>  Structure: General positive-definite, Log-Cholesky parametrization
#>             StdDev   Corr  
#> (Intercept) 28.36990 (Intr)
#> day          1.84375 -0.555
#> 
#>  Formula: ~1 | Side %in% Dog
#>         (Intercept) Residual
#> StdDev:    16.82431 8.989606
#> 
#> Fixed effects:  pixel ~ day + I(day^2) 
#>                 Value Std.Error DF   t-value p-value
#> (Intercept) 1073.3391 10.171686 80 105.52225       0
#> day            6.1296  0.879321 80   6.97083       0
#> I(day^2)      -0.3674  0.033945 80 -10.82179       0
#>  Correlation: 
#>          (Intr) day   
#> day      -0.517       
#> I(day^2)  0.186 -0.668
#> 
#> Standardized Within-Group Residuals:
#>        Min         Q1        Med         Q3        Max 
#> -2.8290572 -0.4491811  0.0255493  0.5572163  2.7519651 
#> 
#> Number of Observations: 102
#> Number of Groups: 
#>           Dog Side %in% Dog 
#>            10            20 
VarCorr(fm1)
#>             Variance       StdDev    Corr  
#> Dog =       pdLogChol(day)                 
#> (Intercept) 804.851443     28.369904 (Intr)
#> day           3.399416      1.843750 -0.555
#> Side =      pdLogChol(1)                   
#> (Intercept) 283.057248     16.824305       
#> Residual     80.813009      8.989606