Growth of soybean plants

The Soybean data frame has 412 rows and 5 columns.

Format

This data frame contains the following columns:

Plot

a factor giving a unique identifier for each plot.

Variety

a factor indicating the variety; Forrest (F) or Plant Introduction #416937 (P).

Year

a factor indicating the year the plot was planted.

Time

a numeric vector giving the time the sample was taken (days after planting).

weight

a numeric vector giving the average leaf weight per plant (g).

Details

These data are described in Davidian and Giltinan (1995, 1.1.3, p.7) as ``Data from an experiment to compare growth patterns of two genotypes of soybeans: Plant Introduction #416937 (P), an experimental strain, and Forrest (F), a commercial variety.''

Source

Pinheiro, J. C. and Bates, D. M. (2000), Mixed-Effects Models in S and S-PLUS, Springer, New York. (Appendix A.27)

Davidian, M. and Giltinan, D. M. (1995), Nonlinear Models for Repeated Measurement Data, Chapman and Hall, London.

Examples

summary(fm1 <- nlsList(SSlogis, data = Soybean))
#> Warning: 2 errors caught in nls(y ~ 1/(1 + exp((xmid - x)/scal)), data = xy, start = list(xmid = aux[[1L]], scal = aux[[2L]]), algorithm = "plinear", ...).  The error messages and their frequencies are
#> 
#>                      number of iterations exceeded maximum of 50 
#>                                                                1 
#> step factor 0.000488281 reduced below 'minFactor' of 0.000976562 
#>                                                                1 
#> Call:
#>   Model: weight ~ SSlogis(Time, Asym, xmid, scal) | Plot 
#>    Data: Soybean 
#> 
#> Coefficients:
#>    Asym 
#>          Estimate  Std. Error    t value     Pr(>|t|)
#> 1988F4  15.151550   0.8738054 17.3397311 4.463275e-07
#> 1988F2  19.745419   1.9433542 10.1604842 3.794022e-07
#> 1988F1  20.338407   2.2589870  9.0033309 6.127446e-08
#> 1988F7  19.871203   1.7791330 11.1690377 5.679788e-06
#> 1988F5  30.648921  11.2081166  2.7345291 3.736542e-03
#> 1988F8  22.777043   2.4318953  9.3659633 2.360279e-07
#> 1988F6  23.293137   2.7299845  8.5323331 7.581537e-09
#> 1988F3  23.697110   1.5879718 14.9228787 3.597063e-07
#> 1988P1  17.300492   0.8640908 20.0216145 1.013670e-08
#> 1988P5  17.703754   1.0257602 17.2591549 3.921123e-09
#> 1988P4  24.008911   2.9706377  8.0820730 3.017104e-03
#> 1988P8  28.249591   4.1988040  6.7280089 2.276684e-05
#> 1988P7  27.485995   3.3945416  8.0971155 1.342293e-08
#> 1988P3  24.939006   1.7455364 14.2873025 1.631779e-05
#> 1988P2  36.661159   6.3486895  5.7746027 2.477014e-02
#> 1988P6 163.702948 255.3305347  0.6411413 8.337208e-01
#> 1989F6   8.509308   1.3635637  6.2404918 5.729741e-05
#> 1989F5   9.669081   0.9877562  9.7889350 2.180771e-05
#> 1989F4  11.247524   1.0049236 11.1924170 7.984003e-08
#> 1989F1  11.251052   1.1113040 10.1241894 7.711603e-05
#> 1989F2  11.233333   0.9989299 11.2453664 6.475196e-09
#> 1989F7  10.071395   0.8452312 11.9155511 3.415718e-05
#> 1989F8  10.609519   0.8302693 12.7784076 1.450821e-05
#> 1989F3  18.419564   2.6808047  6.8709085 6.092551e-06
#> 1989P7  15.471897   0.7861292 19.6811114 8.869028e-07
#> 1989P4  18.177522   1.4204479 12.7970351 7.262489e-07
#> 1989P6  20.498830   1.9022691 10.7759888 5.129593e-06
#> 1989P5         NA          NA         NA           NA
#> 1989P1  21.683985   1.9183399 11.3035158 7.806236e-08
#> 1989P3  22.283750   1.1402473 19.5429100 4.478139e-07
#> 1989P2  28.296972   4.1482952  6.8213495 1.561728e-04
#> 1989P8         NA          NA         NA           NA
#> 1990F2  19.459102   6.1238316  3.1776024 1.175660e-03
#> 1990F3  19.867932   3.2357524  6.1401275 6.112667e-04
#> 1990F4  27.435518   9.4505666  2.9030554 2.326747e-03
#> 1990F5  18.719501   1.2575853 14.8852735 2.425467e-07
#> 1990F1  19.790735   1.9241287 10.2855567 3.680484e-04
#> 1990F8  20.290441   1.5158389 13.3856188 4.113934e-05
#> 1990F7  19.835326   1.3061649 15.1859283 1.585581e-06
#> 1990F6  21.197115   1.7368415 12.2044041 8.578174e-06
#> 1990P8  18.513513   1.4497108 12.7704868 3.942440e-04
#> 1990P7  19.160767   2.1203104  9.0367748 2.394577e-04
#> 1990P3  19.198083   1.3929487 13.7823336 9.374480e-05
#> 1990P1  18.448379   0.9806108 18.8131507 1.488721e-06
#> 1990P6  17.689700   1.0685907 16.5542331 1.302191e-04
#> 1990P5  19.544895   1.1900419 16.4237028 8.984246e-05
#> 1990P2  25.787243   3.8482487  6.7010333 1.186730e-05
#> 1990P4  26.128905   3.2551064  8.0270508 1.385877e-03
#>    xmid 
#>         Estimate Std. Error   t value     Pr(>|t|)
#> 1988F4  52.83408   1.332907 39.638245 1.445923e-09
#> 1988F2  56.57507   2.769461 20.428185 5.866605e-09
#> 1988F1  57.40242   3.006916 19.090129 3.277052e-10
#> 1988F7  56.16159   2.494661 22.512709 8.925168e-08
#> 1988F5  64.12999   9.123542  7.029067 4.563191e-05
#> 1988F8  59.33033   2.684914 22.097663 6.076841e-10
#> 1988F6  59.59762   3.078580 19.358801 2.491787e-11
#> 1988F3  56.42461   1.631701 34.580235 1.056373e-09
#> 1988P1  48.84945   1.347844 36.242652 1.617368e-10
#> 1988P5  51.27156   1.512099 33.907531 3.516646e-11
#> 1988P4  57.75124   3.788657 15.243198 6.825983e-05
#> 1988P8  62.98014   4.006304 15.720259 7.127107e-08
#> 1988P7  61.49839   3.239559 18.983567 3.518826e-11
#> 1988P3  56.32520   1.781868 31.610200 7.333807e-08
#> 1988P2  66.56078   4.701584 14.157096 2.150758e-04
#> 1988P6 104.97349  37.998151  2.762595 3.790336e-01
#> 1989F6  55.27332   5.345507 10.340144 4.786636e-06
#> 1989F5  51.26706   3.391918 15.114475 2.551122e-06
#> 1989F4  53.81049   2.669678 20.156176 4.228986e-09
#> 1989F1  56.62630   2.721298 20.808560 2.227069e-06
#> 1989F2  52.24016   2.860031 18.265595 5.733555e-10
#> 1989F7  51.37754   2.538522 20.239155 2.503115e-06
#> 1989F8  47.96772   2.653212 18.079114 2.608897e-06
#> 1989F3  66.12355   4.004286 16.513192 7.770932e-08
#> 1989P7  46.34313   1.748791 26.500094 2.015245e-07
#> 1989P4  57.18041   2.458381 23.259378 3.691032e-08
#> 1989P6  58.23818   3.195134 18.227147 3.766450e-07
#> 1989P5        NA         NA        NA           NA
#> 1989P1  59.69305   2.877246 20.746595 3.760304e-09
#> 1989P3  53.39565   1.685346 31.682313 4.022773e-08
#> 1989P2  67.17505   4.816749 13.946141 4.718717e-06
#> 1989P8        NA         NA        NA           NA
#> 1990F2  66.28541  10.062524  6.587355 3.658736e-05
#> 1990F3  58.27752   5.679501 10.261027 5.243495e-05
#> 1990F4  70.27180  11.183796  6.283358 6.217567e-05
#> 1990F5  51.27603   1.979733 25.900474 1.528310e-08
#> 1990F1  55.69340   3.010641 18.498849 2.156924e-05
#> 1990F8  55.54949   2.120896 26.191526 1.497685e-06
#> 1990F7  54.73623   1.806128 30.305834 5.068892e-08
#> 1990F6  54.56184   2.526903 21.592380 5.054519e-07
#> 1990P8  52.44794   2.384233 21.997824 2.862983e-05
#> 1990P7  54.80233   3.662114 14.964672 2.073748e-05
#> 1990P3  49.71495   2.322466 21.406105 1.087806e-05
#> 1990P1  47.91706   1.467730 32.647045 9.558243e-08
#> 1990P6  50.23032   1.676583 29.959941 7.154673e-06
#> 1990P5  51.15028   1.753041 29.178026 5.361917e-06
#> 1990P2  62.35966   4.676813 13.333794 3.906739e-07
#> 1990P4  61.19885   3.832619 15.967893 5.339368e-05
#>    scal 
#>         Estimate Std. Error  t value     Pr(>|t|)
#> 1988F4  5.176820   1.091749 4.741768 1.854692e-03
#> 1988F2  8.406559   1.736041 4.842373 2.948576e-05
#> 1988F1  9.604743   1.723335 5.573345 1.654535e-06
#> 1988F7  8.069281   1.632197 4.943814 5.772665e-04
#> 1988F5 11.262835   2.789211 4.038001 6.472908e-04
#> 1988F8  9.000647   1.511850 5.953399 5.185132e-06
#> 1988F6  9.718840   1.633690 5.949010 9.256229e-08
#> 1988F3  7.642380   1.104895 6.916837 6.251672e-05
#> 1988P1  6.362358   1.091349 5.829810 4.497602e-05
#> 1988P5  6.809095   1.176508 5.787547 7.181926e-06
#> 1988P4 11.744707   1.893425 6.202889 1.135466e-02
#> 1988P8 10.947130   1.749418 6.257583 3.661666e-05
#> 1988P7 10.177799   1.561879 6.516380 6.056751e-08
#> 1988P3  8.315917   1.151803 7.219912 1.167177e-03
#> 1988P2 11.916114   1.711035 6.964271 1.092030e-02
#> 1988P6 17.929681   3.017031 5.942822 8.315878e-02
#> 1989F6  8.856063   3.535774 2.504702 3.994054e-03
#> 1989F5  7.205896   2.373316 3.036215 5.159590e-03
#> 1989F4  6.486573   2.153628 3.011928 5.325839e-05
#> 1989F1  6.068136   2.374275 2.555784 3.089681e-02
#> 1989F2  7.016427   2.076585 3.378829 2.598127e-06
#> 1989F7  5.500159   1.926244 2.855381 2.054614e-02
#> 1989F8  5.960926   1.718186 3.469314 6.366473e-03
#> 1989F3  9.224764   2.337292 3.946775 9.227941e-05
#> 1989P7  5.393890   1.125755 4.791354 8.574281e-04
#> 1989P4  8.402098   1.676272 5.012370 7.397974e-05
#> 1989P6 10.613533   1.854108 5.724333 1.138299e-04
#> 1989P5        NA         NA       NA           NA
#> 1989P1  9.972837   1.717799 5.805590 2.155525e-06
#> 1989P3  7.900595   1.179248 6.699689 8.814552e-05
#> 1989P2 12.523509   2.108665 5.939068 3.025485e-04
#> 1989P8        NA         NA       NA           NA
#> 1990F2 13.157219   3.660268 3.594605 6.661875e-04
#> 1990F3 12.796287   2.697127 4.744413 1.984977e-03
#> 1990F4 14.560188   3.415482 4.262996 3.997895e-04
#> 1990F5  7.758448   1.379453 5.624293 3.012976e-05
#> 1990F1  9.617031   1.745763 5.508782 6.062151e-03
#> 1990F8  7.771085   1.327913 5.852106 2.041157e-03
#> 1990F7  6.792194   1.185178 5.730947 1.887271e-04
#> 1990F6  9.263610   1.528612 6.060144 2.589908e-04
#> 1990P8  8.581012   1.561491 5.495395 1.537997e-02
#> 1990P7 10.847282   2.066123 5.250066 2.908553e-03
#> 1990P3  9.322305   1.585599 5.879358 4.746622e-03
#> 1990P1  6.611841   1.153215 5.733397 4.896358e-04
#> 1990P6  6.626907   1.260483 5.257433 2.008824e-02
#> 1990P5  7.293155   1.249137 5.838557 9.704477e-03
#> 1990P2 11.656924   2.055390 5.671392 2.695022e-05
#> 1990P4 10.971539   1.804042 6.081642 4.695487e-03
#> 
#> Residual standard error: 1.020867 on 258 degrees of freedom
#>