Extract variance and correlation components

This function calculates the estimated variances, standard deviations, and correlations between the random-effects terms in a linear mixed-effects model, of class "lme", or a nonlinear mixed-effects model, of class "nlme". The within-group error variance and standard deviation are also calculated.

Usage

VarCorr(x, sigma = 1, ...)
# S3 method for lme
VarCorr(x, sigma = x$sigma, rdig = 3, ...)

# S3 method for pdMat
VarCorr(x, sigma = 1, rdig = 3, ...)
# S3 method for pdBlocked
VarCorr(x, sigma = 1, rdig = 3, ...)

Arguments

x

a fitted model object, usually an object inheriting from class "lme".

sigma

an optional numeric value used as a multiplier for the standard deviations. The default is x$sigma or 1 depending on class(x).

rdig

an optional integer value specifying the number of digits used to represent correlation estimates. Default is 3.

...

further optional arguments passed to other methods (none for the methods documented here).

Value

a matrix with the estimated variances, standard deviations, and correlations for the random effects. The first two columns, named

Variance and StdDev, give, respectively, the variance and the standard deviations. If there are correlation components in the random effects model, the third column, named Corr, and the remaining unnamed columns give the estimated correlations among random effects within the same level of grouping. The within-group error variance and standard deviation are included as the last row in the matrix.

References

Pinheiro, J.C., and Bates, D.M. (2000) Mixed-Effects Models in S and S-PLUS, Springer, esp. pp. 100, 461.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

fm1 <- lme(distance ~ age, data = Orthodont, random = ~age)
VarCorr(fm1)
#> Subject = pdLogChol(age) 
#>             Variance   StdDev    Corr  
#> (Intercept) 5.41508758 2.3270341 (Intr)
#> age         0.05126955 0.2264278 -0.609
#> Residual    1.71620400 1.3100397