Variogram.corRatio {nlme}	R Documentation

Calculate Semi-variogram for a corRatio Object

Description

This method function calculates the semi-variogram values corresponding to the Rational Quadratic correlation model, using the estimated coefficients corresponding to object, at the distances defined by distance.

Usage

## S3 method for class 'corRatio':
Variogram(object, distance, sig2, length.out, ...)

Arguments

object 	 an object inheriting from class corRatio, representing an Rational Quadratic spatial correlation structure.

distance 	 an optional numeric vector with the distances at which the semi-variogram is to be calculated. Defaults to NULL, in which case a sequence of length length.out between the minimum and maximum values of getCovariate(object) is used.

sig2 	 an optional numeric value representing the process variance. Defaults to 1.

length.out 	 an optional integer specifying the length of the sequence of distances to be used for calculating the semi-variogram, when distance = NULL. Defaults to 50.

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

a data frame with columns variog and dist representing, respectively, the semi-variogram values and the corresponding distances. The returned value inherits from class Variogram.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

References

Cressie, N.A.C. (1993), "Statistics for Spatial Data", J. Wiley & Sons.

See Also

corRatio, plot.Variogram Variogram

Examples

cs1 <- corRatio(7, form = ~ Time | Rat)
cs1 <- Initialize(cs1, BodyWeight)
Variogram(cs1)[1:10,]

[Package nlme version 3.1-93 Index]