Calculate Semi-variogram for a corSpatial Object

This method function calculates the semi-variogram values corresponding to the model defined in FUN, using the estimated coefficients corresponding to object, at the distances defined by distance.

Usage

# S3 method for corSpatial
Variogram(object, distance, sig2, length.out, FUN, ...)

Arguments

object

an object inheriting from class "corSpatial", representing spatial correlation structure.

distance

an optional numeric vector with the distances at which the semi-variogram is to be calculated. Defaults to NULL, in which case a sequence of length length.out between the minimum and maximum values of getCovariate(object) is used.

sig2

an optional numeric value representing the process variance. Defaults to 1.

length.out

an optional integer specifying the length of the sequence of distances to be used for calculating the semi-variogram, when distance = NULL. Defaults to 50.

FUN

a function of two arguments, the distance and the range corresponding to object, specifying the semi-variogram model.

...

some methods for this generic require additional arguments. None are used in this method.

Value

a data frame with columns variog and dist representing, respectively, the semi-variogram values and the corresponding distances. The returned value inherits from class Variogram.

References

Cressie, N.A.C. (1993), "Statistics for Spatial Data", J. Wiley & Sons.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

cs1 <- corExp(3, form = ~ Time | Rat)
cs1 <- Initialize(cs1, BodyWeight)
Variogram(cs1, FUN = function(x, y) (1 - exp(-x/y)))[1:10,]
#>       variog      dist
#> 1  0.2834687  1.000000
#> 2  0.5300373  2.265306
#> 3  0.6917582  3.530612
#> 4  0.7978286  4.795918
#> 5  0.8673987  6.061224
#> 6  0.9130287  7.326531
#> 7  0.9429568  8.591837
#> 8  0.9625861  9.857143
#> 9  0.9754608 11.122449
#> 10 0.9839051 12.387755