allCoef {nlme}	R Documentation

Extract Coefficients from a Set of Objects

Description

The extractor function is applied to each object in ..., with the result being converted to a vector. A map attribute is included to indicate which pieces of the returned vector correspond to the original objects in dots.

Usage


allCoef(..., extract)

Arguments

...
objects to which extract will be applied. Generally these will be model components, such as corStruct and varFunc objects.

extract
an optional extractor function. Defaults to coef.

Value

a vector with all elements, generally coefficients, obtained by applying extract to the objects in ....

Author(s)

José' Pinheiro and Douglas Bates

See Also

lmeStruct,nlmeStruct

Examples


cs1 <- corAR1(0.1)
vf1 <- varPower(0.5)
allCoef(cs1, vf1)

[Package nlme version 3.1-137 Index]