Language scores

The bdf data frame has 2287 rows and 25 columns of language scores from grade 8 pupils in elementary schools in The Netherlands.

Usage

data(bdf)

Format

schoolNR

a factor denoting the school.

pupilNR

a factor denoting the pupil.

IQ.verb

a numeric vector of verbal IQ scores

IQ.perf

a numeric vector of IQ scores.

sex

Sex of the student.

Minority

a factor indicating if the student is a member of a minority group.

repeatgr

an ordered factor indicating if one or more grades have been repeated.

aritPRET

a numeric vector

classNR

a numeric vector

aritPOST

a numeric vector

langPRET

a numeric vector

langPOST

a numeric vector

ses

a numeric vector of socioeconomic status indicators.

denomina

a factor indicating of the school is a public school, a Protestant private school, a Catholic private school, or a non-denominational private school.

schoolSES

a numeric vector

satiprin

a numeric vector

natitest

a factor with levels 0 and 1

meetings

a numeric vector

currmeet

a numeric vector

mixedgra

a factor indicating if the class is a mixed-grade class.

percmino

a numeric vector

aritdiff

a numeric vector

homework

a numeric vector

classsiz

a numeric vector

groupsiz

a numeric vector

Source

http://stat.gamma.rug.nl/snijders/multilevel.htm, the first edition of http://www.stats.ox.ac.uk/~snijders/mlbook.htm.

References

Snijders, Tom and Bosker, Roel (1999), Multilevel Analysis: An Introduction to Basic and Advanced Multilevel Modeling, Sage.

Examples

summary(bdf)
#>     schoolNR       pupilNR        IQ.verb         IQ.perf       sex     
#>  40     :  35   17001  :   1   Min.   : 4.00   Min.   : 5.000   0:1194  
#>  155    :  33   17002  :   1   1st Qu.:10.50   1st Qu.: 9.333   1:1093  
#>  183    :  31   17003  :   1   Median :12.00   Median :11.000           
#>  54     :  31   17004  :   1   Mean   :11.83   Mean   :11.047           
#>  159    :  31   17005  :   1   3rd Qu.:13.00   3rd Qu.:12.667           
#>  161    :  31   17006  :   1   Max.   :18.00   Max.   :17.667           
#>  (Other):2095   (Other):2281                                            
#>  Minority repeatgr    aritPRET        classNR         aritPOST    
#>  N:2155   0:1988   Min.   : 1.00   Min.   :  180   Min.   : 2.00  
#>  Y: 132   1: 295   1st Qu.: 9.00   1st Qu.: 6780   1st Qu.:14.00  
#>           2:   4   Median :12.00   Median :14180   Median :20.00  
#>                    Mean   :11.94   Mean   :13382   Mean   :19.44  
#>                    3rd Qu.:14.00   3rd Qu.:19580   3rd Qu.:25.00  
#>                    Max.   :20.00   Max.   :25880   Max.   :30.00  
#>                                                                   
#>     langPRET        langPOST          ses        denomina   schoolSES 
#>  Min.   :11.00   Min.   : 9.00   Min.   :10.00   1:775    Min.   :10  
#>  1st Qu.:30.00   1st Qu.:35.00   1st Qu.:20.00   2:803    1st Qu.:16  
#>  Median :35.00   Median :42.00   Median :27.00   3:617    Median :18  
#>  Mean   :34.19   Mean   :40.93   Mean   :27.81   4: 92    Mean   :19  
#>  3rd Qu.:39.00   3rd Qu.:48.00   3rd Qu.:35.00            3rd Qu.:22  
#>  Max.   :49.00   Max.   :58.00   Max.   :50.00            Max.   :29  
#>                                                                       
#>     satiprin     natitest    meetings        currmeet     mixedgra
#>  Min.   :2.143   0:1056   Min.   :1.100   Min.   :1.000   0:1658  
#>  1st Qu.:3.143   1:1231   1st Qu.:1.800   1st Qu.:1.667   1: 629  
#>  Median :3.286            Median :2.000   Median :1.833           
#>  Mean   :3.325            Mean   :2.055   Mean   :1.973           
#>  3rd Qu.:3.571            3rd Qu.:2.300   3rd Qu.:2.167           
#>  Max.   :4.000            Max.   :3.600   Max.   :3.333           
#>                                                                   
#>     percmino         aritdiff        homework        classsiz    
#>  Min.   : 0.000   Min.   : 8.00   Min.   :1.333   Min.   :10.00  
#>  1st Qu.: 0.000   1st Qu.:13.00   1st Qu.:2.000   1st Qu.:23.00  
#>  Median : 0.000   Median :15.00   Median :2.667   Median :27.00  
#>  Mean   : 6.579   Mean   :16.15   Mean   :2.462   Mean   :26.51  
#>  3rd Qu.: 6.000   3rd Qu.:19.00   3rd Qu.:2.667   3rd Qu.:31.00  
#>  Max.   :90.000   Max.   :27.00   Max.   :3.667   Max.   :39.00  
#>                                                                  
#>     groupsiz      IQ.ver.cen      avg.IQ.ver.cen      grpSiz.cen      
#>  Min.   : 5.0   Min.   :-7.8341   Min.   :-5.0841   Min.   :-18.1006  
#>  1st Qu.:17.0   1st Qu.:-1.3341   1st Qu.:-0.4174   1st Qu.: -6.1006  
#>  Median :24.0   Median : 0.1659   Median : 0.1059   Median :  0.8994  
#>  Mean   :23.1   Mean   : 0.0000   Mean   : 0.0000   Mean   :  0.0000  
#>  3rd Qu.:28.0   3rd Qu.: 1.1659   3rd Qu.: 0.5231   3rd Qu.:  4.8994  
#>  Max.   :37.0   Max.   : 6.1659   Max.   : 1.8993   Max.   : 13.8994  
#>                                                                       

## More examples, including lme() fits  reproducing parts in the above
## book, are available in the R script files
system.file("mlbook", "ch04.R", package ="nlme") # and
#> [1] "C:/Users/neljasonhaw/AppData/Local/Programs/R/R-4.2.2/library/nlme/mlbook/ch04.R"
system.file("mlbook", "ch05.R", package ="nlme")
#> [1] "C:/Users/neljasonhaw/AppData/Local/Programs/R/R-4.2.2/library/nlme/mlbook/ch05.R"