coef.gnls {nlme}	R Documentation

Extract gnls Coefficients

Description

The estimated coefficients for the nonlinear model represented by object are extracted.

Usage

## S3 method for class 'gnls'
coef(object, ...)

Arguments

object
an object inheriting from class "gnls", representing a generalized nonlinear least squares fitted model.

...
some methods for this generic require additional arguments. None are used in this method.

Value

a vector with the estimated coefficients for the nonlinear model represented by object.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

gnls

Examples

fm1 <- gnls(weight ~ SSlogis(Time, Asym, xmid, scal), Soybean,
            weights = varPower())
coef(fm1)

[Package nlme version 3.1-121 Index]