coef.modelStruct {nlme}	R Documentation

Extract modelStruct Object Coefficients

Description

This method function extracts the coefficients associated with each component of the modelStruct list.

Usage

## S3 method for class 'modelStruct'
coef(object, unconstrained, ...)
## S3 replacement method for class 'modelStruct'
coef(object, ...) <- value

Arguments

object
an object inheriting from class "modelStruct", representing a list of model components, such as "corStruct" and "varFunc" objects.

unconstrained
a logical value. If TRUE the coefficients are returned in unconstrained form (the same used in the optimization algorithm). If FALSE the coefficients are returned in "natural", possibly constrained, form. Defaults to TRUE.

value
a vector with the replacement values for the coefficients associated with object. It must be a vector with the same length of coef{object} and must be given in unconstrained form.

...
some methods for this generic require additional arguments. None are used in this method.

Value

a vector with all coefficients corresponding to the components of object.

SIDE EFFECTS

On the left side of an assignment, sets the values of the coefficients of object to value. Object must be initialized (using Initialize) before new values can be assigned to its coefficients.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

Initialize

Examples

lms1 <- lmeStruct(reStruct = reStruct(pdDiag(diag(2), ~age)),
   corStruct = corAR1(0.3))
coef(lms1)

[Package nlme version 3.1-127 Index]