collapse {nlme}	R Documentation

Collapse According to Groups

Description

This function is generic; method functions can be written to handle specific classes of objects. Currently, only a groupedData method is available.

Usage

collapse(object, ...)

Arguments

object
an object to be collapsed, usually a data frame.

...
some methods for the generic may require additional arguments.

Value

will depend on the method function used; see the appropriate documentation.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

collapse.groupedData

Examples

## see the method function documentation

[Package nlme version 3.1-127 Index]