Compare Predictions

Predicted values are obtained at the specified values of primary for each object. If either object1 or object2 have a grouping structure (i.e. getGroups(object) is not NULL), predicted values are obtained for each group. When both objects determine groups, the group levels must be the same. If other covariates besides primary are used in the prediction model, their group-wise averages (numeric covariates) or most frequent values (categorical covariates) are used to obtain the predicted values. The original observations are also included in the returned object.

Usage

comparePred(object1, object2, primary, minimum, maximum,
    length.out, level, ...)

Arguments

object1,object2

fitted model objects, from which predictions can be extracted using the predict method.

primary

an optional one-sided formula specifying the primary covariate to be used to generate the augmented predictions. By default, if a covariate can be extracted from the data used to generate the objects (using getCovariate), it will be used as primary.

minimum

an optional lower limit for the primary covariate. Defaults to min(primary), after primary is evaluated in the data used in fitting object1.

maximum

an optional upper limit for the primary covariate. Defaults to max(primary), after primary is evaluated in the data used in fitting object1.

length.out

an optional integer with the number of primary covariate values at which to evaluate the predictions. Defaults to 51.

level

an optional integer specifying the desired prediction level. Levels increase from outermost to innermost grouping, with level 0 representing the population (fixed effects) predictions. Only one level can be specified. Defaults to the innermost level.

...

some methods for the generic may require additional arguments.

Value

a data frame with four columns representing, respectively, the values of the primary covariate, the groups (if object does not have a grouping structure, all elements will be 1), the predicted or observed values, and the type of value in the third column: the objects' names are used to classify the predicted values and

original is used for the observed values. The returned object inherits from classes comparePred and augPred.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

Note

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include: gls, lme, and lmList.

See also

Examples

fm1 <- lme(distance ~ age * Sex, data = Orthodont, random = ~ age)
fm2 <- update(fm1, distance ~ age)
comparePred(fm1, fm2, length.out = 2)
#>      age .groups distance    .type
#> 1      8     M01 26.00000 original
#> 2     10     M01 25.00000 original
#> 3     12     M01 29.00000 original
#> 4     14     M01 31.00000 original
#> 5      8     M02 21.50000 original
#> 6     10     M02 22.50000 original
#> 7     12     M02 23.00000 original
#> 8     14     M02 26.50000 original
#> 9      8     M03 23.00000 original
#> 10    10     M03 22.50000 original
#> 11    12     M03 24.00000 original
#> 12    14     M03 27.50000 original
#> 13     8     M04 25.50000 original
#> 14    10     M04 27.50000 original
#> 15    12     M04 26.50000 original
#> 16    14     M04 27.00000 original
#> 17     8     M05 20.00000 original
#> 18    10     M05 23.50000 original
#> 19    12     M05 22.50000 original
#> 20    14     M05 26.00000 original
#> 21     8     M06 24.50000 original
#> 22    10     M06 25.50000 original
#> 23    12     M06 27.00000 original
#> 24    14     M06 28.50000 original
#> 25     8     M07 22.00000 original
#> 26    10     M07 22.00000 original
#> 27    12     M07 24.50000 original
#> 28    14     M07 26.50000 original
#> 29     8     M08 24.00000 original
#> 30    10     M08 21.50000 original
#> 31    12     M08 24.50000 original
#> 32    14     M08 25.50000 original
#> 33     8     M09 23.00000 original
#> 34    10     M09 20.50000 original
#> 35    12     M09 31.00000 original
#> 36    14     M09 26.00000 original
#> 37     8     M10 27.50000 original
#> 38    10     M10 28.00000 original
#> 39    12     M10 31.00000 original
#> 40    14     M10 31.50000 original
#> 41     8     M11 23.00000 original
#> 42    10     M11 23.00000 original
#> 43    12     M11 23.50000 original
#> 44    14     M11 25.00000 original
#> 45     8     M12 21.50000 original
#> 46    10     M12 23.50000 original
#> 47    12     M12 24.00000 original
#> 48    14     M12 28.00000 original
#> 49     8     M13 17.00000 original
#> 50    10     M13 24.50000 original
#> 51    12     M13 26.00000 original
#> 52    14     M13 29.50000 original
#> 53     8     M14 22.50000 original
#> 54    10     M14 25.50000 original
#> 55    12     M14 25.50000 original
#> 56    14     M14 26.00000 original
#> 57     8     M15 23.00000 original
#> 58    10     M15 24.50000 original
#> 59    12     M15 26.00000 original
#> 60    14     M15 30.00000 original
#> 61     8     M16 22.00000 original
#> 62    10     M16 21.50000 original
#> 63    12     M16 23.50000 original
#> 64    14     M16 25.00000 original
#> 65     8     F01 21.00000 original
#> 66    10     F01 20.00000 original
#> 67    12     F01 21.50000 original
#> 68    14     F01 23.00000 original
#> 69     8     F02 21.00000 original
#> 70    10     F02 21.50000 original
#> 71    12     F02 24.00000 original
#> 72    14     F02 25.50000 original
#> 73     8     F03 20.50000 original
#> 74    10     F03 24.00000 original
#> 75    12     F03 24.50000 original
#> 76    14     F03 26.00000 original
#> 77     8     F04 23.50000 original
#> 78    10     F04 24.50000 original
#> 79    12     F04 25.00000 original
#> 80    14     F04 26.50000 original
#> 81     8     F05 21.50000 original
#> 82    10     F05 23.00000 original
#> 83    12     F05 22.50000 original
#> 84    14     F05 23.50000 original
#> 85     8     F06 20.00000 original
#> 86    10     F06 21.00000 original
#> 87    12     F06 21.00000 original
#> 88    14     F06 22.50000 original
#> 89     8     F07 21.50000 original
#> 90    10     F07 22.50000 original
#> 91    12     F07 23.00000 original
#> 92    14     F07 25.00000 original
#> 93     8     F08 23.00000 original
#> 94    10     F08 23.00000 original
#> 95    12     F08 23.50000 original
#> 96    14     F08 24.00000 original
#> 97     8     F09 20.00000 original
#> 98    10     F09 21.00000 original
#> 99    12     F09 22.00000 original
#> 100   14     F09 21.50000 original
#> 101    8     F10 16.50000 original
#> 102   10     F10 19.00000 original
#> 103   12     F10 19.00000 original
#> 104   14     F10 19.50000 original
#> 105    8     F11 24.50000 original
#> 106   10     F11 25.00000 original
#> 107   12     F11 28.00000 original
#> 108   14     F11 28.00000 original
#> 109    8     M01 24.84572      fm1
#> 110   14     M01 30.03802      fm1
#> 111    8     M02 21.27478      fm1
#> 112   14     M02 25.83966      fm1
#> 113    8     M03 22.03311      fm1
#> 114   14     M03 26.62726      fm1
#> 115    8     M04 24.46452      fm1
#> 116   14     M04 28.32495      fm1
#> 117    8     M05 20.90249      fm1
#> 118   14     M05 25.55790      fm1
#> 119    8     M06 23.88527      fm1
#> 120   14     M06 28.52762      fm1
#> 121    8     M07 21.57351      fm1
#> 122   14     M07 26.20819      fm1
#> 123    8     M08 21.99187      fm1
#> 124   14     M08 25.95500      fm1
#> 125    8     M09 22.60752      fm1
#> 126   14     M09 27.63182      fm1
#> 127    8     M10 26.47272      fm1
#> 128   14     M10 31.48306      fm1
#> 129    8     M11 21.81724      fm1
#> 130   14     M11 25.68038      fm1
#> 131    8     M12 21.84919      fm1
#> 132   14     M12 26.84421      fm1
#> 133    8     M13 21.15031      fm1
#> 134   14     M13 27.66861      fm1
#> 135    8     M14 22.72716      fm1
#> 136   14     M14 27.01008      fm1
#> 137    8     M15 23.13140      fm1
#> 138   14     M15 28.45567      fm1
#> 139    8     M16 21.12319      fm1
#> 140   14     M16 25.29756      fm1
#> 141    8     F01 20.20973      fm1
#> 142   14     F01 22.81848      fm1
#> 143    8     F02 21.27124      fm1
#> 144   14     F02 24.69027      fm1
#> 145    8     F03 21.86869      fm1
#> 146   14     F03 25.42735      fm1
#> 147    8     F04 23.09591      fm1
#> 148   14     F04 26.14246      fm1
#> 149    8     F05 21.34035      fm1
#> 150   14     F05 23.88784      fm1
#> 151    8     F06 19.99832      fm1
#> 152   14     F06 22.58725      fm1
#> 153    8     F07 21.45516      fm1
#> 154   14     F07 24.47333      fm1
#> 155    8     F08 22.04815      fm1
#> 156   14     F08 24.49475      fm1
#> 157    8     F09 20.07189      fm1
#> 158   14     F09 22.50047      fm1
#> 159    8     F10 17.72334      fm1
#> 160   14     F10 20.22443      fm1
#> 161    8     F11 24.21723      fm1
#> 162   14     F11 27.70339      fm1
#> 1091   8     M01 24.81965      fm2
#> 1101  14     M01 30.07487      fm2
#> 1111   8     M02 21.43115      fm2
#> 1121  14     M02 25.47931      fm2
#> 1131   8     M03 22.15551      fm2
#> 1141  14     M03 26.33173      fm2
#> 1151   8     M04 24.51597      fm2
#> 1161  14     M04 28.08177      fm2
#> 1171   8     M05 21.07073      fm2
#> 1181  14     M05 25.18544      fm2
#> 1191   8     M06 23.92577      fm2
#> 1201  14     M06 28.38602      fm2
#> 1211   8     M07 21.71382      fm2
#> 1221  14     M07 25.88433      fm2
#> 1231   8     M08 22.14500      fm2
#> 1241  14     M08 25.53769      fm2
#> 1251   8     M09 22.68547      fm2
#> 1261  14     M09 27.46203      fm2
#> 1271   8     M10 26.38500      fm2
#> 1281  14     M10 31.61299      fm2
#> 1291   8     M11 21.98248      fm2
#> 1301  14     M11 25.23063      fm2
#> 1311   8     M12 21.96111      fm2
#> 1321  14     M12 26.60961      fm2
#> 1331   8     M13 21.22240      fm2
#> 1341  14     M13 27.66552      fm2
#> 1351   8     M14 22.83398      fm2
#> 1361  14     M14 26.71037      fm2
#> 1371   8     M15 23.17304      fm2
#> 1381  14     M15 28.38322      fm2
#> 1391   8     M16 21.30401      fm2
#> 1401  14     M16 24.85199      fm2
#> 1411   8     F01 20.13096      fm2
#> 1421  14     F01 23.02281      fm2
#> 1431   8     F02 21.10961      fm2
#> 1441  14     F02 25.12987      fm2
#> 1451   8     F03 21.67494      fm2
#> 1461  14     F03 25.93991      fm2
#> 1471   8     F04 22.87286      fm2
#> 1481  14     F04 26.65480      fm2
#> 1491   8     F05 21.21574      fm2
#> 1501  14     F05 24.16910      fm2
#> 1511   8     F06 19.92955      fm2
#> 1521  14     F06 22.77132      fm2
#> 1531   8     F07 21.30401      fm2
#> 1541  14     F07 24.85199      fm2
#> 1551   8     F08 21.89771      fm2
#> 1561  14     F08 24.81242      fm2
#> 1571   8     F09 20.00731      fm2
#> 1581  14     F09 22.66017      fm2
#> 1591   8     F10 17.75649      fm2
#> 1601  14     F10 20.21405      fm2
#> 1611   8     F11 23.92577      fm2
#> 1621  14     F11 28.38602      fm2