Compound Symmetry Correlation Structure

This function is a constructor for the corCompSymm class, representing a compound symmetry structure corresponding to uniform correlation. Objects created using this constructor must later be initialized using the appropriate Initialize method.

Usage

corCompSymm(value, form, fixed)

Arguments

value

the correlation between any two correlated observations. Defaults to 0.

form

a one sided formula of the form ~ t, or ~ t | g, specifying a time covariate t and, optionally, a grouping factor g. When a grouping factor is present in form, the correlation structure is assumed to apply only to observations within the same grouping level; observations with different grouping levels are assumed to be uncorrelated. Defaults to ~ 1, which corresponds to using the order of the observations in the data as a covariate, and no groups.

fixed

an optional logical value indicating whether the coefficients should be allowed to vary in the optimization, or kept fixed at their initial value. Defaults to FALSE, in which case the coefficients are allowed to vary.

Value

an object of class corCompSymm, representing a compound symmetry correlation structure.

References

Milliken, G. A. and Johnson, D. E. (1992) "Analysis of Messy Data, Volume I: Designed Experiments", Van Nostrand Reinhold.

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer, esp. pp. 233-234.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

## covariate is observation order and grouping factor is Subject
cs1 <- corCompSymm(0.5, form = ~ 1 | Subject)

# Pinheiro and Bates, pp. 222-225 
fm1BW.lme <- lme(weight ~ Time * Diet, BodyWeight,
                   random = ~ Time)
# p. 223
fm2BW.lme <- update(fm1BW.lme, weights = varPower())
# p. 225
cs1CompSymm <- corCompSymm(value = 0.3, form = ~ 1 | Subject)
cs2CompSymm <- corCompSymm(value = 0.3, form = ~ age | Subject)
cs1CompSymm <- Initialize(cs1CompSymm, data = Orthodont)
corMatrix(cs1CompSymm)
#> $M01
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M02
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M03
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M04
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M05
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M06
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M07
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M08
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M09
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M10
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M11
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M12
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M13
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M14
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M15
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $M16
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $F01
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $F02
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $F03
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $F04
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $F05
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $F06
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $F07
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $F08
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $F09
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $F10
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 
#> $F11
#>      [,1] [,2] [,3] [,4]
#> [1,]  1.0  0.3  0.3  0.3
#> [2,]  0.3  1.0  0.3  0.3
#> [3,]  0.3  0.3  1.0  0.3
#> [4,]  0.3  0.3  0.3  1.0
#> 

## Print/Summary methods for the empty case:
(cCS <- corCompSymm()) # Uninitialized correlation struc..
#> Uninitialized correlation structure of class corCompSymm
summary(cCS)           #    (ditto)
#> Uninitialized correlation structure of class corCompSymm