corFactor {nlme}	R Documentation

Factor of a Correlation Matrix

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include all corStruct classes.

Usage

corFactor(object, ...)

Arguments

object
an object from which a correlation matrix can be extracted.

...
some methods for this generic function require additional arguments.

Value

will depend on the method function used; see the appropriate documentation.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

corFactor.corStruct, recalc.corStruct

Examples

## see the method function documentation

[Package nlme version 3.1-127 Index]