fitted.glsStruct {nlme}	R Documentation

Calculate glsStruct Fitted Values

Description

The fitted values for the linear model represented by object are extracted.

Usage


## S3 method for class 'glsStruct'
fitted(object, glsFit, ...)

Arguments

object
an object inheriting from class "glsStruct", representing a list of linear model components, such as corStruct and "varFunc" objects.

glsFit
an optional list with components logLik (log-likelihood), beta (coefficients), sigma (standard deviation for error term), varBeta (coefficients' covariance matrix), fitted (fitted values), and residuals (residuals). Defaults to attr(object, "glsFit").

...
some methods for this generic require additional arguments. None are used in this method.

Value

a vector with the fitted values for the linear model represented by object.

Note

This method function is generally only used inside gls and fitted.gls.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

gls, residuals.glsStruct

[Package nlme version 3.1-137 Index]