fixed.effects {nlme}	R Documentation

Extract Fixed Effects

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include lmList and lme.

Usage

fixed.effects(object, ...)
fixef(object, ...)

Arguments

object 	 any fitted model object from which fixed effects estimates can be extracted.

... 	 some methods for this generic function require additional arguments.

Value

will depend on the method function used; see the appropriate documentation.

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

fixef.lmList

Examples

## see the method function documentation

[Package nlme version 3.1-93 Index]