formula.reStruct {nlme}	R Documentation

Extract reStruct Object Formula

Description

This method function extracts a formula from each of the components of x, returning a list of formulas.

Usage

## S3 method for class 'reStruct':
formula(x, asList, ...)

Arguments

x 	 an object inheriting from class reStruct, representing a random effects structure and consisting of a list of pdMat objects.

asList 	 logical. Should the asList argument be applied to each of the components?

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

a list with the formulas of each component of x.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

formula

Examples

rs1 <- reStruct(list(A = pdDiag(diag(2), ~age), B = ~1))
formula(rs1)

[Package nlme version 3.1-93 Index]