getCovariate.corStruct {nlme}	R Documentation

Extract corStruct Object Covariate

Description

This method function extracts the covariate(s) associated with object.

Usage


## S3 method for class 'corStruct'
getCovariate(object, form, data)

Arguments

object
an object inheriting from class corStruct representing a correlation structure.

form
this argument is included to make the method function compatible with the generic. It will be assigned the value of formula(object) and should not be modified.

data
an optional data frame in which to evaluate the variables defined in form, in case object is not initialized and the covariate needs to be evaluated.

Value

when the correlation structure does not include a grouping factor, the returned value will be a vector or a matrix with the covariate(s) associated with object. If a grouping factor is present, the returned value will be a list of vectors or matrices with the covariate(s) corresponding to each grouping level.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

getCovariate

Examples


cs1 <- corAR1(form = ~ 1 | Subject)
getCovariate(cs1, data = Orthodont)

[Package nlme version 3.1-137 Index]