getCovariate.varFunc {nlme}	R Documentation

Extract varFunc Covariate

Description

This method function extracts the covariate(s) associated with the variance function represented by object, if any is present.

Usage


## S3 method for class 'varFunc'
getCovariate(object, form, data)

Arguments

object
an object inheriting from class varFunc, representing a variance function structure.

form
an optional formula specifying the covariate to be evaluated in object. Defaults to formula(object).

data
some methods for this generic require a data object. Not used in this method.

Value

if object has a covariate attribute, its value is returned; else NULL is returned.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

covariate<-.varFunc

Examples


vf1 <- varPower(1.1, form = ~age)
covariate(vf1) <- Orthodont[["age"]]
getCovariate(vf1)

[Package nlme version 3.1-137 Index]