getData {nlme}	R Documentation

Extract Data from an Object

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include gls, lme, and lmList.

Usage

getData(object)

Arguments

object
an object from which a data.frame can be extracted, generally a fitted model object.

Value

will depend on the method function used; see the appropriate documentation.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

getData.gls, getData.lme, getData.lmList

Examples

## see the method function documentation

[Package nlme version 3.1-127 Index]