Extract gls Object Data

If present in the calling sequence used to produce object, the data frame used to fit the model is obtained.

Usage

# S3 method for gls
getData(object)

Arguments

object

an object inheriting from class gls, representing a generalized least squares fitted linear model.

Value

if a data argument is present in the calling sequence that produced object, the corresponding data frame (with

na.action and subset applied to it, if also present in the call that produced object) is returned; else, NULL is returned.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

fm1 <- gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), data = Ovary,
           correlation = corAR1(form = ~ 1 | Mare))
getData(fm1)
#> Grouped Data: follicles ~ Time | Mare
#>     Mare        Time follicles
#> 1      1 -0.13636360        20
#> 2      1 -0.09090910        15
#> 3      1 -0.04545455        19
#> 4      1  0.00000000        16
#> 5      1  0.04545455        13
#> 6      1  0.09090910        10
#> 7      1  0.13636360        12
#> 8      1  0.18181820        14
#> 9      1  0.22727270        13
#> 10     1  0.27272730        20
#> 11     1  0.31818180        22
#> 12     1  0.36363640        15
#> 13     1  0.40909090        18
#> 14     1  0.45454550        17
#> 15     1  0.50000000        14
#> 16     1  0.54545450        18
#> 17     1  0.59090910        14
#> 18     1  0.63636360        16
#> 19     1  0.68181820        17
#> 20     1  0.72727270        18
#> 21     1  0.77272730        18
#> 22     1  0.81818180        17
#> 23     1  0.86363640        14
#> 24     1  0.90909090        12
#> 25     1  0.95454550        12
#> 26     1  1.00000000        14
#> 27     1  1.04545500        10
#> 28     1  1.09090900        11
#> 29     1  1.13636400        16
#> 30     2 -0.15000000         6
#> 31     2 -0.10000000         6
#> 32     2 -0.05000000         8
#> 33     2  0.00000000         7
#> 34     2  0.05000000        16
#> 35     2  0.10000000        10
#> 36     2  0.15000000        13
#> 37     2  0.20000000         9
#> 38     2  0.25000000         7
#> 39     2  0.30000000         6
#> 40     2  0.35000000         8
#> 41     2  0.40000000         8
#> 42     2  0.45000000         6
#> 43     2  0.50000000         8
#> 44     2  0.55000000         7
#> 45     2  0.60000000         9
#> 46     2  0.65000000         6
#> 47     2  0.70000000         4
#> 48     2  0.75000000         5
#> 49     2  0.80000000         8
#> 50     2  0.85000000        11
#> 51     2  0.90000000        13
#> 52     2  0.95000000        10
#> 53     2  1.00000000         6
#> 54     2  1.05000000         7
#> 55     2  1.10000000         6
#> 56     2  1.15000000         5
#> 57     3 -0.15789470        13
#> 58     3 -0.10526320        11
#> 59     3 -0.05263158        10
#> 60     3  0.00000000         6
#> 61     3  0.05263158         8
#> 62     3  0.10526320         6
#> 63     3  0.15789470         9
#> 64     3  0.21052630         9
#> 65     3  0.26315790        10
#> 66     3  0.31578950         8
#> 67     3  0.36842110        14
#> 68     3  0.42105260        13
#> 69     3  0.47368420        14
#> 70     3  0.52631580        16
#> 71     3  0.57894740        20
#> 72     3  0.63157900        21
#> 73     3  0.68421050        25
#> 74     3  0.73684210        23
#> 75     3  0.78947370        19
#> 76     3  0.84210530        22
#> 77     3  0.89473680        16
#> 78     3  0.94736840        21
#> 79     3  1.00000000        19
#> 80     3  1.05263200        20
#> 81     3  1.10526300        17
#> 82     3  1.15789500        24
#> 83     4 -0.13636360         9
#> 84     4 -0.09090910         9
#> 85     4 -0.04545455         7
#> 86     4  0.00000000         6
#> 87     4  0.04545455         7
#> 88     4  0.09090910         6
#> 89     4  0.13636360         1
#> 90     4  0.18181820         1
#> 91     4  0.22727270         1
#> 92     4  0.27272730         5
#> 93     4  0.31818180         6
#> 94     4  0.36363640         3
#> 95     4  0.40909090         5
#> 96     4  0.45454550         3
#> 97     4  0.50000000         6
#> 98     4  0.54545450         8
#> 99     4  0.59090910         6
#> 100    4  0.63636360         5
#> 101    4  0.68181820         6
#> 102    4  0.72727270         8
#> 103    4  0.77272730        11
#> 104    4  0.81818180        14
#> 105    4  0.86363640         8
#> 106    4  0.90909090         9
#> 107    4  0.95454550        10
#> 108    4  1.00000000         7
#> 109    4  1.04545500         7
#> 110    4  1.09090900         6
#> 111    4  1.13636400        11
#> 112    5 -0.13636360        10
#> 113    5 -0.09090910        12
#> 114    5 -0.04545455        12
#> 115    5  0.00000000        17
#> 116    5  0.04545455         9
#> 117    5  0.09090910        10
#> 118    5  0.13636360         3
#> 119    5  0.18181820        12
#> 120    5  0.22727270        13
#> 121    5  0.27272730         9
#> 122    5  0.31818180         4
#> 123    5  0.36363640         7
#> 124    5  0.40909090         4
#> 125    5  0.45454550        12
#> 126    5  0.50000000        14
#> 127    5  0.54545450        12
#> 128    5  0.59090910        15
#> 129    5  0.63636360        17
#> 130    5  0.68181820        15
#> 131    5  0.72727270        13
#> 132    5  0.77272730        18
#> 133    5  0.81818180        19
#> 134    5  0.86363640        13
#> 135    5  0.90909090         9
#> 136    5  0.95454550        12
#> 137    5  1.00000000         8
#> 138    5  1.04545500        10
#> 139    5  1.09090900         5
#> 140    5  1.13636400        14
#> 141    6 -0.13636360        16
#> 142    6 -0.09090910        17
#> 143    6 -0.04545455        13
#> 144    6  0.00000000        17
#> 145    6  0.04545455        15
#> 146    6  0.09090910         9
#> 147    6  0.13636360         8
#> 148    6  0.18181820         5
#> 149    6  0.22727270         9
#> 150    6  0.27272730         8
#> 151    6  0.31818180         8
#> 152    6  0.36363640        13
#> 153    6  0.40909090        14
#> 154    6  0.45454550        13
#> 155    6  0.50000000        14
#> 156    6  0.54545450        14
#> 157    6  0.59090910        11
#> 158    6  0.63636360        17
#> 159    6  0.68181820        21
#> 160    6  0.72727270        21
#> 161    6  0.77272730        21
#> 162    6  0.81818180        20
#> 163    6  0.86363640        17
#> 164    6  0.90909090        18
#> 165    6  0.95454550        22
#> 166    6  1.00000000        10
#> 167    6  1.04545500        11
#> 168    6  1.09090900        11
#> 169    6  1.13636400        12
#> 170    7 -0.15000000        18
#> 171    7 -0.10000000        13
#> 172    7 -0.05000000        14
#> 173    7  0.00000000        12
#> 174    7  0.05000000        11
#> 175    7  0.10000000         8
#> 176    7  0.15000000         5
#> 177    7  0.20000000         8
#> 178    7  0.25000000        10
#> 179    7  0.30000000        11
#> 180    7  0.35000000        10
#> 181    7  0.40000000        12
#> 182    7  0.45000000        10
#> 183    7  0.50000000         9
#> 184    7  0.55000000        12
#> 185    7  0.60000000        14
#> 186    7  0.65000000        16
#> 187    7  0.70000000        13
#> 188    7  0.75000000        11
#> 189    7  0.80000000        13
#> 190    7  0.85000000        13
#> 191    7  0.90000000        11
#> 192    7  0.95000000        11
#> 193    7  1.00000000         8
#> 194    7  1.05000000        14
#> 195    7  1.10000000         4
#> 196    7  1.15000000         7
#> 197    8 -0.12500000        13
#> 198    8 -0.08333333         9
#> 199    8 -0.04166667        15
#> 200    8  0.00000000        15
#> 201    8  0.04166667        12
#> 202    8  0.08333333         8
#> 203    8  0.12500000        10
#> 204    8  0.16666670         6
#> 205    8  0.20833330         9
#> 206    8  0.25000000         8
#> 207    8  0.29166670        10
#> 208    8  0.33333330         6
#> 209    8  0.37500000         8
#> 210    8  0.41666670        13
#> 211    8  0.45833330        12
#> 212    8  0.50000000        12
#> 213    8  0.54166670        15
#> 214    8  0.58333330        21
#> 215    8  0.62500000        25
#> 216    8  0.66666670        21
#> 217    8  0.70833330        21
#> 218    8  0.75000000        24
#> 219    8  0.79166670        20
#> 220    8  0.83333330        20
#> 221    8  0.87500000        18
#> 222    8  0.91666670        20
#> 223    8  0.95833330        20
#> 224    8  1.00000000        19
#> 225    8  1.04166700        12
#> 226    8  1.08333300         7
#> 227    8  1.12500000         8
#> 228    9 -0.16666670        10
#> 229    9 -0.11111110        14
#> 230    9 -0.05555556        12
#> 231    9  0.00000000        10
#> 232    9  0.05555556         7
#> 233    9  0.11111110        12
#> 234    9  0.16666670        10
#> 235    9  0.22222220         8
#> 236    9  0.27777780        10
#> 237    9  0.33333330        15
#> 238    9  0.38888890        15
#> 239    9  0.44444440        12
#> 240    9  0.50000000        19
#> 241    9  0.55555560        15
#> 242    9  0.61111110        16
#> 243    9  0.66666670        15
#> 244    9  0.72222220        17
#> 245    9  0.77777780        14
#> 246    9  0.83333330        16
#> 247    9  0.88888890        15
#> 248    9  0.94444440        11
#> 249    9  1.00000000        10
#> 250    9  1.05555600         7
#> 251    9  1.11111100         4
#> 252    9  1.16666700         8
#> 253   10 -0.13636360        11
#> 254   10 -0.09090910        16
#> 255   10 -0.04545455        15
#> 256   10  0.00000000        12
#> 257   10  0.04545455        11
#> 258   10  0.09090910         6
#> 259   10  0.13636360        11
#> 260   10  0.18181820        12
#> 261   10  0.22727270        11
#> 262   10  0.27272730        16
#> 263   10  0.31818180        15
#> 264   10  0.36363640        11
#> 265   10  0.40909090         7
#> 266   10  0.45454550        14
#> 267   10  0.50000000        20
#> 268   10  0.54545450        22
#> 269   10  0.59090910        23
#> 270   10  0.63636360        21
#> 271   10  0.68181820        21
#> 272   10  0.72727270        23
#> 273   10  0.77272730        22
#> 274   10  0.81818180        22
#> 275   10  0.86363640        17
#> 276   10  0.90909090        17
#> 277   10  0.95454550        17
#> 278   10  1.00000000        17
#> 279   10  1.04545500        14
#> 280   10  1.09090900        12
#> 281   10  1.13636400        11
#> 282   11 -0.15000000         9
#> 283   11 -0.10000000         8
#> 284   11 -0.05000000         8
#> 285   11  0.00000000         8
#> 286   11  0.05000000         8
#> 287   11  0.10000000         6
#> 288   11  0.15000000         7
#> 289   11  0.20000000         8
#> 290   11  0.25000000        10
#> 291   11  0.30000000        10
#> 292   11  0.35000000        14
#> 293   11  0.40000000        13
#> 294   11  0.45000000         8
#> 295   11  0.50000000         8
#> 296   11  0.55000000         8
#> 297   11  0.60000000         9
#> 298   11  0.65000000        16
#> 299   11  0.70000000        12
#> 300   11  0.75000000        10
#> 301   11  0.80000000        12
#> 302   11  0.85000000        12
#> 303   11  0.90000000         9
#> 304   11  0.95000000         6
#> 305   11  1.00000000         9
#> 306   11  1.05000000         7
#> 307   11  1.10000000         5
#> 308   11  1.15000000         5