Extract Groups from a Data Frame

Each variable named in the expression after the | operator on the right hand side of form is evaluated in object. If more than one variable is indicated in level they are combined into a data frame; else the selected variable is returned as a vector. When multiple grouping levels are defined in form and level > 1, the levels of the returned factor are obtained by pasting together the levels of the grouping factors of level greater or equal to level, to ensure their uniqueness.

Usage

# S3 method for data.frame
getGroups(object, form, level, data, sep)

Arguments

object

an object inheriting from class data.frame.

form

an optional formula with a conditioning expression on its right hand side (i.e. an expression involving the | operator). Defaults to formula(object).

level

a positive integer vector with the level(s) of grouping to be used when multiple nested levels of grouping are present. Defaults to all levels of nesting.

data

unused

sep

character, the separator to use between group levels when multiple levels are collapsed. The default is '/'.

Value

either a data frame with columns given by the grouping factors indicated in level, from outer to inner, or, when a single level is requested, a factor representing the selected grouping factor.

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer, esp. pp. 100, 461.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

getGroups(Pixel)
#>     Dog Side
#> 1     1    R
#> 2     1    R
#> 3     1    R
#> 4     1    R
#> 5     1    R
#> 6     1    R
#> 7     1    R
#> 8     2    R
#> 9     2    R
#> 10    2    R
#> 11    2    R
#> 12    2    R
#> 13    2    R
#> 14    2    R
#> 15    3    R
#> 16    3    R
#> 17    3    R
#> 18    3    R
#> 19    3    R
#> 20    3    R
#> 21    3    R
#> 22    4    R
#> 23    4    R
#> 24    4    R
#> 25    4    R
#> 26    4    R
#> 27    4    R
#> 28    4    R
#> 29    5    R
#> 30    5    R
#> 31    5    R
#> 32    5    R
#> 33    5    R
#> 34    6    R
#> 35    6    R
#> 36    6    R
#> 37    6    R
#> 38    6    R
#> 39    7    R
#> 40    7    R
#> 41    7    R
#> 42    7    R
#> 43    8    R
#> 44    8    R
#> 45    8    R
#> 46    8    R
#> 47    9    R
#> 48    9    R
#> 49   10    R
#> 50   10    R
#> 51   10    R
#> 52    1    L
#> 53    1    L
#> 54    1    L
#> 55    1    L
#> 56    1    L
#> 57    1    L
#> 58    1    L
#> 59    2    L
#> 60    2    L
#> 61    2    L
#> 62    2    L
#> 63    2    L
#> 64    2    L
#> 65    2    L
#> 66    3    L
#> 67    3    L
#> 68    3    L
#> 69    3    L
#> 70    3    L
#> 71    3    L
#> 72    3    L
#> 73    4    L
#> 74    4    L
#> 75    4    L
#> 76    4    L
#> 77    4    L
#> 78    4    L
#> 79    4    L
#> 80    5    L
#> 81    5    L
#> 82    5    L
#> 83    5    L
#> 84    5    L
#> 85    6    L
#> 86    6    L
#> 87    6    L
#> 88    6    L
#> 89    6    L
#> 90    7    L
#> 91    7    L
#> 92    7    L
#> 93    7    L
#> 94    8    L
#> 95    8    L
#> 96    8    L
#> 97    8    L
#> 98    9    L
#> 99    9    L
#> 100  10    L
#> 101  10    L
#> 102  10    L
getGroups(Pixel, level = 2)
#>   [1] 1/R  1/R  1/R  1/R  1/R  1/R  1/R  2/R  2/R  2/R  2/R  2/R  2/R  2/R  3/R 
#>  [16] 3/R  3/R  3/R  3/R  3/R  3/R  4/R  4/R  4/R  4/R  4/R  4/R  4/R  5/R  5/R 
#>  [31] 5/R  5/R  5/R  6/R  6/R  6/R  6/R  6/R  7/R  7/R  7/R  7/R  8/R  8/R  8/R 
#>  [46] 8/R  9/R  9/R  10/R 10/R 10/R 1/L  1/L  1/L  1/L  1/L  1/L  1/L  2/L  2/L 
#>  [61] 2/L  2/L  2/L  2/L  2/L  3/L  3/L  3/L  3/L  3/L  3/L  3/L  4/L  4/L  4/L 
#>  [76] 4/L  4/L  4/L  4/L  5/L  5/L  5/L  5/L  5/L  6/L  6/L  6/L  6/L  6/L  7/L 
#>  [91] 7/L  7/L  7/L  8/L  8/L  8/L  8/L  9/L  9/L  10/L 10/L 10/L
#> 20 Levels: 1/R < 2/R < 3/R < 4/R < 5/R < 6/R < 7/R < 8/R < 9/R < ... < 10/L