Extract lme Object Groups

The grouping factors corresponding to the linear mixed-effects model represented by object are extracted. If more than one level is indicated in level, the corresponding grouping factors are combined into a data frame; else the selected grouping factor is returned as a vector.

Usage

# S3 method for lme
getGroups(object, form, level, data, sep)

Arguments

object

an object inheriting from class lme, representing a fitted linear mixed-effects model.

form

this argument is included to make the method function compatible with the generic and is ignored in this method.

level

an optional integer vector giving the level(s) of grouping to be extracted from object. Defaults to the highest or innermost level of grouping.

data

unused

sep

character, the separator to use between group levels when multiple levels are collapsed. The default is '/'.

Value

either a data frame with columns given by the grouping factors indicated in level, or, when a single level is requested, a factor representing the selected grouping factor.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

fm1 <- lme(pixel ~ day + day^2, Pixel,
  random = list(Dog = ~day, Side = ~1))
getGroups(fm1, level = 1:2)
#>     Dog Side
#> 1     1  1/R
#> 2     1  1/R
#> 3     1  1/R
#> 4     1  1/R
#> 5     1  1/R
#> 6     1  1/R
#> 7     1  1/R
#> 8     2  2/R
#> 9     2  2/R
#> 10    2  2/R
#> 11    2  2/R
#> 12    2  2/R
#> 13    2  2/R
#> 14    2  2/R
#> 15    3  3/R
#> 16    3  3/R
#> 17    3  3/R
#> 18    3  3/R
#> 19    3  3/R
#> 20    3  3/R
#> 21    3  3/R
#> 22    4  4/R
#> 23    4  4/R
#> 24    4  4/R
#> 25    4  4/R
#> 26    4  4/R
#> 27    4  4/R
#> 28    4  4/R
#> 29    5  5/R
#> 30    5  5/R
#> 31    5  5/R
#> 32    5  5/R
#> 33    5  5/R
#> 34    6  6/R
#> 35    6  6/R
#> 36    6  6/R
#> 37    6  6/R
#> 38    6  6/R
#> 39    7  7/R
#> 40    7  7/R
#> 41    7  7/R
#> 42    7  7/R
#> 43    8  8/R
#> 44    8  8/R
#> 45    8  8/R
#> 46    8  8/R
#> 47    9  9/R
#> 48    9  9/R
#> 49   10 10/R
#> 50   10 10/R
#> 51   10 10/R
#> 52    1  1/L
#> 53    1  1/L
#> 54    1  1/L
#> 55    1  1/L
#> 56    1  1/L
#> 57    1  1/L
#> 58    1  1/L
#> 59    2  2/L
#> 60    2  2/L
#> 61    2  2/L
#> 62    2  2/L
#> 63    2  2/L
#> 64    2  2/L
#> 65    2  2/L
#> 66    3  3/L
#> 67    3  3/L
#> 68    3  3/L
#> 69    3  3/L
#> 70    3  3/L
#> 71    3  3/L
#> 72    3  3/L
#> 73    4  4/L
#> 74    4  4/L
#> 75    4  4/L
#> 76    4  4/L
#> 77    4  4/L
#> 78    4  4/L
#> 79    4  4/L
#> 80    5  5/L
#> 81    5  5/L
#> 82    5  5/L
#> 83    5  5/L
#> 84    5  5/L
#> 85    6  6/L
#> 86    6  6/L
#> 87    6  6/L
#> 88    6  6/L
#> 89    6  6/L
#> 90    7  7/L
#> 91    7  7/L
#> 92    7  7/L
#> 93    7  7/L
#> 94    8  8/L
#> 95    8  8/L
#> 96    8  8/L
#> 97    8  8/L
#> 98    9  9/L
#> 99    9  9/L
#> 100  10 10/L
#> 101  10 10/L
#> 102  10 10/L