Extract varFunc Groups

This method function extracts the grouping factor associated with the variance function represented by object, if any is present.

Usage

# S3 method for varFunc
getGroups(object, form, level, data, sep)

Arguments

object

an object inheriting from class varFunc, representing a variance function structure.

form

an optional formula with a conditioning expression on its right hand side (i.e. an expression involving the | operator). Defaults to formula(object). Not used.

level

a positive integer vector with the level(s) of grouping to be used when multiple nested levels of grouping are present. This argument is optional for most methods of this generic function and defaults to all levels of nesting. Not used.

data

a data frame in which to interpret the variables named in form. Optional for most methods. Not used.

sep

character, the separator to use between group levels when multiple levels are collapsed. The default is '/'. Not used.

Value

if object has a groups attribute, its value is returned; else NULL is returned.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

Examples

vf1 <- varPower(form = ~ age | Sex)
vf1 <- Initialize(vf1, Orthodont)
getGroups(vf1)
#>   [1] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>   [9] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [17] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [25] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [33] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [41] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [49] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [57] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [65] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#>  [73] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#>  [81] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#>  [89] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#>  [97] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [105] "Female" "Female" "Female" "Female"