getGroupsFormula {nlme}	R Documentation

Extract Grouping Formula

Description

The conditioning expression associated with formula(object) (i.e. the expression after the | operator) is returned either as a named list of one-sided formulas, or a single one-sided formula, depending on the value of asList. The components of the returned list are ordered from outermost to innermost level and are named after the grouping factor expression.

Usage

getGroupsFormula(object, asList, sep)

Arguments

object
any object from which a formula can be extracted.

asList
an optional logical value. If TRUE the returned value with be a list of formulas; else, if FALSE the returned value will be a one-sided formula. Defaults to FALSE.

sep
character, the separator to use between group levels when multiple levels are collapsed. The default is '/'.

Value

a one-sided formula, or a list of one-sided formulas, with the grouping structure associated with formula(object). If no conditioning expression is present in formula(object) a NULL value is returned.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

getGroupsFormula.gls, getGroupsFormula.lmList, getGroupsFormula.lme, getGroupsFormula.reStruct, getGroups

Examples

getGroupsFormula(y ~ x | g1/g2)

[Package nlme version 3.1-121 Index]