getResponse {nlme}	R Documentation

Extract Response Variable from an Object

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include data.frame, gls, lme, and lmList.

Usage


getResponse(object, form)

Arguments

object
any object

form
an optional two-sided formula. Defaults to formula(object).

Value

will depend on the method function used; see the appropriate documentation.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

getResponseFormula

Examples


getResponse(Orthodont)

[Package nlme version 3.1-137 Index]