gls-internal {nlme}	R Documentation

Auxiliary functions used by gls

Description

These are functions used by gls to call its compiled C code. They are exported to allow experimentation with modified versions.

Usage


glsApVar(glsSt, sigma, conLin = attr(glsSt, "conLin"),
         .relStep = .Machine$double.eps^(1/3), minAbsPar = 0, natural = TRUE)
glsEstimate(object, conLin = attr(object, "conLin"),
            control = list(singular.ok = FALSE))

Arguments

glsSt, object
An object inheriting from class "glsStruct".

sigma
the estimated residual standard error: see glsObject.

conLin
A ‘condensed linear model’: see logLik.glsStruct.

.relStep, minAbsPar, natural
Control values: see glsControl.

control
The relevant part of a glsControl return value.

[Package nlme version 3.1-137 Index]