glsStruct {nlme}	R Documentation

Generalized Least Squares Structure

Description

A generalized least squares structure is a list of model components representing different sets of parameters in the linear model. A glsStruct may contain corStruct and varFunc objects. NULL arguments are not included in the glsStruct list.

Usage


glsStruct(corStruct, varStruct)

Arguments

corStruct
an optional corStruct object, representing a correlation structure. Default is NULL.

varStruct
an optional varFunc object, representing a variance function structure. Default is NULL.

Value

a list of model variance-covariance components determining the parameters to be estimated for the associated linear model.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

corClasses, gls, residuals.glsStruct, varFunc

Examples


gls1 <- glsStruct(corAR1(), varPower())

[Package nlme version 3.1-137 Index]