gnlsControl {nlme}	R Documentation

Control Values for gnls Fit

Description

The values supplied in the function call replace the defaults and a list with all possible arguments is returned. The returned list is used as the control argument to the gnls function.

Usage


gnlsControl(maxIter = 50, nlsMaxIter = 7, msMaxIter = 50, minScale = 0.001,
            tolerance = 1e-6, nlsTol = 0.001, msTol = 1e-7,
            returnObject = FALSE, msVerbose = FALSE,
            apVar = TRUE, .relStep =,
            opt = c("nlminb", "optim"), optimMethod = "BFGS",
            minAbsParApVar = 0.05, sigma = NULL)

Arguments

maxIter
maximum number of iterations for the gnls optimization algorithm. Default is 50.

nlsMaxIter
maximum number of iterations for the nls optimization step inside the gnls optimization. Default is 7.

msMaxIter
maximum number of iterations for the ms optimization step inside the gnls optimization. Default is 50.

minScale
minimum factor by which to shrink the default step size in an attempt to decrease the sum of squares in the nls step. Default 0.001.

tolerance
tolerance for the convergence criterion in the gnls algorithm. Default is 1e-6.

nlsTol
tolerance for the convergence criterion in nls step. Default is 1e-3.

msTol
tolerance for the convergence criterion of the first outer iteration when optim is used. Default is 1e-7.

returnObject
a logical value indicating whether the fitted object should be returned with a warning (instead of an error via stop()) when the maximum number of iterations is reached without convergence of the algorithm.

msVerbose
a logical value passed as the trace argument to the optimizer chosen by opt; see documentation on that. Default is FALSE.

apVar
a logical value indicating whether the approximate covariance matrix of the variance-covariance parameters should be calculated. Default is TRUE.

.relStep
relative step for numerical derivatives calculations. Default is .Machine$double.eps^(1/3) (about 6e-6).

opt
the optimizer to be used, either "nlminb" (the current default) or "optim" (the previous default).

optimMethod
character - the optimization method to be used with the optim optimizer. The default is "BFGS". An alternative is "L-BFGS-B".

minAbsParApVar
numeric value - minimum absolute parameter value in the approximate variance calculation. The default is 0.05.

sigma
optionally a positive number to fix the residual error at. If NULL, as by default, or 0, sigma is estimated.

Value

a list with components for each of the possible arguments.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu; the sigma option: Siem Heisterkamp and Bert van Willigen.

See Also

gnls

Examples


# decrease the maximum number iterations in the ms call and
# request that information on the evolution of the ms iterations be printed
gnlsControl(msMaxIter = 20, msVerbose = TRUE)

[Package nlme version 3.1-137 Index]