gnlsStruct {nlme}	R Documentation

Generalized Nonlinear Least Squares Structure

Description

A generalized nonlinear least squares structure is a list of model components representing different sets of parameters in the nonlinear model. A gnlsStruct may contain corStruct and varFunc objects. NULL arguments are not included in the gnlsStruct list.

Usage

gnlsStruct(corStruct, varStruct)

Arguments

corStruct
an optional corStruct object, representing a correlation structure. Default is NULL.

varStruct
an optional varFunc object, representing a variance function structure. Default is NULL.

Value

a list of model variance-covariance components determining the parameters to be estimated for the associated nonlinear model.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

gnls, corClasses, residuals.gnlsStruct varFunc

Examples

gnls1 <- gnlsStruct(corAR1(), varPower())

[Package nlme version 3.1-121 Index]