intervals {nlme}	R Documentation

Confidence Intervals on Coefficients

Description

Confidence intervals on the parameters associated with the model represented by object are obtained. This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include: gls, lme, and lmList.

Usage

intervals(object, level, ...)

Arguments

object
a fitted model object from which parameter estimates can be extracted.

level
an optional numeric value for the interval confidence level. Defaults to 0.95.

...
some methods for the generic may require additional arguments.

Value

will depend on the method function used; see the appropriate documentation.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

intervals.lme, intervals.lmList, intervals.gls

Examples

## see the method documentation

[Package nlme version 3.1-121 Index]