Confidence Intervals on lmList Coefficients

Confidence intervals on the linear model coefficients are obtained for each lm component of object and organized into a three dimensional array. The first dimension corresponding to the names of the object components. The second dimension is given by lower, est., and upper corresponding, respectively, to the lower confidence limit, estimated coefficient, and upper confidence limit. The third dimension is given by the coefficients names.

Usage

# S3 method for lmList
intervals(object, level = 0.95, pool = attr(object, "pool"), ...)

Arguments

object

an object inheriting from class "lmList", representing a list of lm objects with a common model.

level

an optional numeric value with the confidence level for the intervals. Defaults to 0.95.

pool

an optional logical value indicating whether a pooled estimate of the residual standard error should be used. Default is attr(object, "pool").

...

some methods for this generic require additional arguments. None are used in this method.

Value

a three dimensional array with the confidence intervals and estimates for the coefficients of each lm component of object.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See also

Examples

fm1 <- lmList(distance ~ age | Subject, Orthodont)
intervals(fm1)
#> , , (Intercept)
#> 
#>        lower  est.    upper
#> M16 10.35761 16.95 23.54239
#> M05  7.05761 13.65 20.24239
#> M02  8.25761 14.85 21.44239
#> M11 13.45761 20.05 26.64239
#> M07  8.35761 14.95 21.54239
#> M08 13.15761 19.75 26.34239
#> M03  9.40761 16.00 22.59239
#> M12  6.65761 13.25 19.84239
#> M13 -3.79239  2.80  9.39239
#> M14 12.50761 19.10 25.69239
#> M09  7.80761 14.40 20.99239
#> M15  6.90761 13.50 20.09239
#> M06 12.35761 18.95 25.54239
#> M04 18.10761 24.70 31.29239
#> M01 10.70761 17.30 23.89239
#> M10 14.65761 21.25 27.84239
#> F10  6.95761 13.55 20.14239
#> F09 11.50761 18.10 24.69239
#> F06 10.40761 17.00 23.59239
#> F01 10.65761 17.25 23.84239
#> F05 13.00761 19.60 26.19239
#> F07 10.35761 16.95 23.54239
#> F02  7.60761 14.20 20.79239
#> F08 14.85761 21.45 28.04239
#> F03  7.80761 14.40 20.99239
#> F04 13.05761 19.65 26.24239
#> F11 12.35761 18.95 25.54239
#> 
#> , , age
#> 
#>           lower  est.     upper
#> M16 -0.03729682 0.550 1.1372968
#> M05  0.26270318 0.850 1.4372968
#> M02  0.18770318 0.775 1.3622968
#> M11 -0.26229682 0.325 0.9122968
#> M07  0.21270318 0.800 1.3872968
#> M08 -0.21229682 0.375 0.9622968
#> M03  0.16270318 0.750 1.3372968
#> M12  0.41270318 1.000 1.5872968
#> M13  1.36270318 1.950 2.5372968
#> M14 -0.06229682 0.525 1.1122968
#> M09  0.38770318 0.975 1.5622968
#> M15  0.53770318 1.125 1.7122968
#> M06  0.08770318 0.675 1.2622968
#> M04 -0.41229682 0.175 0.7622968
#> M01  0.36270318 0.950 1.5372968
#> M10  0.16270318 0.750 1.3372968
#> F10 -0.13729682 0.450 1.0372968
#> F09 -0.31229682 0.275 0.8622968
#> F06 -0.21229682 0.375 0.9622968
#> F01 -0.21229682 0.375 0.9622968
#> F05 -0.31229682 0.275 0.8622968
#> F07 -0.03729682 0.550 1.1372968
#> F02  0.21270318 0.800 1.3872968
#> F08 -0.41229682 0.175 0.7622968
#> F03  0.26270318 0.850 1.4372968
#> F04 -0.11229682 0.475 1.0622968
#> F11  0.08770318 0.675 1.2622968
#>