lmList {nlme}	R Documentation

List of lm Objects with a Common Model

Description

Data is partitioned according to the levels of the grouping factor g and individual lm fits are obtained for each data partition, using the model defined in object.

Usage


lmList(object, data, level, subset, na.action = na.fail,
       pool = TRUE, warn.lm = TRUE)
## S3 method for class 'lmList'
update(object, formula., ..., evaluate = TRUE)
## S3 method for class 'lmList'
print(x, pool, ...)

Arguments

object
For lmList, either a linear formula object of the form y ~ x1+...+xn | g or a groupedData object. In the formula object, y represents the response, x1,...,xn the covariates, and g the grouping factor specifying the partitioning of the data according to which different lm fits should be performed. The grouping factor g may be omitted from the formula, in which case the grouping structure will be obtained from data, which must inherit from class groupedData. The method function lmList.groupedData is documented separately. For the method update.lmList, object is an object inheriting from class lmList.

formula
(used in update.lmList only) a two-sided linear formula with the common model for the individuals lm fits.

formula.
Changes to the formula – see update.formula for details.

data
a data frame in which to interpret the variables named in object.

level
an optional integer specifying the level of grouping to be used when multiple nested levels of grouping are present.

subset
an optional expression indicating which subset of the rows of data should be used in the fit. This can be a logical vector, or a numeric vector indicating which observation numbers are to be included, or a character vector of the row names to be included. All observations are included by default.

na.action
a function that indicates what should happen when the data contain NAs. The default action (na.fail) causes lmList to print an error message and terminate if there are any incomplete observations.

pool
an optional logical value indicating whether a pooled estimate of the residual standard error should be used in calculations of standard deviations or standard errors for summaries.

warn.lm
logical indicating if lm() errors (all of which are caught by tryCatch) should be signalled as a “summarizing” warning.

x
an object inheriting from class lmList to be printed.

...
some methods for this generic require additional arguments. None are used in this method.

evaluate
If TRUE evaluate the new call else return the call.

Value

a list of lm objects with as many components as the number of groups defined by the grouping factor. Generic functions such as coef, fixed.effects, lme, pairs, plot, predict, random.effects, summary, and update have methods that can be applied to an lmList object.

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

lm, lme.lmList, plot.lmList, pooledSD, predict.lmList, residuals.lmList, summary.lmList

Examples


fm1 <- lmList(distance ~ age | Subject, Orthodont)
summary(fm1)

[Package nlme version 3.1-137 Index]