lmeScale {nlme}	R Documentation

Scale for lme Optimization

Description

This function calculates the scales to be used for each coefficient estimated through an nlm optimization in the lme function. If all initial values are zero, the scale is set to one for all coefficients; else, the scale for a coefficient with non-zero initial value is equal to the inverse of its initial value and the scale for the coefficients with initial value equal to zero is set to the median of the non-zero initial value coefficients.

Usage

lmeScale(start)

Arguments

start 	 the starting values for the coefficients to be estimated.

Value

a vector with the scales to be used in nlm for estimating the coefficients.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

nlm

[Package nlme version 3.1-93 Index]