lmeStruct {nlme}	R Documentation

Linear Mixed-Effects Structure

Description

A linear mixed-effects structure is a list of model components representing different sets of parameters in the linear mixed-effects model. An lmeStruct list must contain at least a reStruct object, but may also contain corStruct and varFunc objects. NULL arguments are not included in the lmeStruct list.

Usage

lmeStruct(reStruct, corStruct, varStruct)

Arguments

reStruct
a reStruct representing a random effects structure.

corStruct
an optional corStruct object, representing a correlation structure. Default is NULL.

varStruct
an optional varFunc object, representing a variance function structure. Default is NULL.

Value

a list of model components determining the parameters to be estimated for the associated linear mixed-effects model.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

corClasses, lme, residuals.lmeStruct, reStruct, varFunc

Examples

lms1 <- lmeStruct(reStruct(~age), corAR1(), varPower())

[Package nlme version 3.1-127 Index]