logDet {nlme}	R Documentation

Extract the Logarithm of the Determinant

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include: corStruct, several pdMat classes, and reStruct.

Usage

logDet(object, ...)

Arguments

object
any object from which a matrix, or list of matrices, can be extracted

...
some methods for this generic function require additional arguments.

Value

will depend on the method function used; see the appropriate documentation.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

logLik, logDet.corStruct, logDet.pdMat, logDet.reStruct

Examples

## see the method function documentation

[Package nlme version 3.1-121 Index]