logDet.corStruct {nlme}	R Documentation

Extract corStruct Log-Determinant

Description

This method function extracts the logarithm of the determinant of a square-root factor of the correlation matrix associated with object, or the sum of the log-determinants of square-root factors of the list of correlation matrices associated with object.

Usage

## S3 method for class 'corStruct'
logDet(object, covariate, ...)

Arguments

object
an object inheriting from class "corStruct", representing a correlation structure.

covariate
an optional covariate vector (matrix), or list of covariate vectors (matrices), at which values the correlation matrix, or list of correlation matrices, are to be evaluated. Defaults to getCovariate(object).

...
some methods for this generic require additional arguments. None are used in this method.

Value

the log-determinant of a square-root factor of the correlation matrix associated with object, or the sum of the log-determinants of square-root factors of the list of correlation matrices associated with object.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

logLik.corStruct, corMatrix.corStruct, logDet

Examples

cs1 <- corAR1(0.3)
logDet(cs1, covariate = 1:4)

[Package nlme version 3.1-127 Index]