logDet.reStruct {nlme}	R Documentation

Extract reStruct Log-Determinants

Description

Calculates, for each of the pdMat components of object, the logarithm of the determinant of a square-root factor.

Usage


## S3 method for class 'reStruct'
logDet(object, ...)

Arguments

object
an object inheriting from class "reStruct", representing a random effects structure and consisting of a list of pdMat objects.

...
some methods for this generic require additional arguments. None are used in this method.

Value

a vector with the log-determinants of square-root factors of the pdMat components of object.

Author(s)

José Pinheiro

See Also

reStruct, pdMat, logDet

Examples


rs1 <- reStruct(list(A = pdSymm(diag(1:3), form = ~Score),
  B = pdDiag(2 * diag(4), form = ~Educ)))
logDet(rs1)

[Package nlme version 3.1-137 Index]