logLik.corStruct {nlme}	R Documentation

Extract corStruct Log-Likelihood

Description

This method function extracts the component of a Gaussian log-likelihood associated with the correlation structure, which is equal to the negative of the logarithm of the determinant (or sum of the logarithms of the determinants) of the matrix (or matrices) represented by object.

Usage

## S3 method for class 'corStruct'
logLik(object, data, ...)

Arguments

object
an object inheriting from class "corStruct", representing a correlation structure.

data
this argument is included to make this method function compatible with other logLik methods and will be ignored.

...
some methods for this generic require additional arguments. None are used in this method.

Value

the negative of the logarithm of the determinant (or sum of the logarithms of the determinants) of the correlation matrix (or matrices) represented by object.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

logDet.corStruct, logLik.lme,

Examples

cs1 <- corAR1(0.2)
cs1 <- Initialize(cs1, data = Orthodont)
logLik(cs1)

[Package nlme version 3.1-127 Index]