logLik.glsStruct {nlme}	R Documentation

Log-Likelihood of a glsStruct Object

Description

Pars is used to update the coefficients of the model components of object and the individual (restricted) log-likelihood contributions of each component are added together. The type of log-likelihood (restricted or not) is determined by the settings attribute of object.

Usage

## S3 method for class 'glsStruct'
logLik(object, Pars, conLin, ...)

Arguments

object
an object inheriting from class "glsStruct", representing a list of linear model components, such as corStruct and "varFunc" objects.

Pars
the parameter values at which the (restricted) log-likelihood is to be evaluated.

conLin
an optional condensed linear model object, consisting of a list with components "Xy", corresponding to a regression matrix (X) combined with a response vector (y), and "logLik", corresponding to the log-likelihood of the underlying linear model. Defaults to attr(object, "conLin").

...
some methods for this generic require additional arguments. None are used in this method.

Value

the (restricted) log-likelihood for the linear model described by object, evaluated at Pars.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

gls, glsStruct, logLik.lme

[Package nlme version 3.1-127 Index]