logLik.gnls {nlme}	R Documentation

Log-Likelihood of a gnls Object

Description

Returns the log-likelihood value of the nonlinear model represented by object evaluated at the estimated coefficients.

Usage

## S3 method for class 'gnls'
logLik(object, REML, ...)

Arguments

object
an object inheriting from class "gnls", representing a generalized nonlinear least squares fitted model.

REML
an logical value for consistency with logLik,gls, but only FALSE is accepted..

...
some methods for this generic require additional arguments. None are used in this method.

Value

the log-likelihood of the linear model represented by object evaluated at the estimated coefficients.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

gnls, logLik.lme

Examples

fm1 <- gnls(weight ~ SSlogis(Time, Asym, xmid, scal), Soybean,
            weights = varPower())
logLik(fm1)

[Package nlme version 3.1-127 Index]