logLik.gnlsStruct {nlme}	R Documentation

Log-Likelihood of a gnlsStruct Object

Description

Pars is used to update the coefficients of the model components of object and the individual log-likelihood contributions of each component are added together.

Usage

## S3 method for class 'gnlsStruct':
logLik(object, Pars, conLin, ...)

Arguments

object 	 an object inheriting from class gnlsStruct, representing a list of model components, such as corStruct and varFunc objects, and attributes specifying the underlying nonlinear model and the response variable.

Pars 	 the parameter values at which the log-likelihood is to be evaluated.

conLin 	 an optional condensed linear model object, consisting of a list with components "Xy", corresponding to a regression matrix (X) combined with a response vector (y), and "logLik", corresponding to the log-likelihood of the underlying nonlinear model. Defaults to attr(object, "conLin").

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

the log-likelihood for the linear model described by object, evaluated at Pars.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

gnls, gnlsStruct, logLik.gnls

[Package nlme version 3.1-93 Index]