logLik.reStruct {nlme}	R Documentation

Calculate reStruct Log-Likelihood

Description

Calculates the log-likelihood, or restricted log-likelihood, of the Gaussian linear mixed-effects model represented by object and conLin (assuming spherical within-group covariance structure), evaluated at coef(object). The settings attribute of object determines whether the log-likelihood, or the restricted log-likelihood, is to be calculated. The computational methods are described in Bates and Pinheiro (1998).

Usage


## S3 method for class 'reStruct'
logLik(object, conLin, ...)

Arguments

object
an object inheriting from class "reStruct", representing a random effects structure and consisting of a list of pdMat objects.

conLin
a condensed linear model object, consisting of a list with components "Xy", corresponding to a regression matrix (X) combined with a response vector (y), and "logLik", corresponding to the log-likelihood of the underlying model.

...
some methods for this generic require additional arguments. None are used in this method.

Value

the log-likelihood, or restricted log-likelihood, of linear mixed-effects model represented by object and conLin, evaluated at coef{object}.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

reStruct, pdMat, logLik.lme

[Package nlme version 3.1-137 Index]