reStruct Model Matrix

The model matrices for each element of formula(object), calculated using data, are bound together column-wise. When multiple grouping levels are present (i.e. when length(object) > 1), the individual model matrices are combined from innermost (at the leftmost position) to outermost (at the rightmost position).

Usage

# S3 method for reStruct
model.matrix(object, data, contrast, ...)

Arguments

object

an object inheriting from class "reStruct", representing a random effects structure and consisting of a list of pdMat objects.

data

a data frame in which to evaluate the variables defined in formula(object).

contrast

an optional named list specifying the contrasts to be used for representing the factor variables in data. The components names should match the names of the variables in data for which the contrasts are to be specified. The components of this list will be used as the contrasts attribute of the corresponding factor. If missing, the default contrast specification is used.

...

some methods for this generic require additional arguments. None are used in this method.

Value

a matrix obtained by binding together, column-wise, the model matrices for each element of formula(object).

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

rs1 <- reStruct(list(Dog = ~day, Side = ~1), data = Pixel)
model.matrix(rs1, Pixel)
#>     Side.(Intercept) Dog.(Intercept) Dog.day
#> 1                  1               1       0
#> 2                  1               1       1
#> 3                  1               1       2
#> 4                  1               1       4
#> 5                  1               1       6
#> 6                  1               1      10
#> 7                  1               1      14
#> 8                  1               1       0
#> 9                  1               1       1
#> 10                 1               1       2
#> 11                 1               1       4
#> 12                 1               1       6
#> 13                 1               1      10
#> 14                 1               1      14
#> 15                 1               1       0
#> 16                 1               1       4
#> 17                 1               1       6
#> 18                 1               1       8
#> 19                 1               1      10
#> 20                 1               1      14
#> 21                 1               1      21
#> 22                 1               1       0
#> 23                 1               1       4
#> 24                 1               1       6
#> 25                 1               1       8
#> 26                 1               1      10
#> 27                 1               1      14
#> 28                 1               1      21
#> 29                 1               1       4
#> 30                 1               1       6
#> 31                 1               1       8
#> 32                 1               1      10
#> 33                 1               1      14
#> 34                 1               1       4
#> 35                 1               1       6
#> 36                 1               1       8
#> 37                 1               1      10
#> 38                 1               1      14
#> 39                 1               1       4
#> 40                 1               1       6
#> 41                 1               1      10
#> 42                 1               1      14
#> 43                 1               1       4
#> 44                 1               1       6
#> 45                 1               1      10
#> 46                 1               1      14
#> 47                 1               1       4
#> 48                 1               1       8
#> 49                 1               1       4
#> 50                 1               1       6
#> 51                 1               1       8
#> 52                 1               1       0
#> 53                 1               1       1
#> 54                 1               1       2
#> 55                 1               1       4
#> 56                 1               1       6
#> 57                 1               1      10
#> 58                 1               1      14
#> 59                 1               1       0
#> 60                 1               1       1
#> 61                 1               1       2
#> 62                 1               1       4
#> 63                 1               1       6
#> 64                 1               1      10
#> 65                 1               1      14
#> 66                 1               1       0
#> 67                 1               1       4
#> 68                 1               1       6
#> 69                 1               1       8
#> 70                 1               1      10
#> 71                 1               1      14
#> 72                 1               1      21
#> 73                 1               1       0
#> 74                 1               1       4
#> 75                 1               1       6
#> 76                 1               1       8
#> 77                 1               1      10
#> 78                 1               1      14
#> 79                 1               1      21
#> 80                 1               1       4
#> 81                 1               1       6
#> 82                 1               1       8
#> 83                 1               1      10
#> 84                 1               1      14
#> 85                 1               1       4
#> 86                 1               1       6
#> 87                 1               1       8
#> 88                 1               1      10
#> 89                 1               1      14
#> 90                 1               1       4
#> 91                 1               1       6
#> 92                 1               1      10
#> 93                 1               1      14
#> 94                 1               1       4
#> 95                 1               1       6
#> 96                 1               1      10
#> 97                 1               1      14
#> 98                 1               1       4
#> 99                 1               1       8
#> 100                1               1       4
#> 101                1               1       6
#> 102                1               1       8
#> attr(,"ncols")
#> Side  Dog 
#>    1    2 
#> attr(,"nams")
#> attr(,"nams")$Side
#> [1] "(Intercept)"
#> 
#> attr(,"nams")$Dog
#> [1] "(Intercept)" "day"        
#> 
#> attr(,"contr")
#> attr(,"contr")$day
#> NULL
#>