nlme-deprecated {nlme}	R Documentation

Deprecated Functions in Package nlme

Description

These functions are provided for compatibility with older versions of nlme only, and may be defunct as soon as the next release.

Usage

## internal functions wrapped by groupedData()
nfGroupedData(formula, data, order.groups, FUN, outer, inner,
              labels, units)
nmGroupedData(formula, data, order.groups, FUN, outer, inner,
              labels, units)

## trivial "identity" correlation, never usefully implemented:
corIdent(form = NULL)

Examples

assertDeprecation <- function(expr)
  tools::assertCondition(expr, verbose = TRUE,
    if(getRversion() >= "3.6.0") "deprecatedWarning" else "warning")

assertDeprecation(
  nlme::nfGroupedData(height ~ age | Subject, as.data.frame(Oxboys))
)
assertDeprecation( csId <-  corIdent(~ 1 | Subject) )
assertDeprecation( csI. <- Initialize(csId, data = Orthodont) )
assertDeprecation( corMatrix(csI.) )  # actually errors

[Package nlme version 3.1-162 Index]