NLME fit from nlsList Object

If the random effects names defined in random are a subset of the lmList object coefficient names, initial estimates for the covariance matrix of the random effects are obtained (overwriting any values given in random). formula(fixed) and the data argument in the calling sequence used to obtain fixed are passed as the fixed and data arguments to nlme.formula, together with any other additional arguments in the function call. See the documentation on nlme.formula for a description of that function.

Usage

# S3 method for nlsList
nlme(model, data, fixed, random, groups, start, correlation, weights,
     subset, method, na.action, naPattern, control, verbose)

Arguments

model

an object inheriting from class "nlsList", representing a list of nls fits with a common model.

data

this argument is included for consistency with the generic function. It is ignored in this method function.

fixed

this argument is included for consistency with the generic function. It is ignored in this method function.

random

an optional one-sided linear formula with no conditioning expression, or a pdMat object with a formula attribute. Multiple levels of grouping are not allowed with this method function. Defaults to a formula consisting of the right hand side of formula(fixed).

groups

an optional one-sided formula of the form ~g1 (single level of nesting) or ~g1/.../gQ (multiple levels of nesting), specifying the partitions of the data over which the random effects vary. g1,...,gQ must evaluate to factors in data. The order of nesting, when multiple levels are present, is taken from left to right (i.e. g1 is the first level, g2 the second, etc.).

start

an optional numeric vector, or list of initial estimates for the fixed effects and random effects. If declared as a numeric vector, it is converted internally to a list with a single component fixed, given by the vector. The fixed component is required, unless the model function inherits from class selfStart, in which case initial values will be derived from a call to nlsList. An optional random component is used to specify initial values for the random effects and should consist of a matrix, or a list of matrices with length equal to the number of grouping levels. Each matrix should have as many rows as the number of groups at the corresponding level and as many columns as the number of random effects in that level.

correlation

an optional corStruct object describing the within-group correlation structure. See the documentation of corClasses for a description of the available corStruct classes. Defaults to NULL, corresponding to no within-group correlations.

weights

an optional varFunc object or one-sided formula describing the within-group heteroscedasticity structure. If given as a formula, it is used as the argument to varFixed, corresponding to fixed variance weights. See the documentation on varClasses for a description of the available varFunc classes. Defaults to NULL, corresponding to homoscedastic within-group errors.

subset

an optional expression indicating the subset of the rows of data that should be used in the fit. This can be a logical vector, or a numeric vector indicating which observation numbers are to be included, or a character vector of the row names to be included. All observations are included by default.

method

a character string. If "REML" the model is fit by maximizing the restricted log-likelihood. If "ML" the log-likelihood is maximized. Defaults to "ML".

na.action

a function that indicates what should happen when the data contain NAs. The default action (na.fail) causes nlme to print an error message and terminate if there are any incomplete observations.

naPattern

an expression or formula object, specifying which returned values are to be regarded as missing.

control

a list of control values for the estimation algorithm to replace the default values returned by the function nlmeControl. Defaults to an empty list.

verbose

an optional logical value. If TRUE information on the evolution of the iterative algorithm is printed. Default is FALSE.

Value

an object of class nlme representing the linear mixed-effects model fit. Generic functions such as print, plot and

summary have methods to show the results of the fit. See

nlmeObject for the components of the fit. The functions

resid, coef, fitted, fixed.effects, and

random.effects can be used to extract some of its components.

References

The computational methods follow on the general framework of Lindstrom, M.J. and Bates, D.M. (1988). The model formulation is described in Laird, N.M. and Ware, J.H. (1982). The variance-covariance parametrizations are described in <Pinheiro, J.C. and Bates., D.M. (1996). The different correlation structures available for the correlation argument are described in Box, G.E.P., Jenkins, G.M., and Reinsel G.C. (1994), Littel, R.C., Milliken, G.A., Stroup, W.W., and Wolfinger, R.D. (1996), and Venables, W.N. and Ripley, B.D. (2002). The use of variance functions for linear and nonlinear mixed effects models is presented in detail in Davidian, M. and Giltinan, D.M. (1995).

Box, G.E.P., Jenkins, G.M., and Reinsel G.C. (1994) "Time Series Analysis: Forecasting and Control", 3rd Edition, Holden-Day.

Davidian, M. and Giltinan, D.M. (1995) "Nonlinear Mixed Effects Models for Repeated Measurement Data", Chapman and Hall.

Laird, N.M. and Ware, J.H. (1982) "Random-Effects Models for Longitudinal Data", Biometrics, 38, 963-974.

Lindstrom, M.J. and Bates, D.M. (1988) "Newton-Raphson and EM Algorithms for Linear Mixed-Effects Models for Repeated-Measures Data", Journal of the American Statistical Association, 83, 1014-1022.

Littel, R.C., Milliken, G.A., Stroup, W.W., and Wolfinger, R.D. (1996) "SAS Systems for Mixed Models", SAS Institute.

Pinheiro, J.C. and Bates., D.M. (1996) "Unconstrained Parametrizations for Variance-Covariance Matrices", Statistics and Computing, 6, 289-296.

Venables, W.N. and Ripley, B.D. (2002) "Modern Applied Statistics with S", 4th Edition, Springer-Verlag.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

fm1 <- nlsList(SSasymp, data = Loblolly)
fm2 <- nlme(fm1, random = Asym ~ 1)
summary(fm1)
#> Call:
#>   Model: height ~ SSasymp(age, Asym, R0, lrc) | Seed 
#>    Data: Loblolly 
#> 
#> Coefficients:
#>    Asym 
#>      Estimate Std. Error   t value     Pr(>|t|)
#> 329  94.12820   7.854473 11.984025 0.0015250903
#> 327  94.94058   8.034073 11.817241 0.0012377238
#> 325  89.88487   6.107469 14.717206 0.0006432969
#> 307 110.69919  10.969922 10.091156 0.0025753993
#> 331 111.00287  11.125081  9.977713 0.0032192120
#> 311 109.98575  10.393098 10.582575 0.0032302622
#> 315 101.05622   8.088408 12.493957 0.0006267587
#> 321 127.13400  15.734953  8.079719 0.0041760502
#> 319 101.08748   7.800912 12.958419 0.0009328591
#> 301  95.66688   6.469242 14.787959 0.0004839491
#> 323  95.55627   6.178278 15.466489 0.0004783920
#> 309 113.51390  10.199357 11.129515 0.0006540778
#> 303 105.71792   7.936191 13.320991 0.0006901480
#> 305  99.17191   6.091461 16.280482 0.0006761877
#>    R0 
#>      Estimate Std. Error   t value    Pr(>|t|)
#> 329 -8.250753   1.146068 -7.199181 0.006700134
#> 327 -7.757495   1.143211 -6.785709 0.006226357
#> 325 -8.759017   1.181264 -7.414952 0.004805581
#> 307 -8.169431   1.108136 -7.372226 0.006382602
#> 331 -8.462608   1.106725 -7.646532 0.006927323
#> 311 -8.558543   1.113691 -7.684848 0.008098734
#> 315 -8.443628   1.142368 -7.391340 0.002941994
#> 321 -7.679356   1.075754 -7.138578 0.005963178
#> 319 -8.502343   1.146942 -7.413052 0.004783021
#> 301 -9.078241   1.172624 -7.741816 0.003259525
#> 323 -9.665035   1.179503 -8.194157 0.003113070
#> 309 -7.595622   1.115674 -6.808100 0.002782876
#> 303 -8.906081   1.144754 -7.779906 0.003356654
#> 305 -9.916654   1.181667 -8.392087 0.004722701
#>    lrc 
#>      Estimate Std. Error   t value     Pr(>|t|)
#> 329 -3.217578  0.1295370 -24.83907 1.750358e-04
#> 327 -3.229325  0.1313231 -24.59068 1.399840e-04
#> 325 -3.086219  0.1120207 -27.55043 9.918120e-05
#> 307 -3.390345  0.1439118 -23.55849 2.091688e-04
#> 331 -3.397574  0.1447737 -23.46817 2.570655e-04
#> 311 -3.362518  0.1383662 -24.30158 2.770780e-04
#> 315 -3.232825  0.1237859 -26.11625 6.945353e-05
#> 321 -3.575331  0.1692979 -21.11858 2.452645e-04
#> 319 -3.214021  0.1204012 -26.69426 1.083870e-04
#> 301 -3.116381  0.1099075 -28.35459 6.931135e-05
#> 323 -3.092266  0.1058796 -29.20549 7.171926e-05
#> 309 -3.352815  0.1334770 -25.11905 5.763078e-05
#> 303 -3.222957  0.1165915 -27.64315 7.822383e-05
#> 305 -3.084841  0.1011074 -30.51055 1.039483e-04
#> 
#> Residual standard error: 0.7003965 on 42 degrees of freedom
#> 
summary(fm2)
#> Nonlinear mixed-effects model fit by maximum likelihood
#>   Model: height ~ SSasymp(age, Asym, R0, lrc) 
#>   Data: Loblolly 
#>       AIC      BIC   logLik
#>   239.486 251.6401 -114.743
#> 
#> Random effects:
#>  Formula: Asym ~ 1 | Seed
#>             Asym  Residual
#> StdDev: 3.650645 0.7188624
#> 
#> Fixed effects:  list(Asym ~ 1, R0 ~ 1, lrc ~ 1) 
#>          Value Std.Error DF   t-value p-value
#> Asym 101.44830 2.4616151 68  41.21209       0
#> R0    -8.62749 0.3179519 68 -27.13459       0
#> lrc   -3.23373 0.0342695 68 -94.36168       0
#>  Correlation: 
#>     Asym   R0    
#> R0   0.704       
#> lrc -0.908 -0.827
#> 
#> Standardized Within-Group Residuals:
#>         Min          Q1         Med          Q3         Max 
#> -2.23604174 -0.62389999  0.05912777  0.65724316  1.95790785 
#> 
#> Number of Observations: 84
#> Number of Groups: 14