nlsList Fit from a selfStart Function

The response variable and primary covariate in formula(data) are used together with model to construct the nonlinear model formula. This is used in the nls calls and, because a selfStarting model function can calculate initial estimates for its parameters from the data, no starting estimates need to be provided.

Usage

# S3 method for selfStart
nlsList(model, data, start, control, level, subset,
        na.action = na.fail, pool = TRUE, warn.nls = NA)

Arguments

model

a "selfStart" model function, which calculates initial estimates for the model parameters from data.

data

a data frame in which to interpret the variables in model. Because no grouping factor can be specified in model, data must inherit from class "groupedData".

start

an optional named list with initial values for the parameters to be estimated in model. It is passed as the start argument to each nls call and is required when the nonlinear function in model does not inherit from class selfStart.

control

a list of control values passed as the control argument to nls. Defaults to an empty list.

level

an optional integer specifying the level of grouping to be used when multiple nested levels of grouping are present.

subset

an optional expression indicating the subset of the rows of data that should be used in the fit. This can be a logical vector, or a numeric vector indicating which observation numbers are to be included, or a character vector of the row names to be included. All observations are included by default.

na.action

a function that indicates what should happen when the data contain NAs. The default action (na.fail) causes nlsList to print an error message and terminate if there are any incomplete observations.

pool, warn.nls

optional logicals, see nlsList.

Value

a list of nls objects with as many components as the number of groups defined by the grouping factor. A NULL value is assigned to the components corresponding to clusters for which the nls

algorithm failed to converge. Generic functions such as coef,

fixed.effects, lme, pairs, plot,

predict, random.effects, summary, and

update have methods that can be applied to an nlsList

object.

See also

Examples

fm1 <- nlsList(SSasympOff, CO2)
summary(fm1)
#> Call:
#>   Model: uptake ~ SSasympOff(conc, Asym, lrc, c0) | Plant 
#>    Data: CO2 
#> 
#> Coefficients:
#>    Asym 
#>     Estimate Std. Error  t value     Pr(>|t|)
#> Qn1 38.13978  0.9911157 38.48167 1.991995e-06
#> Qn2 42.87169  1.0932085 39.21638 2.583951e-06
#> Qn3 44.22800  1.0241029 43.18706 1.809264e-07
#> Qc1 36.42873  1.1941583 30.50578 1.140081e-05
#> Qc3 40.68370  1.2480885 32.59681 1.424622e-04
#> Qc2 39.81950  1.0167249 39.16447 2.692304e-06
#> Mn3 28.48285  1.0624237 26.80932 1.066432e-06
#> Mn2 32.12827  1.0174819 31.57625 3.488780e-06
#> Mn1 34.08481  1.3400580 25.43532 4.199318e-06
#> Mc2 13.55520  1.0506456 12.90178 4.385958e-06
#> Mc3 18.53506  0.8363374 22.16218 1.461565e-06
#> Mc1 21.78723  1.4113332 15.43734 5.756890e-06
#>    lrc 
#>      Estimate Std. Error    t value     Pr(>|t|)
#> Qn1 -4.380647  0.2208027 -19.839647 2.789814e-05
#> Qn2 -4.665728  0.1581312 -29.505427 8.036968e-06
#> Qn3 -4.486118  0.1698748 -26.408376 1.291542e-06
#> Qc1 -4.861741  0.2124535 -22.883785 3.574868e-05
#> Qc3 -4.945218  0.1794898 -27.551534 2.755569e-04
#> Qc2 -4.463838  0.1564976 -28.523374 9.531808e-06
#> Mn3 -4.591566  0.2690656 -17.064855 6.469578e-06
#> Mn2 -4.466157  0.2330265 -19.165873 2.548164e-05
#> Mn1 -5.064579  0.2069978 -24.466818 4.902553e-06
#> Mc2 -4.560851  0.8205907  -5.558010 1.242239e-04
#> Mc3 -3.465158  1.2239475  -2.831133 4.548516e-03
#> Mc1 -5.142256  0.4475243 -11.490450 1.865741e-05
#>    c0 
#>      Estimate Std. Error    t value     Pr(>|t|)
#> Qn1  51.22324  12.621453  4.0584263 0.0117884093
#> Qn2  55.85816   9.998214  5.5868139 0.0051398604
#> Qn3  54.64958   9.814188  5.5684259 0.0006120970
#> Qc1  31.07538  18.988521  1.6365349 0.2228630363
#> Qc3  35.08889  15.994755  2.1937748 0.3930903187
#> Qc2  72.09422   6.924937 10.4108114 0.0005089519
#> Mn3  46.97188  18.133572  2.5903269 0.0092919673
#> Mn2  56.03863  13.055031  4.2924933 0.0079486674
#> Mn1  36.40805  19.079675  1.9082112 0.0606738848
#> Mc2  13.05675  78.006927  0.1673794 0.6806349044
#> Mc3  67.84877  33.737187  2.0110975 0.0150825747
#> Mc1 -20.39998  62.466523 -0.3265746 0.5367826131
#> 
#> Residual standard error: 1.79822 on 48 degrees of freedom
#>