pairs.compareFits {nlme}	R Documentation

Pairs Plot of compareFits Object

Description

Scatter plots of the values being compared are generated for each pair of coefficients in x. Different symbols (colors) are used for each object being compared and values corresponding to the same group are joined by a line, to facilitate comparison of fits. If only two coefficients are present, the trellis function xyplot is used; otherwise the trellis function splom is used.

Usage


## S3 method for class 'compareFits'
pairs(x, subset, key, ...)

Arguments

x
an object of class compareFits.

subset
an optional logical or integer vector specifying which rows of x should be used in the plots. If missing, all rows are used.

key
an optional logical value, or list. If TRUE, a legend is included at the top of the plot indicating which symbols (colors) correspond to which objects being compared. If FALSE, no legend is included. If given as a list, key is passed down as an argument to the trellis function generating the plots (splom or xyplot). Defaults to TRUE.

...
optional arguments passed down to the trellis function generating the plots.

Value

Pairwise scatter plots of the values being compared, with different symbols (colors) used for each object under comparison.

Author(s)

José Pinheiro and Douglas Bates

See Also

compareFits, plot.compareFits, pairs.lme, pairs.lmList, xyplot, splom

Examples


fm1 <- lmList(Orthodont)
fm2 <- lme(Orthodont)
pairs(compareFits(coef(fm1), coef(fm2)))

[Package nlme version 3.1-137 Index]