pdClasses {nlme}	R Documentation

Positive-Definite Matrix Classes

Description

Standard classes of positive-definite matrices (pdMat) structures available in the nlme package.

Value

Available standard classes:

pdSymm
general positive-definite matrix, with no additional structure

pdLogChol
general positive-definite matrix, with no additional structure, using a log-Cholesky parameterization

pdDiag
diagonal

pdIdent
multiple of an identity

pdCompSymm
compound symmetry structure (constant diagonal and constant off-diagonal elements)

pdBlocked
block-diagonal matrix, with diagonal blocks of any "atomic" pdMat class

pdNatural
general positive-definite matrix in natural parametrization (i.e. parametrized in terms of standard deviations and correlations). The underlying coefficients are not unrestricted, so this class should NOT be used for optimization.

Note

Users may define their own pdMat classes by specifying a constructor function and, at a minimum, methods for the functions pdConstruct, pdMatrix and coef. For examples of these functions, see the methods for classes pdSymm and pdDiag.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

pdBlocked, pdCompSymm, pdDiag, pdFactor, pdIdent, pdMat, pdMatrix, pdNatural, pdSymm, pdLogChol

[Package nlme version 3.1-127 Index]