Square-Root Factor of a Positive-Definite Matrix

A square-root factor of the positive-definite matrix represented by object is obtained. Letting \(\Sigma\) denote a positive-definite matrix, a square-root factor of \(\Sigma\) is any square matrix \(L\) such that \(\Sigma = L'L\). This function extracts \(L\).

Usage

pdFactor(object)

Arguments

object

an object inheriting from class pdMat, representing a positive definite matrix, which must have been initialized (i.e. length(coef(object)) > 0).

Value

a vector with a square-root factor of the positive-definite matrix associated with object stacked column-wise.

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

Note

This function is used intensively in optimization algorithms and its value is returned as a vector for efficiency reasons. The pdMatrix function can be used to obtain square-root factors in matrix form.

See also

Examples

pd1 <- pdCompSymm(4 * diag(3) + 1)
pdFactor(pd1)
#> [1]  1.5275252 -1.4142136 -0.8164966  1.5275252  1.4142136 -0.8164966  1.5275252
#> [8]  0.0000000  1.6329932