phenoModel {nlme}	R Documentation

Model function for the Phenobarb data

Description

A model function for a model used with the Phenobarb data. This function uses compiled C code to improve execution speed.

Usage


phenoModel(Subject, time, dose, lCl, lV)

Arguments

Subject
an integer vector of subject identifiers. These should be sorted in increasing order.

time
numeric. A vector of the times at which the sample was drawn or the drug administered (hr).

dose
numeric. A vector of doses of drug administered (μg/kg).

lCl
numeric. A vector of values of the natural log of the clearance parameter according to Subject and time.

lV
numeric. A vector of values of the natural log of the effective volume of distribution according to Subject and time.

Details

See the details section of Phenobarb for a description of the model function that phenoModel evaluates.

Value

a numeric vector of predicted phenobarbital concentrations.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J. C. and Bates, D. M. (2000) Mixed-effects Models in S and S-PLUS, Springer. (section 6.4)

[Package nlme version 3.1-137 Index]