plot.augPred {nlme}	R Documentation

Plot an augPred Object

Description

A Trellis xyplot of predictions versus the primary covariate is generated, with a different panel for each value of the grouping factor. Predicted values are joined by lines, with different line types (colors) being used for each level of grouping. Original observations are represented by circles.

Usage

## S3 method for class 'augPred'
plot(x, key, grid, ...)

Arguments

x
an object of class "augPred".

key
an optional logical value, or list. If TRUE, a legend is included at the top of the plot indicating which symbols (colors) correspond to which prediction levels. If FALSE, no legend is included. If given as a list, key is passed down as an argument to the trellis function generating the plots (xyplot). Defaults to TRUE.

grid
an optional logical value indicating whether a grid should be added to plot. Default is FALSE.

...
optional arguments passed down to the trellis function generating the plots.

Value

A Trellis plot of predictions versus the primary covariate, with panels determined by the grouping factor.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

augPred, xyplot

Examples

fm1 <- lme(Orthodont)
plot(augPred(fm1, level = 0:1, length.out = 2))

[Package nlme version 3.1-121 Index]