predict.gnls {nlme}	R Documentation

Predictions from a gnls Object

Description

The predictions for the nonlinear model represented by object are obtained at the covariate values defined in newdata.

Usage


## S3 method for class 'gnls'
predict(object, newdata, na.action, naPattern, ...)

Arguments

object
an object inheriting from class "gnls", representing a generalized nonlinear least squares fitted model.

newdata
an optional data frame to be used for obtaining the predictions. All variables used in the nonlinear model must be present in the data frame. If missing, the fitted values are returned.

na.action
a function that indicates what should happen when newdata contains NAs. The default action (na.fail) causes the function to print an error message and terminate if there are any incomplete observations.

naPattern
an expression or formula object, specifying which returned values are to be regarded as missing.

...
some methods for this generic require additional arguments. None are used in this method.

Value

a vector with the predicted values.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

gnls

Examples


fm1 <- gnls(weight ~ SSlogis(Time, Asym, xmid, scal), Soybean,
            weights = varPower())
newSoybean <- data.frame(Time = c(10,30,50,80,100))
predict(fm1, newSoybean)

[Package nlme version 3.1-137 Index]