recalc.corStruct {nlme}	R Documentation

Recalculate for corStruct Object

Description

This method function pre-multiples the "Xy" component of conLin by the transpose square-root factor(s) of the correlation matrix (matrices) associated with object and adds the log-likelihood contribution of object, given by logLik(object), to the "logLik" component of conLin.

Usage

## S3 method for class 'corStruct'
recalc(object, conLin, ...)

Arguments

object
an object inheriting from class "corStruct", representing a correlation structure.

conLin
a condensed linear model object, consisting of a list with components "Xy", corresponding to a regression matrix (X) combined with a response vector (y), and "logLik", corresponding to the log-likelihood of the underlying model.

...
some methods for this generic require additional arguments. None are used in this method.

Value

the recalculated condensed linear model object.

Note

This method function is only used inside model fitting functions, such as lme and gls, that allow correlated error terms.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

corFactor, logLik.corStruct

[Package nlme version 3.1-121 Index]