residuals.gnlsStruct {nlme}	R Documentation

Calculate gnlsStruct Residuals

Description

The residuals for the nonlinear model represented by object are extracted.

Usage

## S3 method for class 'gnlsStruct'
residuals(object, ...)

Arguments

object
an object inheriting from class "gnlsStruct", representing a list of model components, such as corStruct and varFunc objects, and attributes specifying the underlying nonlinear model and the response variable.

...
some methods for this generic require additional arguments. None are used in this method.

Value

a vector with the residuals for the nonlinear model represented by object.

Note

This method function is primarily used inside gnls and residuals.gnls.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

gnls, residuals.gnls, fitted.gnlsStruct

[Package nlme version 3.1-121 Index]