Extract lmList Residuals

The residuals are extracted from each lm component of object and arranged into a list with as many components as object, or combined into a single vector.

Usage

# S3 method for lmList
residuals(object, type, subset, asList, ...)

Arguments

object

an object inheriting from class "lmList", representing a list of lm objects with a common model.

subset

an optional character or integer vector naming the lm components of object from which the residuals are to be extracted. Default is NULL, in which case all components are used.

type

an optional character string specifying the type of residuals to be extracted. Options include "response" for the "raw" residuals (observed - fitted), "pearson" for the standardized residuals (raw residuals divided by the estimated residual standard error) using different standard errors for each lm fit, and "pooled.pearson" for the standardized residuals using a pooled estimate of the residual standard error. Partial matching of arguments is used, so only the first character needs to be provided. Defaults to "response".

asList

an optional logical value. If TRUE, the returned object is a list with the residuals split by groups; else the returned value is a vector. Defaults to FALSE.

...

some methods for this generic require additional arguments. None are used in this method.

Value

a list with components given by the residuals of each lm

component of object, or a vector with the residuals for all

lm components of object.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

fm1 <- lmList(distance ~ age | Subject, Orthodont)
residuals(fm1)
#>   M01   M01   M01   M01   M02   M02   M02   M02   M03   M03   M03   M03   M04 
#>  1.10 -1.80  0.30  0.40  0.45 -0.10 -1.15  0.80  1.00 -1.00 -1.00  1.00 -0.60 
#>   M04   M04   M04   M05   M05   M05   M05   M06   M06   M06   M06   M07   M07 
#>  1.05 -0.30 -0.15 -0.45  1.35 -1.35  0.45  0.15 -0.20 -0.05  0.10  0.65 -0.95 
#>   M07   M07   M08   M08   M08   M08   M09   M09   M09   M09   M10   M10   M10 
#> -0.05  0.35  1.25 -2.00  0.25  0.50  0.80 -3.65  4.90 -2.05  0.25 -0.75  0.75 
#>   M10   M11   M11   M11   M11   M12   M12   M12   M12   M13   M13   M13   M13 
#> -0.25  0.35 -0.30 -0.45  0.40  0.25  0.25 -1.25  0.75 -1.40  2.20 -0.20 -0.60 
#>   M14   M14   M14   M14   M15   M15   M15   M15   M16   M16   M16   M16   F01 
#> -0.80  1.15  0.10 -0.45  0.50 -0.25 -1.00  0.75  0.65 -0.95 -0.05  0.35  0.75 
#>   F01   F01   F01   F02   F02   F02   F02   F03   F03   F03   F03   F04   F04 
#> -1.00 -0.25  0.50  0.40 -0.70  0.20  0.10 -0.70  1.10 -0.10 -0.30  0.05  0.10 
#>   F04   F04   F05   F05   F05   F05   F06   F06   F06   F06   F07   F07   F07 
#> -0.35  0.20 -0.30  0.65 -0.40  0.05  0.00  0.25 -0.50  0.25  0.15  0.05 -0.55 
#>   F07   F08   F08   F08   F08   F09   F09   F09   F09   F10   F10   F10   F10 
#>  0.35  0.15 -0.20 -0.05  0.10 -0.30  0.15  0.60 -0.45 -0.65  0.95  0.05 -0.35 
#>   F11   F11   F11   F11 
#>  0.15 -0.70  0.95 -0.40 
#> attr(,"label")
#> [1] "Residuals (mm)"