splitFormula {nlme}	R Documentation

Split a Formula

Description

Splits the right hand side of form into a list of subformulas according to the presence of sep. The left hand side of form, if present, will be ignored. The length of the returned list will be equal to the number of occurrences of sep in form plus one.

Usage

splitFormula(form, sep)

Arguments

form
a formula object.

sep
an optional character string specifying the separator to be used for splitting the formula. Defaults to "/".

Value

a list of formulas, corresponding to the split of form according to sep.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

formula

Examples

splitFormula(~ g1/g2/g3)

[Package nlme version 3.1-127 Index]