Summarize a Generalized Least Squares gls Object

Additional information about the linear model fit represented by object is extracted and included as components of object.

Usage

# S3 method for gls
summary(object, verbose, ...)

Arguments

object

an object inheriting from class "gls", representing a generalized least squares fitted linear model.

verbose

an optional logical value used to control the amount of output when the object is printed. Defaults to FALSE.

...

some methods for this generic require additional arguments. None are used in this method.

Value

an object inheriting from class summary.gls with all components included in object (see glsObject for a full description of the components) plus the following components:

corBeta

approximate correlation matrix for the coefficients estimates

tTable

a matrix with columns Value, Std. Error, t-value, and p-value representing respectively the coefficients estimates, their approximate standard errors, the ratios between the estimates and their standard errors, and the associated p-value under a \(t\) approximation. Rows correspond to the different coefficients.

residuals

if more than five observations are used in the gls fit, a vector with the minimum, first quartile, median, third quartile, and maximum of the residuals distribution; else the residuals.

AIC

the Akaike Information Criterion corresponding to object.

BIC

the Bayesian Information Criterion corresponding to object.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

fm1 <- gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), Ovary,
           correlation = corAR1(form = ~ 1 | Mare))
summary(fm1)
#> Generalized least squares fit by REML
#>   Model: follicles ~ sin(2 * pi * Time) + cos(2 * pi * Time) 
#>   Data: Ovary 
#>        AIC      BIC    logLik
#>   1571.455 1590.056 -780.7273
#> 
#> Correlation Structure: AR(1)
#>  Formula: ~1 | Mare 
#>  Parameter estimate(s):
#>       Phi 
#> 0.7532079 
#> 
#> Coefficients:
#>                        Value Std.Error   t-value p-value
#> (Intercept)        12.216398 0.6646437 18.380373  0.0000
#> sin(2 * pi * Time) -2.774712 0.6450478 -4.301561  0.0000
#> cos(2 * pi * Time) -0.899605 0.6975383 -1.289685  0.1981
#> 
#>  Correlation: 
#>                    (Intr) s(*p*T
#> sin(2 * pi * Time)  0.000       
#> cos(2 * pi * Time) -0.294  0.000
#> 
#> Standardized residuals:
#>         Min          Q1         Med          Q3         Max 
#> -2.41180365 -0.75405234 -0.02923628  0.63156880  3.16247697 
#> 
#> Residual standard error: 4.616172 
#> Degrees of freedom: 308 total; 305 residual
coef(summary(fm1)) # "the matrix"
#>                         Value Std.Error   t-value      p-value
#> (Intercept)        12.2163982 0.6646437 18.380373 2.618737e-51
#> sin(2 * pi * Time) -2.7747122 0.6450478 -4.301561 2.286284e-05
#> cos(2 * pi * Time) -0.8996047 0.6975383 -1.289685 1.981371e-01