Summarize an lme Object

Additional information about the linear mixed-effects fit represented by object is extracted and included as components of object. The returned object has a print and a coef method, the latter returning the coefficient's tTtable.

Usage

# S3 method for lme
summary(object, adjustSigma, verbose, ...)
# S3 method for summary.lme
print(x, verbose = FALSE, ...)

Arguments

object

an object inheriting from class "lme", representing a fitted linear mixed-effects model.

adjustSigma

an optional logical value. If TRUE and the estimation method used to obtain object was maximum likelihood, the residual standard error is multiplied by \(\sqrt{n_{obs}/(n_{obs} - n_{par})}\), converting it to a REML-like estimate. This argument is only used when a single fitted object is passed to the function. Default is TRUE.

verbose

an optional logical value used to control the amount of output in the print.summary.lme method. Defaults to FALSE.

...

additional optional arguments passed to methods, mainly for the print method.

x

a "summary.lme" object.

Value

an object inheriting from class summary.lme with all components included in object (see lmeObject for a full description of the components) plus the following components:

corFixed

approximate correlation matrix for the fixed effects estimates.

tTable

a matrix with columns named Value, Std. Error, DF, t-value, and p-value representing respectively the fixed effects estimates, their approximate standard errors, the denominator degrees of freedom, the ratios between the estimates and their standard errors, and the associated p-value from a t distribution. Rows correspond to the different fixed effects.

residuals

if more than five observations are used in the lme fit, a vector with the minimum, first quartile, median, third quartile, and maximum of the innermost grouping level residuals distribution; else the innermost grouping level residuals.

AIC

the Akaike Information Criterion corresponding to object.

BIC

the Bayesian Information Criterion corresponding to object.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

fm1 <- lme(distance ~ age, Orthodont, random = ~ age | Subject)
(s1 <- summary(fm1))
#> Linear mixed-effects model fit by REML
#>   Data: Orthodont 
#>        AIC      BIC    logLik
#>   454.6367 470.6173 -221.3183
#> 
#> Random effects:
#>  Formula: ~age | Subject
#>  Structure: General positive-definite, Log-Cholesky parametrization
#>             StdDev    Corr  
#> (Intercept) 2.3270341 (Intr)
#> age         0.2264278 -0.609
#> Residual    1.3100397       
#> 
#> Fixed effects:  distance ~ age 
#>                 Value Std.Error DF   t-value p-value
#> (Intercept) 16.761111 0.7752460 80 21.620377       0
#> age          0.660185 0.0712533 80  9.265333       0
#>  Correlation: 
#>     (Intr)
#> age -0.848
#> 
#> Standardized Within-Group Residuals:
#>          Min           Q1          Med           Q3          Max 
#> -3.223106060 -0.493761143  0.007316631  0.472151118  3.916033216 
#> 
#> Number of Observations: 108
#> Number of Groups: 27 
coef(s1) # the (coef | Std.E | t | P-v ) matrix
#>                  Value  Std.Error DF   t-value      p-value
#> (Intercept) 16.7611111 0.77524603 80 21.620377 3.737299e-35
#> age          0.6601852 0.07125326 80  9.265333 2.645956e-14
stopifnot(is.matrix(coef(s1)))