summary.modelStruct {nlme}	R Documentation

Summarize a modelStruct Object

Description

This method function applies summary to each element of object.

Usage

## S3 method for class 'modelStruct'
summary(object, ...)

Arguments

object
an object inheriting from class "modelStruct", representing a list of model components, such as reStruct, corStruct and varFunc objects.

...
some methods for this generic require additional arguments. None are used in this method.

Value

a list with elements given by the summarized components of object. The returned value is of class summary.modelStruct, also inheriting from the same classes as object.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

reStruct, summary

Examples

lms1 <- lmeStruct(reStruct = reStruct(pdDiag(diag(2), ~age)),
   corStruct = corAR1(0.3))
summary(lms1)

[Package nlme version 3.1-121 Index]