Summarize an nlsList Object

The summary function is applied to each nls component of object to produce summary information on the individual fits, which is organized into a list of summary statistics. The returned object is suitable for printing with the print.summary.nlsList method.

Usage

# S3 method for nlsList
summary(object, ...)

Arguments

object

an object inheriting from class "nlsList", representing a list of nls fitted objects.

...

optional arguments to the summary.lmList method. One such optional argument is pool, a logical value indicating whether a pooled estimate of the residual standard error should be used. Default is attr(object, "pool").

Value

a list with summary statistics obtained by applying summary

to the elements of object, inheriting from class

summary.nlsList. The components of value are:

call

a list containing an image of the nlsList call that produced object.

parameters

a three dimensional array with summary information on the nls coefficients. The first dimension corresponds to the names of the object components, the second dimension is given by "Value", "Std. Error", "t value", and "Pr(>|t|)", corresponding, respectively, to the coefficient estimates and their associated standard errors, t-values, and p-values. The third dimension is given by the coefficients names.

correlation

a three dimensional array with the correlations between the individual nls coefficient estimates. The first dimension corresponds to the names of the object components. The third dimension is given by the coefficients names. For each coefficient, the rows of the associated array give the correlations between that coefficient and the remaining coefficients, by nls component.

cov.unscaled

a three dimensional array with the unscaled variances/covariances for the individual lm coefficient estimates (giving the estimated variance/covariance for the coefficients, when multiplied by the estimated residual errors). The first dimension corresponds to the names of the object components. The third dimension is given by the coefficients names. For each coefficient, the rows of the associated array give the unscaled covariances between that coefficient and the remaining coefficients, by nls component.

df

an array with the number of degrees of freedom for the model and for residuals, for each nls component.

df.residual

the total number of degrees of freedom for residuals, corresponding to the sum of residuals df of all nls components.

pool

the value of the pool argument to the function.

RSE

the pooled estimate of the residual standard error.

sigma

a vector with the residual standard error estimates for the individual lm fits.

Author

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See also

Examples

fm1 <- nlsList(SSasymp, Loblolly)
summary(fm1)
#> Call:
#>   Model: height ~ SSasymp(age, Asym, R0, lrc) | Seed 
#>    Data: Loblolly 
#> 
#> Coefficients:
#>    Asym 
#>      Estimate Std. Error   t value     Pr(>|t|)
#> 329  94.12820   7.854473 11.984025 0.0015250903
#> 327  94.94058   8.034073 11.817241 0.0012377238
#> 325  89.88487   6.107469 14.717206 0.0006432969
#> 307 110.69919  10.969922 10.091156 0.0025753993
#> 331 111.00287  11.125081  9.977713 0.0032192120
#> 311 109.98575  10.393098 10.582575 0.0032302622
#> 315 101.05622   8.088408 12.493957 0.0006267587
#> 321 127.13400  15.734953  8.079719 0.0041760502
#> 319 101.08748   7.800912 12.958419 0.0009328591
#> 301  95.66688   6.469242 14.787959 0.0004839491
#> 323  95.55627   6.178278 15.466489 0.0004783920
#> 309 113.51390  10.199357 11.129515 0.0006540778
#> 303 105.71792   7.936191 13.320991 0.0006901480
#> 305  99.17191   6.091461 16.280482 0.0006761877
#>    R0 
#>      Estimate Std. Error   t value    Pr(>|t|)
#> 329 -8.250753   1.146068 -7.199181 0.006700134
#> 327 -7.757495   1.143211 -6.785709 0.006226357
#> 325 -8.759017   1.181264 -7.414952 0.004805581
#> 307 -8.169431   1.108136 -7.372226 0.006382602
#> 331 -8.462608   1.106725 -7.646532 0.006927323
#> 311 -8.558543   1.113691 -7.684848 0.008098734
#> 315 -8.443628   1.142368 -7.391340 0.002941994
#> 321 -7.679356   1.075754 -7.138578 0.005963178
#> 319 -8.502343   1.146942 -7.413052 0.004783021
#> 301 -9.078241   1.172624 -7.741816 0.003259525
#> 323 -9.665035   1.179503 -8.194157 0.003113070
#> 309 -7.595622   1.115674 -6.808100 0.002782876
#> 303 -8.906081   1.144754 -7.779906 0.003356654
#> 305 -9.916654   1.181667 -8.392087 0.004722701
#>    lrc 
#>      Estimate Std. Error   t value     Pr(>|t|)
#> 329 -3.217578  0.1295370 -24.83907 1.750358e-04
#> 327 -3.229325  0.1313231 -24.59068 1.399840e-04
#> 325 -3.086219  0.1120207 -27.55043 9.918120e-05
#> 307 -3.390345  0.1439118 -23.55849 2.091688e-04
#> 331 -3.397574  0.1447737 -23.46817 2.570655e-04
#> 311 -3.362518  0.1383662 -24.30158 2.770780e-04
#> 315 -3.232825  0.1237859 -26.11625 6.945353e-05
#> 321 -3.575331  0.1692979 -21.11858 2.452645e-04
#> 319 -3.214021  0.1204012 -26.69426 1.083870e-04
#> 301 -3.116381  0.1099075 -28.35459 6.931135e-05
#> 323 -3.092266  0.1058796 -29.20549 7.171926e-05
#> 309 -3.352815  0.1334770 -25.11905 5.763078e-05
#> 303 -3.222957  0.1165915 -27.64315 7.822383e-05
#> 305 -3.084841  0.1011074 -30.51055 1.039483e-04
#> 
#> Residual standard error: 0.7003965 on 42 degrees of freedom
#>