summary.pdMat {nlme}	R Documentation

Summarize a pdMat Object

Description

Attributes structName and noCorrelation, with the values of the corresponding arguments to the method function, are appended to object and its class is changed to summary.pdMat.

Usage


## S3 method for class 'pdMat'
summary(object, structName, noCorrelation, ...)

Arguments

object
an object inheriting from class "pdMat", representing a positive definite matrix.

structName
an optional character string with a description of the pdMat class. Default depends on the method function: "Blocked" for pdBlocked, "Compound Symmetry" for pdCompSymm, "Diagonal" for pdDiag, "Multiple of an Identity" for pdIdent, "General Positive-Definite, Natural Parametrization" for pdNatural, "General Positive-Definite" for pdSymm, and data.class(object) for pdMat.

noCorrelation
an optional logical value indicating whether correlations are to be printed in print.summary.pdMat. Default depends on the method function: FALSE for pdDiag and pdIdent, and TRUE for all other classes.

...
some methods for this generic require additional arguments. None are used in this method.

Value

an object similar to object, with additional attributes structName and noCorrelation, inheriting from class summary.pdMat.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

print.summary.pdMat, pdMat

Examples


summary(pdSymm(diag(4)))

[Package nlme version 3.1-137 Index]