summary.varFunc {nlme}	R Documentation

Summarize varFunc Object

Description

A structName attribute, with the value of corresponding argument, is appended to object and its class is changed to summary.varFunc.

Usage


## S3 method for class 'varFunc'
summary(object, structName, ...)

Arguments

object
an object inheriting from class "varFunc", representing a variance function structure.

structName
an optional character string with a description of the varFunc class. Default depends on the method function: "Combination of variance functions" for varComb, "Constant plus power of covariate" for varConstPower, "Exponential of variance covariate" for varExp, "Different standard deviations per stratum" for varIdent, "Power of variance covariate" for varPower, and data.class(object) for varFunc.

...
some methods for this generic require additional arguments. None are used in this method.

Value

an object similar to object, with an additional attribute structName, inheriting from class summary.varFunc.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

varClasses, varFunc

Examples


vf1 <- varPower(0.3, form = ~age)
vf1 <- Initialize(vf1, Orthodont)
summary(vf1)

[Package nlme version 3.1-137 Index]