Coef {nlme}	R Documentation

Assign Values to Coefficients

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include all "pdMat", "corStruct" and "varFunc" classes, "reStruct", and "modelStruct".

Usage


coef(object, ...) <-  value

Arguments

object
any object representing a fitted model, or, by default, any object with a coef component.

...
some methods for this generic function may require additional arguments.

value
a value to be assigned to the coefficients associated with object.

Value

will depend on the method function; see the appropriate documentation.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

coef

Examples


## see the method function documentation

[Package nlme version 3.1-137 Index]