Covariate.varFunc {nlme}	R Documentation

Assign varFunc Covariate

Description

The covariate(s) used in the calculation of the weights of the variance function represented by object is (are) replaced by value. If object has been initialized, value must have the same dimensions as getCovariate(object).

Usage


## S3 replacement method for class 'varFunc'
covariate(object) <- value

Arguments

object
an object inheriting from class "varFunc", representing a variance function structure.

value
a value to be assigned to the covariate associated with object.

Value

a varFunc object similar to object, but with its covariate attribute replaced by value.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

getCovariate.varFunc

Examples


vf1 <- varPower(1.1, form = ~age)
covariate(vf1) <- Orthodont[["age"]]

[Package nlme version 3.1-137 Index]