Dim {nlme}	R Documentation

Extract Dimensions from an Object

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include: corSpatial, corStruct, pdCompSymm, pdDiag, pdIdent, pdMat, and pdSymm.

Usage

Dim(object, ...)

Arguments

object 	 any object for which dimensions can be extracted.

... 	 some methods for this generic function require additional arguments.

Value

will depend on the method function used; see the appropriate documentation.

Note

If dim allowed more than one argument, there would be no need for this generic function.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

Dim.pdMat, Dim.corStruct

Examples

## see the method function documentation

[Package nlme version 3.1-93 Index]