Dim.corSpatial {nlme}	R Documentation

Dimensions of a corSpatial Object

Description

if groups is missing, it returns the Dim attribute of object; otherwise, calculates the dimensions associated with the grouping factor.

Usage

## S3 method for class 'corSpatial':
Dim(object, groups, ...)

Arguments

object 	 an object inheriting from class corSpatial, representing a spatial correlation structure.

groups 	 an optional factor defining the grouping of the observations; observations within a group are correlated and observations in different groups are uncorrelated.

... 	 further arguments to be passed to or from methods.

Value

a list with components:

N 	 length of groups

M 	 number of groups

spClass 	 an integer representing the spatial correlation class; 0 = user defined class, 1 = corSpher, 2 = corExp, 3 = corGaus, 4 = corLin

sumLenSq 	 sum of the squares of the number of observations per group

len 	 an integer vector with the number of observations per group

start 	 an integer vector with the starting position for the distance vectors in each group, beginning from zero

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

Dim, Dim.corStruct

Examples

Dim(corGaus(), getGroups(Orthodont))

cs1ARMA <- corARMA(0.4, form = ~ 1 | Subject, q = 1)
cs1ARMA <- Initialize(cs1ARMA, data = Orthodont)
Dim(cs1ARMA)

[Package nlme version 3.1-93 Index]