[.pdMat {nlme}	R Documentation

Subscript a pdMat Object

Description

This method function extracts sub-matrices from the positive-definite matrix represented by x.

Usage

## S3 method for class 'pdMat':
x[i, j, drop = TRUE]
## S3 method for class 'pdMat':
x[i, j] <- value

Arguments

x 	 an object inheriting from class pdMat representing a positive-definite matrix.

i, j 	 optional subscripts applying respectively to the rows and columns of the positive-definite matrix represented by object. When i (j) is omitted, all rows (columns) are extracted.

drop 	 a logical value. If TRUE, single rows or columns are converted to vectors. If FALSE the returned value retains its matrix representation.

value 	 a vector, or matrix, with the replacement values for the relevant piece of the matrix represented by x.

Value

if i and j are identical, the returned value will be pdMat object with the same class as x. Otherwise, the returned value will be a matrix. In the case a single row (or column) is selected, the returned value may be converted to a vector, according to the rules above.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

[, pdMat

Examples

pd1 <- pdSymm(diag(3))
pd1[1, , drop = FALSE]
pd1[1:2, 1:2] <- 3 * diag(2)

[Package nlme version 3.1-93 Index]