Initialize {nlme}	R Documentation

Initialize Object

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include: corStruct, lmeStruct, reStruct, and varFunc.

Usage

Initialize(object, data, ...)

Arguments

object 	 any object requiring initialization, e.g. "plug-in" structures such as corStruct and varFunc objects.

data 	 a data frame to be used in the initialization procedure.

... 	 some methods for this generic function require additional arguments.

Value

an initialized object with the same class as object. Changes introduced by the initialization procedure will depend on the method function used; see the appropriate documentation.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

Initialize.corStruct, Initialize.lmeStruct, Initialize.glsStruct, Initialize.varFunc, isInitialized

Examples

## see the method function documentation

[Package nlme version 3.1-93 Index]