Initialize.varFunc {nlme}	R Documentation

Initialize varFunc Object

Description

This method initializes object by evaluating its associated covariate(s) and grouping factor, if any is present, in data; determining if the covariate(s) need to be updated when the values of the coefficients associated with object change; initializing the log-likelihood and the weights associated with object; and assigning initial values for the coefficients in object, if none were present. The covariate(s) will only be initialized if no update is needed when coef(object) changes.

Usage


## S3 method for class 'varFunc'
Initialize(object, data, ...)

Arguments

object
an object inheriting from class "varFunc", representing a variance function structure.

data
a data frame in which to evaluate the variables named in formula(object).

...
this argument is included to make this method compatible with the generic.

Value

an initialized object with the same class as object representing a variance function structure.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

Initialize

Examples


vf1 <- varPower( form = ~ age | Sex )
vf1 <- Initialize( vf1, Orthodont )

[Package nlme version 3.1-137 Index]