MathAchSchool {nlme}	R Documentation

School demographic data for MathAchieve

Description

The MathAchSchool data frame has 160 rows and 7 columns.

Format

This data frame contains the following columns:

School

a factor giving the school on which the measurement is made.

Size

a numeric vector giving the number of students in the school

Sector

a factor with levels Public Catholic

PRACAD

a numeric vector giving the percentage of students on the academic track

DISCLIM

a numeric vector measuring the discrimination climate

HIMINTY

a factor with levels 0 1

MEANSES

a numeric vector giving the mean SES score.

Details

These variables give the school-level demographic data to accompany the MathAchieve data.

[Package nlme version 3.1-93 Index]