Matrix.pdMat {nlme}	R Documentation

Assign Matrix to a pdMat Object

Description

The positive-definite matrix represented by object is replaced by value. If the original matrix had row and/or column names, the corresponding names for value can either be NULL, or a permutation of the original names.

Usage

## S3 method for class 'pdMat':
matrix(object) <- value

Arguments

object 	 an object inheriting from class pdMat, representing a positive definite matrix.

value 	 a matrix with the new values to be assigned to the positive-definite matrix represented by object. Must have the same dimensions as as.matrix(object).

Value

a pdMat object similar to object, but with its coefficients modified to produce the matrix in value.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

pdMat,"matrix<-"

Examples

pd1 <- pdSymm(diag(3))
matrix(pd1) <- diag(1:3)
pd1

[Package nlme version 3.1-93 Index]