Milk {nlme}	R Documentation

Protein content of cows' milk

Description

The Milk data frame has 1337 rows and 4 columns.

Format

This data frame contains the following columns:

protein

a numeric vector giving the protein content of the milk.

Time

a numeric vector giving the time since calving (weeks).

Cow

an ordered factor giving a unique identifier for each cow.

Diet

a factor with levels barley, barley+lupins, and lupins identifying the diet for each cow.

Details

Diggle, Liang, and Zeger (1994) describe data on the protein content of cows' milk in the weeks following calving. The cattle are grouped according to whether they are fed a diet with barley alone, with barley and lupins, or with lupins alone.

Source

Diggle, Peter J., Liang, Kung-Yee and Zeger, Scott L. (1994), Analysis of longitudinal data, Oxford University Press, Oxford.

[Package nlme version 3.1-93 Index]