Names {nlme}	R Documentation

Names Associated with an Object

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include: formula, modelStruct, pdBlocked, pdMat, and reStruct.

Usage

Names(object, ...)
Names(object, ...) <- value

Arguments

object 	 any object for which names can be extracted and/or assigned.

... 	 some methods for this generic function require additional arguments.

value 	 names to be assigned to object.

Value

will depend on the method function used; see the appropriate documentation.

SIDE EFFECTS

On the left side of an assignment, sets the names associated with object to value, which must have an appropriate length.

Note

If names were generic, there would be no need for this generic function.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

Names.formula, Names.pdMat

Examples

## see the method function documentation

[Package nlme version 3.1-93 Index]