Names.formula {nlme}	R Documentation

Extract Names from a formula

Description

This method function returns the names of the terms corresponding to the right hand side of object (treated as a linear formula), obtained as the column names of the corresponding model.matrix.

Usage

## S3 method for class 'formula':
Names(object, data, exclude, ...)

Arguments

object 	 an object inheriting from class formula.

data 	 an optional data frame containing the variables specified in object. By default the variables are taken from the environment from which Names.formula is called.

exclude 	 an optional character vector with names to be excluded from the returned value. Default is c("pi",".").

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

a character vector with the column names of the model.matrix corresponding to the right hand side of object which are not listed in excluded.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

model.matrix, terms, Names

Examples

Names(distance ~ Sex * age, data = Orthodont)

[Package nlme version 3.1-93 Index]