Names.reStruct {nlme}	R Documentation

Names of an reStruct Object

Description

This method function extracts the column names of each of the positive-definite matrices represented the pdMat elements of object.

Usage

## S3 method for class 'reStruct':
Names(object, ...)
## S3 method for class 'reStruct':
Names(object, ...) <- value

Arguments

object 	 an object inheriting from class reStruct, representing a random effects structure and consisting of a list of pdMat objects.

value 	 a list of character vectors with the replacement values for the names of the individual pdMat objects that form object. It must have the same length as object.

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

a list containing the column names of each of the positive-definite matrices represented by the pdMat elements of object.

SIDE EFFECTS

On the left side of an assignment, sets the Names of the pdMat elements of object to the corresponding element of value.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

reStruct, pdMat, Names.pdMat

Examples

rs1 <- reStruct(list(Dog = ~day, Side = ~1), data = Pixel)
Names(rs1)

[Package nlme version 3.1-93 Index]