VarCorr {nlme}	R Documentation

Extract variance and correlation components

Description

This function calculates the estimated variances, standard deviations, and correlations between the random-effects terms in a linear mixed-effects model, of class lme, or a nonlinear mixed-effects model, of class nlme. The within-group error variance and standard deviation are also calculated.

Usage

VarCorr(x, sigma, rdig)

Arguments

x 	 a fitted model object, usually an object inheriting from class lme.

sigma 	 an optional numeric value used as a multiplier for the standard deviations. Default is 1.

rdig 	 an optional integer value specifying the number of digits used to represent correlation estimates. Default is 3.

Value

a matrix with the estimated variances, standard deviations, and correlations for the random effects. The first two columns, named Variance and StdDev, give, respectively, the variance and the standard deviations. If there are correlation components in the random effects model, the third column, named Corr, and the remaining unnamed columns give the estimated correlations among random effects within the same level of grouping. The within-group error variance and standard deviation are included as the last row in the matrix.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer, esp. pp. 100, 461.

See Also

lme, nlme

Examples

fm1 <- lme(distance ~ age, data = Orthodont, random = ~age)
VarCorr(fm1)

[Package nlme version 3.1-93 Index]