asOneFormula {nlme}	R Documentation

Combine Formulas of a Set of Objects

Description

The names of all variables used in the formulas extracted from the objects defined in ... are converted into a single linear formula, with the variables names separated by +.

Usage


asOneFormula(..., omit)

Arguments

...
objects, or lists of objects, from which a formula can be extracted.

omit
an optional character vector with the names of variables to be omitted from the returned formula. Defaults to c(".", "pi").

Value

a one-sided linear formula with all variables named in the formulas extracted from the objects in ..., except the ones listed in omit.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

formula, all.vars

Examples


asOneFormula(y ~ x + z | g, list(~ w, ~ t * sin(2 * pi)))

[Package nlme version 3.1-137 Index]