coef.pdMat {nlme}	R Documentation

pdMat Object Coefficients

Description

This method function extracts the coefficients associated with the positive-definite matrix represented by object.

Usage

## S3 method for class 'pdMat':
coef(object, unconstrained, ...)
## S3 method for class 'pdMat':
coef(object, ...) <- value

Arguments

object 	 an object inheriting from class pdMat, representing a positive-definite matrix.

unconstrained 	 a logical value. If TRUE the coefficients are returned in unconstrained form (the same used in the optimization algorithm). If FALSE the upper triangular elements of the positive-definite matrix represented by object are returned. Defaults to TRUE.

value 	 a vector with the replacement values for the coefficients associated with object. It must be a vector with the same length of coef{object} and must be given in unconstrained form.

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

a vector with the coefficients corresponding to object.

SIDE EFFECTS

On the left side of an assignment, sets the values of the coefficients of object to value.

Author(s)

Jose Pinheiro and Douglas Bates

References

Pinheiro, J.C. and Bates., D.M. (1996) "Unconstrained Parametrizations for Variance-Covariance Matrices", Statistics and Computing, 6, 289-296.

See Also

pdMat

Examples

coef(pdSymm(diag(3)))

[Package nlme version 3.1-93 Index]