corClasses {nlme}	R Documentation

Correlation Structure Classes

Description

Standard classes of correlation structures (corStruct) available in the nlme library.

Value

Available standard classes:

corAR1 	 autoregressive process of order 1.

corARMA 	 autoregressive moving average process, with arbitrary orders for the autoregressive and moving average components.

corCAR1 	 continuous autoregressive process (AR(1) process for a continuous time covariate).

corCompSymm 	 compound symmetry structure corresponding to a constant correlation.

corExp 	 exponential spatial correlation.

corGaus 	 Gaussian spatial correlation.

corLin 	 linear spatial correlation.

corRatio 	 Rational quadratics spatial correlation.

corSpher 	 spherical spatial correlation.

corSymm 	 general correlation matrix, with no additional structure.

Note

Users may define their own corStruct classes by specifying a constructor function and, at a minimum, methods for the functions corMatrix and coef. For examples of these functions, see the methods for classes corSymm and corAR1.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

corAR1, corARMA, corCAR1, corCompSymm, corExp, corGaus, corLin, corRatio, corSpher, corSymm, summary.corStruct

[Package nlme version 3.1-93 Index]