corFactor.corStruct {nlme}	R Documentation

Factor of a corStruct Object Matrix

Description

This method function extracts a transpose inverse square-root factor, or a series of transpose inverse square-root factors, of the correlation matrix, or list of correlation matrices, represented by object. Letting S denote a correlation matrix, a square-root factor of S is any square matrix L such that S = L'L. This method extracts L^(-t).

Usage

## S3 method for class 'corStruct':
corFactor(object, ...)

Arguments

object 	 an object inheriting from class corStruct representing a correlation structure, which must have been initialized (using Initialize).

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

If the correlation structure does not include a grouping factor, the returned value will be a vector with a transpose inverse square-root factor of the correlation matrix associated with object stacked column-wise. If the correlation structure includes a grouping factor, the returned value will be a vector with transpose inverse square-root factors of the correlation matrices for each group, stacked by group and stacked column-wise within each group.

Note

This method function is used intensively in optimization algorithms and its value is returned as a vector for efficiency reasons. The corMatrix method function can be used to obtain transpose inverse square-root factors in matrix form.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

corFactor, corMatrix.corStruct, recalc.corStruct, Initialize.corStruct

Examples

cs1 <- corAR1(form = ~1 | Subject)
cs1 <- Initialize(cs1, data = Orthodont)
corFactor(cs1)

[Package nlme version 3.1-93 Index]