fixef.lmList {nlme}	R Documentation

Extract lmList Fixed Effects

Description

The average of the coefficients corresponding to the lm components of object is calculated.

Usage

## S3 method for class 'lmList'
fixef(object, ...)

Arguments

object
an object inheriting from class "lmList", representing a list of lm objects with a common model.

...
some methods for this generic require additional arguments. None are used in this method.

Value

a vector with the average of the individual lm coefficients in object.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

lmList, random.effects.lmList

Examples

fm1 <- lmList(distance ~ age | Subject, Orthodont)
fixed.effects(fm1)

[Package nlme version 3.1-153 Index]