formula.pdBlocked {nlme}	R Documentation

Extract pdBlocked Formula

Description

The formula attributes of the pdMat elements of x are extracted and returned as a list, in case asList=TRUE, or converted to a single one-sided formula when asList=FALSE. If the pdMat elements do not have a formula attribute, a NULL value is returned.

Usage


## S3 method for class 'pdBlocked'
formula(x, asList, ...)

Arguments

x
an object inheriting from class "pdBlocked", representing a positive definite block diagonal matrix.

asList
an optional logical value. If TRUE, a list with the formulas for the individual block diagonal elements of x is returned; else, if FALSE, a one-sided formula combining all individual formulas is returned. Defaults to FALSE.

...
some methods for this generic require additional arguments. None are used in this method.

Value

a list of one-sided formulas, or a single one-sided formula, or NULL.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

pdBlocked, pdMat

Examples


pd1 <- pdBlocked(list(~ age, ~ Sex - 1))
formula(pd1)
formula(pd1, asList = TRUE)

[Package nlme version 3.1-137 Index]