formula.pdMat {nlme}	R Documentation

Extract pdMat Formula

Description

This method function extracts the formula associated with a pdMat object, in which the column and row names are specified.

Usage


## S3 method for class 'pdMat'
formula(x, asList, ...)

Arguments

x
an object inheriting from class "pdMat", representing a positive definite matrix.

asList
logical. Should the asList argument be applied to each of the components? Never used.

...
some methods for this generic require additional arguments. None are used in this method.

Value

if x has a formula attribute, its value is returned, else NULL is returned.

Note

Because factors may be present in formula(x), the pdMat object needs to have access to a data frame where the variables named in the formula can be evaluated, before it can resolve its row and column names from the formula.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

pdMat

Examples


pd1 <- pdSymm(~Sex*age)
formula(pd1)

[Package nlme version 3.1-137 Index]