getCovariate.data.frame {nlme}	R Documentation

Extract Data Frame Covariate

Description

The right hand side of form, stripped of any conditioning expression (i.e. an expression following a | operator), is evaluated in object.

Usage


## S3 method for class 'data.frame'
getCovariate(object, form, data)

Arguments

object
an object inheriting from class data.frame.

form
an optional formula specifying the covariate to be evaluated in object. Defaults to formula(object).

data
some methods for this generic require a separate data frame. Not used in this method.

Value

the value of the right hand side of form, stripped of any conditional expression, evaluated in object.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

getCovariateFormula

Examples


getCovariate(Orthodont)

[Package nlme version 3.1-137 Index]