getData.gls {nlme}	R Documentation

Extract gls Object Data

Description

If present in the calling sequence used to produce object, the data frame used to fit the model is obtained.

Usage

## S3 method for class 'gls'
getData(object)

Arguments

object
an object inheriting from class gls, representing a generalized least squares fitted linear model.

Value

if a data argument is present in the calling sequence that produced object, the corresponding data frame (with na.action and subset applied to it, if also present in the call that produced object) is returned; else, NULL is returned.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

gls, getData

Examples

fm1 <- gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), data = Ovary,
           correlation = corAR1(form = ~ 1 | Mare))
getData(fm1)

[Package nlme version 3.1-121 Index]