getData.lmList {nlme}	R Documentation

Extract lmList Object Data

Description

If present in the calling sequence used to produce object, the data frame used to fit the model is obtained.

Usage


## S3 method for class 'lmList'
getData(object)

Arguments

object
an object inheriting from class lmList, representing a list of lm objects with a common model.

Value

if a data argument is present in the calling sequence that produced object, the corresponding data frame (with na.action and subset applied to it, if also present in the call that produced object) is returned; else, NULL is returned.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

lmList, getData

Examples


fm1 <- lmList(distance ~ age | Subject, Orthodont)
getData(fm1)

[Package nlme version 3.1-137 Index]