getGroups.corStruct {nlme}	R Documentation

Extract corStruct Groups

Description

This method function extracts the grouping factor associated with object, if any is present.

Usage

## S3 method for class 'corStruct'
getGroups(object, form, level, data, sep)

Arguments

object
an object inheriting from class corStruct representing a correlation structure.

form
this argument is included to make the method function compatible with the generic. It will be assigned the value of formula(object) and should not be modified.

level
this argument is included to make the method function compatible with the generic and is not used.

data
an optional data frame in which to evaluate the variables defined in form, in case object is not initialized and the grouping factor needs to be evaluated.

sep
character, the separator to use between group levels when multiple levels are collapsed. The default is '/'.

Value

if a grouping factor is present in the correlation structure represented by object, the function returns the corresponding factor vector; else the function returns NULL.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

getGroups

Examples

cs1 <- corAR1(form = ~ 1 | Subject)
getGroups(cs1, data = Orthodont)

[Package nlme version 3.1-121 Index]