getGroups.data.frame {nlme}	R Documentation

Extract Groups from a Data Frame

Description

Each variable named in the expression after the | operator on the right hand side of form is evaluated in object. If more than one variable is indicated in level they are combined into a data frame; else the selected variable is returned as a vector. When multiple grouping levels are defined in form and level > 1, the levels of the returned factor are obtained by pasting together the levels of the grouping factors of level greater or equal to level, to ensure their uniqueness.

Usage

## S3 method for class 'data.frame'
getGroups(object, form, level, data, sep)

Arguments

object
an object inheriting from class data.frame.

form
an optional formula with a conditioning expression on its right hand side (i.e. an expression involving the | operator). Defaults to formula(object).

level
a positive integer vector with the level(s) of grouping to be used when multiple nested levels of grouping are present. Defaults to all levels of nesting.

data
unused

sep
character, the separator to use between group levels when multiple levels are collapsed. The default is '/'.

Value

either a data frame with columns given by the grouping factors indicated in level, from outer to inner, or, when a single level is requested, a factor representing the selected grouping factor.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer, esp. pp. 100, 461.

See Also

getGroupsFormula

Examples

getGroups(Pixel)
getGroups(Pixel, level = 2)

[Package nlme version 3.1-121 Index]