getGroups.gls {nlme}	R Documentation

Extract gls Object Groups

Description

If present, the grouping factor associated to the correlation structure for the linear model represented by object is extracted.

Usage


## S3 method for class 'gls'
getGroups(object, form, level, data, sep)

Arguments

object
an object inheriting from class gls, representing a generalized least squares fitted linear model.

form
an optional formula with a conditioning expression on its right hand side (i.e. an expression involving the | operator). Defaults to formula(object). Not used.

level
a positive integer vector with the level(s) of grouping to be used when multiple nested levels of grouping are present. This argument is optional for most methods of this generic function and defaults to all levels of nesting. Not used.

data
a data frame in which to interpret the variables named in form. Optional for most methods. Not used.

sep
character, the separator to use between group levels when multiple levels are collapsed. The default is '/'. Not used.

Value

if the linear model represented by object incorporates a correlation structure and the corresponding corStruct object has a grouping factor, a vector with the group values is returned; else, NULL is returned.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

gls, corClasses

Examples


fm1 <- gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), Ovary,
           correlation = corAR1(form = ~ 1 | Mare))
getGroups(fm1)

[Package nlme version 3.1-137 Index]