getGroups.lmList {nlme}	R Documentation

Extract lmList Object Groups

Description

The grouping factor determining the partitioning of the observations used to produce the lm components of object is extracted.

Usage

## S3 method for class 'lmList':
getGroups(object, form, level, data, sep)

Arguments

object 	 an object inheriting from class lmList, representing a list of lm objects with a common model.

form 	 an optional formula with a conditioning expression on its right hand side (i.e. an expression involving the | operator). Defaults to formula(object). Not used.

level 	 a positive integer vector with the level(s) of grouping to be used when multiple nested levels of grouping are present. This argument is optional for most methods of this generic function and defaults to all levels of nesting. Not used.

data 	 a data frame in which to interpret the variables named in form. Optional for most methods. Not used.

sep 	 character, the separator to use between group levels when multiple levels are collapsed. The default is '/'. Not used.

Value

a vector with the grouping factor corresponding to the lm components of object.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

lmList

Examples

fm1 <- lmList(distance ~ age | Subject, Orthodont)
getGroups(fm1)

[Package nlme version 3.1-93 Index]