getResponseFormula {nlme}	R Documentation

Extract Formula Specifying Response Variable

Description

The left hand side of formula{object} is returned as a one-sided formula.

Usage


getResponseFormula(object)

Arguments

object
any object from which a formula can be extracted.

Value

a one-sided formula with the response variable associated with formula{object}.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

getResponse

Examples


getResponseFormula(y ~ x | g)

[Package nlme version 3.1-137 Index]