glsControl {nlme}	R Documentation

Control Values for gls Fit

Description

The values supplied in the function call replace the defaults and a list with all possible arguments is returned. The returned list is used as the control argument to the gls function.

Usage

glsControl(maxIter, msMaxIter, tolerance, msTol, msScale, msVerbose,
           singular.ok, qrTol, returnObject, apVar, .relStep,
           nlmStepMax, opt=c("nlminb", "optim"), optimMethod, 
           minAbsParApVar, natural)

Arguments

maxIter 	 maximum number of iterations for the gls optimization algorithm. Default is 50.

msMaxIter 	 maximum number of iterations for the optimization step inside the gls optimization. Default is 50.

tolerance 	 tolerance for the convergence criterion in the gls algorithm. Default is 1e-6.

msTol 	 tolerance for the convergence criterion in ms, passed as the rel.tolerance argument to the function (see documentation on ms). Default is 1e-7.

msScale 	 scale function passed as the scale argument to the ms function (see documentation on that function). Default is lmeScale.

msVerbose 	 a logical value passed as the trace argument to ms (see documentation on that function). Default is FALSE.

singular.ok 	 a logical value indicating whether non-estimable coefficients (resulting from linear dependencies among the columns of the regression matrix) should be allowed. Default is FALSE.

qrTol 	 a tolerance for detecting linear dependencies among the columns of the regression matrix in its QR decomposition. Default is .Machine$single.eps.

returnObject 	 a logical value indicating whether the fitted object should be returned when the maximum number of iterations is reached without convergence of the algorithm. Default is FALSE.

apVar 	 a logical value indicating whether the approximate covariance matrix of the variance-covariance parameters should be calculated. Default is TRUE.

.relStep 	 relative step for numerical derivatives calculations. Default is .Machine$double.eps^(1/3).

nlmStepMax 	 stepmax value to be passed to nlm. See nlm for details. Default is 100.0

opt 	 the optimizer to be used, either nlminb (the default since (R 2.2.0) or optim (the previous default).

optimMethod 	 character - the optimization method to be used with the optim optimizer. The default is "BFGS". An alternative is "L-BFGS-B".

minAbsParApVar 	 numeric value - minimum absolute parameter value in the approximate variance calculation. The default is 0.05.

natural 	 logical. Should the natural parameterization be used for the approximate variance calculations? Default is TRUE.

Value

a list with components for each of the possible arguments.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

gls, lmeScale

Examples

# decrease the maximum number iterations in the optimization call and
# request that information on the evolution of the ms iterations be printed
glsControl(msMaxIter = 20, msVerbose = TRUE)

[Package nlme version 3.1-93 Index]