gnlsObject {nlme}	R Documentation

Fitted gnls Object

Description

An object returned by the gnls function, inheriting from class gnls and also from class gls, and representing a generalized nonlinear least squares fitted model. Objects of this class have methods for the generic functions anova, coef, fitted, formula, getGroups, getResponse, intervals, logLik, plot, predict, print, residuals, summary, and update.

Value

The following components must be included in a legitimate gnls object.

apVar 	 an approximate covariance matrix for the variance-covariance coefficients. If apVar = FALSE in the list of control values used in the call to gnls, this component is equal to NULL.

call 	 a list containing an image of the gnls call that produced the object.

coefficients 	 a vector with the estimated nonlinear model coefficients.

contrasts 	 a list with the contrasts used to represent factors in the model formula. This information is important for making predictions from a new data frame in which not all levels of the original factors are observed. If no factors are used in the model, this component will be an empty list.

dims 	 a list with basic dimensions used in the model fit, including the components N - the number of observations used in the fit and p - the number of coefficients in the nonlinear model.

fitted 	 a vector with the fitted values.

modelStruct 	 an object inheriting from class gnlsStruct, representing a list of model components, such as corStruct and varFunc objects.

groups 	 a vector with the correlation structure grouping factor, if any is present.

logLik 	 the log-likelihood at convergence.

numIter 	 the number of iterations used in the iterative algorithm.

plist

pmap

residuals 	 a vector with the residuals.

sigma 	 the estimated residual standard error.

varBeta 	 an approximate covariance matrix of the coefficients estimates.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

gnls, gnlsStruct

[Package nlme version 3.1-93 Index]