intervals.gls {nlme}	R Documentation

Confidence Intervals on gls Parameters

Description

Approximate confidence intervals for the parameters in the linear model represented by object are obtained, using a normal approximation to the distribution of the (restricted) maximum likelihood estimators (the estimators are assumed to have a normal distribution centered at the true parameter values and with covariance matrix equal to the negative inverse Hessian matrix of the (restricted) log-likelihood evaluated at the estimated parameters). Confidence intervals are obtained in an unconstrained scale first, using the normal approximation, and, if necessary, transformed to the constrained scale.

Usage

## S3 method for class 'gls':
intervals(object, level, which, ...)

Arguments

object 	 an object inheriting from class gls, representing a generalized least squares fitted linear model.

level 	 an optional numeric value for the interval confidence level. Defaults to 0.95.

which 	 an optional character string specifying the subset of parameters for which to construct the confidence intervals. Possible values are "all" for all parameters, "var-cov" for the variance-covariance parameters only, and "coef" for the linear model coefficients only. Defaults to "all".

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

a list with components given by data frames with rows corresponding to parameters and columns lower, est., and upper representing respectively lower confidence limits, the estimated values, and upper confidence limits for the parameters. Possible components are:

coef 	 linear model coefficients, only present when which is not equal to "var-cov".

corStruct 	 correlation parameters, only present when which is not equal to "coef" and a correlation structure is used in object.

varFunc 	 variance function parameters, only present when which is not equal to "coef" and a variance function structure is used in object.

sigma 	 residual standard error.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

gls, intervals, print.intervals.gls

Examples

fm1 <- gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), Ovary,
           correlation = corAR1(form = ~ 1 | Mare))
intervals(fm1)

[Package nlme version 3.1-93 Index]