intervals.lmList {nlme}	R Documentation

Confidence Intervals on lmList Coefficients

Description

Confidence intervals on the linear model coefficients are obtained for each lm component of object and organized into a three dimensional array. The first dimension corresponding to the names of the object components. The second dimension is given by lower, est., and upper corresponding, respectively, to the lower confidence limit, estimated coefficient, and upper confidence limit. The third dimension is given by the coefficients names.

Usage

## S3 method for class 'lmList':
intervals(object, level, pool, ...)

Arguments

object 	 an object inheriting from class lmList, representing a list of lm objects with a common model.

level 	 an optional numeric value with the confidence level for the intervals. Defaults to 0.95.

pool 	 an optional logical value indicating whether a pooled estimate of the residual standard error should be used. Default is attr(object, "pool").

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

a three dimensional array with the confidence intervals and estimates for the coefficients of each lm component of object.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

lmList, intervals, plot.intervals.lmList

Examples

fm1 <- lmList(distance ~ age | Subject, Orthodont)
intervals(fm1)

[Package nlme version 3.1-93 Index]