intervals.lme {nlme}	R Documentation

Confidence Intervals on lme Parameters

Description

Approximate confidence intervals for the parameters in the linear mixed-effects model represented by object are obtained, using a normal approximation to the distribution of the (restricted) maximum likelihood estimators (the estimators are assumed to have a normal distribution centered at the true parameter values and with covariance matrix equal to the negative inverse Hessian matrix of the (restricted) log-likelihood evaluated at the estimated parameters). Confidence intervals are obtained in an unconstrained scale first, using the normal approximation, and, if necessary, transformed to the constrained scale. The pdNatural parametrization is used for general positive-definite matrices.

Usage

## S3 method for class 'lme':
intervals(object, level, which, ...)

Arguments

object 	 an object inheriting from class lme, representing a fitted linear mixed-effects model.

level 	 an optional numeric value with the confidence level for the intervals. Defaults to 0.95.

which 	 an optional character string specifying the subset of parameters for which to construct the confidence intervals. Possible values are "all" for all parameters, "var-cov" for the variance-covariance parameters only, and "fixed" for the fixed effects only. Defaults to "all".

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

a list with components given by data frames with rows corresponding to parameters and columns lower, est., and upper representing respectively lower confidence limits, the estimated values, and upper confidence limits for the parameters. Possible components are:

fixed 	 fixed effects, only present when which is not equal to "var-cov".

reStruct 	 random effects variance-covariance parameters, only present when which is not equal to "fixed".

corStruct 	 within-group correlation parameters, only present when which is not equal to "fixed" and a correlation structure is used in object.

varFunc 	 within-group variance function parameters, only present when which is not equal to "fixed" and a variance function structure is used in object.

sigma 	 within-group standard deviation.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

lme, intervals, print.intervals.lme, pdNatural

Examples

fm1 <- lme(distance ~ age, Orthodont, random = ~ age | Subject)
intervals(fm1)

[Package nlme version 3.1-93 Index]