isInitialized {nlme}	R Documentation

Check if Object is Initialized

Description

Checks if object has been initialized (generally through a call to Initialize), by searching for components and attributes which are modified during initialization.

Usage

isInitialized(object)

Arguments

object 	 any object requiring initialization.

Value

a logical value indicating whether object has been initialized.

Author(s)

Jose Pinheiro and Douglas Bates

See Also

Initialize

Examples

pd1 <- pdDiag(~age)
isInitialized(pd1)

[Package nlme version 3.1-93 Index]