lmList.groupedData {nlme}	R Documentation

lmList Fit from a groupedData Object

Description

The response variable and primary covariate in formula(object) are used to construct the linear model formula. This formula and the groupedData object are passed as the object and data arguments to lmList.formula, together with any other additional arguments in the function call. See the documentation on lmList.formula for a description of that function.

Usage

## S3 method for class 'groupedData':
lmList(object, data, level, subset, na.action, pool)

Arguments

object 	 a data frame inheriting from class groupedData.

data 	 this argument is included for consistency with the generic function. It is ignored in this method function.

level 	 an optional integer specifying the level of grouping to be used when multiple nested levels of grouping are present.

subset 	 an optional expression indicating which subset of the rows of data should be used in the fit. This can be a logical vector, or a numeric vector indicating which observation numbers are to be included, or a character vector of the row names to be included. All observations are included by default.

na.action 	 a function that indicates what should happen when the data contain NAs. The default action (na.fail) causes lmList to print an error message and terminate if there are any incomplete observations.

pool 	 an optional logical value that is preserved as an attribute of the returned value. This will be used as the default for pool in calculations of standard deviations or standard errors for summaries.

Value

a list of lm objects with as many components as the number of groups defined by the grouping factor. Generic functions such as coef, fixed.effects, lme, pairs, plot, predict, random.effects, summary, and update have methods that can be applied to an lmList object.

See Also

groupedData, lm, lme.lmList, lmList, lmList.formula

Examples

fm1 <- lmList(Orthodont)
summary(fm1)

[Package nlme version 3.1-93 Index]