lmeControl {nlme}	R Documentation

Control Values for lme Fit

Description

The values supplied in the function call replace the defaults and a list with all possible arguments is returned. The returned list is used as the control argument to the lme function.

Usage

lmeControl(maxIter, msMaxIter, tolerance, niterEM, msMaxEval, msTol,
           msScale, msVerbose, returnObject, gradHess, apVar,
           .relStep, minAbsParApVar, nlmStepMax,
           opt = c("nlminb", "optim"), optimMethod,
           natural)

Arguments

maxIter 	 maximum number of iterations for the lme optimization algorithm. Default is 50.

msMaxIter 	 maximum number of iterations for the nlm optimization step inside the lme optimization. Default is 50.

tolerance 	 tolerance for the convergence criterion in the lme algorithm. Default is 1e-6.

niterEM 	 number of iterations for the EM algorithm used to refine the initial estimates of the random effects variance-covariance coefficients. Default is 25.

msMaxEval 	 maximum number of evaluations of the objective function permitted for nlminb. Default is 200.

msTol 	 tolerance for the convergence criterion in nlm, passed as the rel.tolerance argument to the function (see documentation on nlm). Default is 1e-7.

msScale 	 scale function passed as the scale argument to the nlm function (see documentation on that function). Default is lmeScale.

msVerbose 	 a logical value passed as the trace argument to nlm (see documentation on that function). Default is FALSE.

returnObject 	 a logical value indicating whether the fitted object should be returned when the maximum number of iterations is reached without convergence of the algorithm. Default is FALSE.

gradHess 	 a logical value indicating whether numerical gradient vectors and Hessian matrices of the log-likelihood function should be used in the nlm optimization. This option is only available when the correlation structure (corStruct) and the variance function structure (varFunc) have no "varying" parameters and the pdMat classes used in the random effects structure are pdSymm (general positive-definite), pdDiag (diagonal), pdIdent (multiple of the identity), or pdCompSymm (compound symmetry). Default is TRUE.

apVar 	 a logical value indicating whether the approximate covariance matrix of the variance-covariance parameters should be calculated. Default is TRUE.

.relStep 	 relative step for numerical derivatives calculations. Default is .Machine$double.eps^(1/3).

nlmStepMax 	 stepmax value to be passed to nlm. See nlm for details. Default is 100.0

opt 	 the optimizer to be used, either nlminb (the default since (R 2.2.0) or optim (the previous default).

optimMethod 	 character - the optimization method to be used with the optim optimizer. The default is "BFGS". An alternative is "L-BFGS-B".

minAbsParApVar 	 numeric value - minimum absolute parameter value in the approximate variance calculation. The default is 0.05.

natural 	 a logical value indicating whether the pdNatural parametrization should be used for general positive-definite matrices (pdSymm) in reStruct, when the approximate covariance matrix of the estimators is calculated. Default is TRUE.

Value

a list with components for each of the possible arguments.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

lme, nlm, optim, lmeScale

Examples

# decrease the maximum number iterations in the ms call and
# request that information on the evolution of the ms iterations be printed
lmeControl(msMaxIter = 20, msVerbose = TRUE)

[Package nlme version 3.1-93 Index]