logLik.lmeStruct {nlme}	R Documentation

Log-Likelihood of an lmeStruct Object

Description

Pars is used to update the coefficients of the model components of object and the individual (restricted) log-likelihood contributions of each component are added together. The type of log-likelihood (restricted or not) is determined by the settings attribute of object.

Usage

## S3 method for class 'lmeStruct':
logLik(object, Pars, conLin, ...)

Arguments

object 	 an object inheriting from class lmeStruct, representing a list of linear mixed-effects model components, such as reStruct, corStruct, and varFunc objects.

Pars 	 the parameter values at which the (restricted) log-likelihood is to be evaluated.

conLin 	 an optional condensed linear model object, consisting of a list with components "Xy", corresponding to a regression matrix (X) combined with a response vector (y), and "logLik", corresponding to the log-likelihood of the underlying lme model. Defaults to attr(object, "conLin").

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

the (restricted) log-likelihood for the linear mixed-effects model described by object, evaluated at Pars.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

lme, lmeStruct, logLik.lme

[Package nlme version 3.1-93 Index]