logLik.varFunc {nlme}	R Documentation

Extract varFunc logLik

Description

This method function extracts the component of a Gaussian log-likelihood associated with the variance function structure represented by object, which is equal to the sum of the logarithms of the corresponding weights.

Usage

## S3 method for class 'varFunc':
logLik(object, data, ...)

Arguments

object 	 an object inheriting from class varFunc, representing a variance function structure.

data 	 this argument is included to make this method function compatible with other logLik methods and will be ignored.

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

the sum of the logarithms of the weights corresponding to the variance function structure represented by object.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

logLik.lme

Examples

vf1 <- varPower(form = ~age)
vf1 <- Initialize(vf1, Orthodont)
coef(vf1) <- 0.1
logLik(vf1)

[Package nlme version 3.1-93 Index]