needUpdate.modelStruct {nlme}	R Documentation

Check if a modelStruct Object Needs Updating

Description

This method function checks if any of the elements of object needs to be updated. Updating of objects usually takes place in iterative algorithms in which auxiliary quantities associated with the object, and not being optimized over, may change.

Usage

## S3 method for class 'modelStruct':
needUpdate(object)

Arguments

object 	 an object inheriting from class modelStruct, representing a list of model components, such as corStruct and varFunc objects.

Value

a logical value indicating whether any element of object needs to be updated.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

needUpdate

Examples

lms1 <- lmeStruct(reStruct = reStruct(pdDiag(diag(2), ~age)),
   varStruct = varPower(form = ~age))
needUpdate(lms1)

[Package nlme version 3.1-93 Index]