nlmeControl {nlme}	R Documentation

Control Values for nlme Fit

Description

The values supplied in the function call replace the defaults and a list with all possible arguments is returned. The returned list is used as the control argument to the nlme function.

Usage

nlmeControl(maxIter, pnlsMaxIter, msMaxIter, minScale,
            tolerance, niterEM, pnlsTol, msTol,
            returnObject, msVerbose, gradHess, apVar, .relStep,
            minAbsParApVar = 0.05,
            opt = c("nlminb", "nlm"), natural = TRUE, ...)

Arguments

maxIter
maximum number of iterations for the nlme optimization algorithm. Default is 50.

pnlsMaxIter
maximum number of iterations for the PNLS optimization step inside the nlme optimization. Default is 7.

msMaxIter
maximum number of iterations for the nlm optimization step inside the nlme optimization. Default is 50.

minScale
minimum factor by which to shrink the default step size in an attempt to decrease the sum of squares in the PNLS step. Default 0.001.

tolerance
tolerance for the convergence criterion in the nlme algorithm. Default is 1e-6.

niterEM
number of iterations for the EM algorithm used to refine the initial estimates of the random effects variance-covariance coefficients. Default is 25.

pnlsTol
tolerance for the convergence criterion in PNLS step. Default is 1e-3.

msTol
tolerance for the convergence criterion in nlm, passed as the gradtol argument to the function (see documentation on nlm). Default is 1e-7.

returnObject
a logical value indicating whether the fitted object should be returned when the maximum number of iterations is reached without convergence of the algorithm. Default is FALSE.

msVerbose
a logical value passed as the trace argument to nlm (see documentation on that function). Default is FALSE.

gradHess
a logical value indicating whether numerical gradient vectors and Hessian matrices of the log-likelihood function should be used in the nlm optimization. This option is only available when the correlation structure (corStruct) and the variance function structure (varFunc) have no "varying" parameters and the pdMat classes used in the random effects structure are pdSymm (general positive-definite), pdDiag (diagonal), pdIdent (multiple of the identity), or pdCompSymm (compound symmetry). Default is TRUE.

apVar
a logical value indicating whether the approximate covariance matrix of the variance-covariance parameters should be calculated. Default is TRUE.

.relStep
relative step for numerical derivatives calculations. Default is .Machine$double.eps^(1/3).

minAbsParApVar
numeric value - minimum absolute parameter value in the approximate variance calculation. The default is 0.05.

opt
the optimizer to be used, either "nlminb" (the default) or "nlm".

natural
a logical value indicating whether the pdNatural parametrization should be used for general positive-definite matrices (pdSymm) in reStruct, when the approximate covariance matrix of the estimators is calculated. Default is TRUE.

...
Further named control arguments to be passed to nlminb, where used (eval.max and those from abs.tol down).

Value

a list with components for each of the possible arguments.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

nlme, nlm, optim, nlmeStruct

Examples

# decrease the maximum number iterations in the ms call and
# request that information on the evolution of the ms iterations be printed
nlmeControl(msMaxIter = 20, msVerbose = TRUE)

[Package nlme version 3.1-121 Index]