nlmeStruct {nlme}	R Documentation

Nonlinear Mixed-Effects Structure

Description

A nonlinear mixed-effects structure is a list of model components representing different sets of parameters in the nonlinear mixed-effects model. An nlmeStruct list must contain at least a reStruct object, but may also contain corStruct and varFunc objects. NULL arguments are not included in the nlmeStruct list.

Usage


nlmeStruct(reStruct, corStruct, varStruct)

Arguments

reStruct
a reStruct representing a random effects structure.

corStruct
an optional corStruct object, representing a correlation structure. Default is NULL.

varStruct
an optional varFunc object, representing a variance function structure. Default is NULL.

Value

a list of model components determining the parameters to be estimated for the associated nonlinear mixed-effects model.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

corClasses, nlme, residuals.nlmeStruct, reStruct, varFunc

Examples


nlms1 <- nlmeStruct(reStruct(~age), corAR1(), varPower())

[Package nlme version 3.1-137 Index]