nlsList.selfStart {nlme}	R Documentation

nlsList Fit from a selfStart Function

Description

The response variable and primary covariate in formula(data) are used together with model to construct the nonlinear model formula. This is used in the nls calls and, because a selfStarting model function can calculate initial estimates for its parameters from the data, no starting estimates need to be provided.

Usage


## S3 method for class 'selfStart'
nlsList(model, data, start, control, level, subset,
        na.action = na.fail, pool = TRUE, warn.nls = NA)

Arguments

model
a "selfStart" model function, which calculates initial estimates for the model parameters from data.

data
a data frame in which to interpret the variables in model. Because no grouping factor can be specified in model, data must inherit from class "groupedData".

start
an optional named list with initial values for the parameters to be estimated in model. It is passed as the start argument to each nls call and is required when the nonlinear function in model does not inherit from class selfStart.

control
a list of control values passed as the control argument to nls. Defaults to an empty list.

level
an optional integer specifying the level of grouping to be used when multiple nested levels of grouping are present.

subset
an optional expression indicating the subset of the rows of data that should be used in the fit. This can be a logical vector, or a numeric vector indicating which observation numbers are to be included, or a character vector of the row names to be included. All observations are included by default.

na.action
a function that indicates what should happen when the data contain NAs. The default action (na.fail) causes nlsList to print an error message and terminate if there are any incomplete observations.

pool, warn.nls
optional logicals, see nlsList.

Value

a list of nls objects with as many components as the number of groups defined by the grouping factor. A NULL value is assigned to the components corresponding to clusters for which the nls algorithm failed to converge. Generic functions such as coef, fixed.effects, lme, pairs, plot, predict, random.effects, summary, and update have methods that can be applied to an nlsList object.

See Also

selfStart, groupedData, nls, nlsList, nlme.nlsList, nlsList.formula

Examples


fm1 <- nlsList(SSasympOff, CO2)
summary(fm1)

[Package nlme version 3.1-137 Index]