pdFactor {nlme}	R Documentation

Square-Root Factor of a Positive-Definite Matrix

Description

A square-root factor of the positive-definite matrix represented by object is obtained. Letting S denote a positive-definite matrix, a square-root factor of S is any square matrix L such that S = L'L. This function extracts L.

Usage


pdFactor(object)

Arguments

object
an object inheriting from class pdMat, representing a positive definite matrix, which must have been initialized (i.e. length(coef(object)) > 0).

Value

a vector with a square-root factor of the positive-definite matrix associated with object stacked column-wise.

Note

This function is used intensively in optimization algorithms and its value is returned as a vector for efficiency reasons. The pdMatrix function can be used to obtain square-root factors in matrix form.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

pdMatrix

Examples


pd1 <- pdCompSymm(4 * diag(3) + 1)
pdFactor(pd1)

[Package nlme version 3.1-137 Index]