plot.ACF {nlme}	R Documentation

Plot an ACF Object

Description

an xyplot of the autocorrelations versus the lags, with type = "h", is produced. If alpha > 0, curves representing the critical limits for a two-sided test of level alpha for the autocorrelations are added to the plot.

Usage

## S3 method for class 'ACF':
plot(x, alpha, xlab, ylab, grid, ...)

Arguments

x 	 an object inheriting from class ACF, consisting of a data frame with two columns named lag and ACF, representing the autocorrelation values and the corresponding lags.

alpha 	 an optional numeric value with the significance level for testing if the autocorrelations are zero. Lines corresponding to the lower and upper critical values for a test of level alpha are added to the plot. Default is 0, in which case no lines are plotted.

xlab,ylab 	 optional character strings with the x- and y-axis labels. Default respectively to "Lag" and "Autocorrelation".

grid 	 an optional logical value indicating whether a grid should be added to plot. Default is FALSE.

... 	 optional arguments passed to the xyplot function.

Value

an xyplot Trellis plot.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

ACF, xyplot

Examples

fm1 <- lme(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), Ovary)
plot(ACF(fm1, maxLag = 10), alpha = 0.01)

[Package nlme version 3.1-93 Index]