plot.Variogram {nlme}	R Documentation

Plot a Variogram Object

Description

an xyplot of the semi-variogram versus the distances is produced. If smooth = TRUE, a loess smoother is added to the plot. If showModel = TRUE and x includes an "modelVariog" attribute, the corresponding semi-variogram is added to the plot.

Usage

## S3 method for class 'Variogram':
plot(x, smooth, showModel, sigma, span, xlab,
     ylab, type, ylim, grid, ...)

Arguments

x 	 an object inheriting from class Variogram, consisting of a data frame with two columns named variog and dist, representing the semi-variogram values and the corresponding distances.

smooth 	 an optional logical value controlling whether a loess smoother should be added to the plot. Defaults to TRUE, when showModel is FALSE.

showModel 	 an optional logical value controlling whether the semi-variogram corresponding to an "modelVariog" attribute of x, if any is present, should be added to the plot. Defaults to TRUE, when the "modelVariog" attribute is present.

sigma 	 an optional numeric value used as the height of a horizontal line displayed in the plot. Can be used to represent the process standard deviation. Default is NULL, implying that no horizontal line is drawn.

span 	 an optional numeric value with the smoothing parameter for the loess fit. Default is 0.6.

xlab,ylab 	 optional character strings with the x- and y-axis labels. Default respectively to "Distance" and "SemiVariogram".

type 	 an optional character indicating the type of plot. Defaults to "p".

ylim 	 an optional numeric vector with the limits for the y-axis. Defaults to c(0, max(x$variog)).

grid 	 an optional logical value indicating whether a grid should be added to plot. Default is FALSE.

... 	 optional arguments passed to the Trellis xyplot function.

Value

an xyplot Trellis plot.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

Variogram, xyplot, loess

Examples

fm1 <- lme(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), Ovary)
plot(Variogram(fm1, form = ~ Time | Mare, maxDist = 0.7))

[Package nlme version 3.1-93 Index]