plot.compareFits {nlme}	R Documentation

Plot a compareFits Object

Description

A Trellis dotplot of the values being compared, with different rows per group, is generated, with a different panel for each coefficient. Different symbols (colors) are used for each object being compared.

Usage


## S3 method for class 'compareFits'
plot(x, subset, key, mark, ...)

Arguments

x
an object of class "compareFits".

subset
an optional logical or integer vector specifying which rows of x should be used in the plots. If missing, all rows are used.

key
an optional logical value, or list. If TRUE, a legend is included at the top of the plot indicating which symbols (colors) correspond to which objects being compared. If FALSE, no legend is included. If given as a list, key is passed down as an argument to the trellis function generating the plots (dotplot). Defaults to TRUE.

mark
an optional numeric vector, of length equal to the number of coefficients being compared, indicating where vertical lines should be drawn in the plots. If missing, no lines are drawn.

...
optional arguments passed down to the trellis function generating the plots.

Value

A Trellis dotplot of the values being compared, with rows determined by the groups and panels by the coefficients.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

compareFits, pairs.compareFits, dotplot

Examples


## Not run: 
fm1 <- lmList(Orthodont)
fm2 <- lme(Orthodont)
plot(compareFits(coef(fm1), coef(fm2)))

## End(Not run)

[Package nlme version 3.1-137 Index]