plot.intervals.lmList {nlme}	R Documentation

Plot lmList Confidence Intervals

Description

A Trellis dot-plot of the confidence intervals on the linear model coefficients is generated, with a different panel for each coefficient. Rows in the dot-plot correspond to the names of the lm components of the lmList object used to produce x. The lower and upper confidence limits are connected by a line segment and the estimated coefficients are marked with a "+". The Trellis function dotplot is used in this method function.

Usage

## S3 method for class 'intervals.lmList':
plot(x, ...)

Arguments

x 	 an object inheriting from class intervals.lmList, representing confidence intervals and estimates for the coefficients in the lm components of the lmList object used to produce x.

... 	 optional arguments passed to the Trellis dotplot function.

Value

a Trellis plot with the confidence intervals on the coefficients of the individual lm components of the lmList that generated x.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

intervals.lmList, lmList, dotplot

Examples

fm1 <- lmList(distance ~ age | Subject, Orthodont)
plot(intervals(fm1))

[Package nlme version 3.1-93 Index]