plot.intervals.lmList {nlme}	R Documentation

Plot lmList Confidence Intervals

Description

A Trellis dot-plot of the confidence intervals on the linear model coefficients is generated, with a different panel for each coefficient. Rows in the dot-plot correspond to the names of the lm components of the lmList object used to produce x. The lower and upper confidence limits are connected by a line segment and the estimated coefficients are marked with a "+".

This is based on function dotplot() from package lattice.

Usage

## S3 method for class 'intervals.lmList'
plot(x, xlab = "", ylab = attr(x, "groupsName"),
     strip = function(...) strip.default(..., style = 1),
     ...)

Arguments

x
an object inheriting from class "intervals.lmList", representing confidence intervals and estimates for the coefficients in the lm components of the lmList object used to produce x.

xlab, ylab
axis labels, each with a sensible default.

strip
a function or FALSE, see dotplot() from package lattice.

...
optional arguments passed to the dotplot function (see above).

Value

a Trellis plot with the confidence intervals on the coefficients of the individual lm components of the lmList that generated x.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

intervals.lmList, lmList, dotplot

Examples

fm1 <- lmList(distance ~ age | Subject, Orthodont)
plot(intervals(fm1))

[Package nlme version 3.1-147 Index]