pooledSD {nlme}	R Documentation

Extract Pooled Standard Deviation

Description

The pooled estimated standard deviation is obtained by adding together the residual sum of squares for each non-null element of object, dividing by the sum of the corresponding residual degrees-of-freedom, and taking the square-root.

Usage


pooledSD(object)

Arguments

object
an object inheriting from class lmList.

Value

the pooled standard deviation for the non-null elements of object, with an attribute df with the number of degrees-of-freedom used in the estimation.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

lmList, lm

Examples


fm1 <- lmList(Orthodont)
pooledSD(fm1)

[Package nlme version 3.1-137 Index]