predict.gls {nlme}	R Documentation

Predictions from a gls Object

Description

The predictions for the linear model represented by object are obtained at the covariate values defined in newdata.

Usage

## S3 method for class 'gls':
predict(object, newdata, na.action, ...)

Arguments

object 	 an object inheriting from class gls, representing a generalized least squares fitted linear model.

newdata 	 an optional data frame to be used for obtaining the predictions. All variables used in the linear model must be present in the data frame. If missing, the fitted values are returned.

na.action 	 a function that indicates what should happen when newdata contains NAs. The default action (na.fail) causes the function to print an error message and terminate if there are any incomplete observations.

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

a vector with the predicted values.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

gls

Examples

fm1 <- gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), Ovary,
           correlation = corAR1(form = ~ 1 | Mare))
newOvary <- data.frame(Time = c(-0.75, -0.5, 0, 0.5, 0.75))
predict(fm1, newOvary)

[Package nlme version 3.1-93 Index]