predict.lmList {nlme}	R Documentation

Predictions from an lmList Object

Description

If the grouping factor corresponding to object is included in newdata, the data frame is partitioned according to the grouping factor levels; else, newdata is repeated for all lm components. The predictions and, optionally, the standard errors for the predictions, are obtained for each lm component of object, using the corresponding element of the partitioned newdata, and arranged into a list with as many components as object, or combined into a single vector or data frame (if se.fit=TRUE).

Usage

## S3 method for class 'lmList':
predict(object, newdata, subset, pool, asList, se.fit, ...)

Arguments

object 	 an object inheriting from class lmList, representing a list of lm objects with a common model.

newdata 	 an optional data frame to be used for obtaining the predictions. All variables used in the object model formula must be present in the data frame. If missing, the same data frame used to produce object is used.

subset 	 an optional character or integer vector naming the lm components of object from which the predictions are to be extracted. Default is NULL, in which case all components are used.

asList 	 an optional logical value. If TRUE, the returned object is a list with the predictions split by groups; else the returned value is a vector. Defaults to FALSE.

pool 	 an optional logical value indicating whether a pooled estimate of the residual standard error should be used. Default is attr(object, "pool").

se.fit 	 an optional logical value indicating whether pointwise standard errors should be computed along with the predictions. Default is FALSE.

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

a list with components given by the predictions (and, optionally, the standard errors for the predictions) from each lm component of object, a vector with the predictions from all lm components of object, or a data frame with columns given by the predictions and their corresponding standard errors.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

lmList, predict.lm

Examples

fm1 <- lmList(distance ~ age | Subject, Orthodont)
predict(fm1, se.fit = TRUE)

[Package nlme version 3.1-93 Index]