print.summary.pdMat {nlme}	R Documentation

Print a summary.pdMat Object

Description

The standard deviations and correlations associated with the positive-definite matrix represented by object (considered as a variance-covariance matrix) are printed, together with the formula and the grouping level associated object, if any are present.

Usage


## S3 method for class 'summary.pdMat'
print(x, sigma, rdig, Level, resid, ...)

Arguments

x
an object inheriting from class "summary.pdMat", generally resulting from applying summary to an object inheriting from class "pdMat".

sigma
an optional numeric value used as a multiplier for the square-root factor of the positive-definite matrix represented by object (usually the estimated within-group standard deviation from a mixed-effects model). Defaults to 1.

rdig
an optional integer value with the number of significant digits to be used in printing correlations. Defaults to 3.

Level
an optional character string with a description of the grouping level associated with object (generally corresponding to levels of grouping in a mixed-effects model). Defaults to NULL.

resid
an optional logical value. If TRUE an extra row with the "residual" standard deviation given in sigma will be included in the output. Defaults to FALSE.

...
optional arguments passed to print.default; see the documentation on that method function.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

summary.pdMat,pdMat

Examples


pd1 <- pdCompSymm(3 * diag(2) + 1, form = ~age + age^2,
         data = Orthodont)
print(summary(pd1), sigma = 1.2, resid = TRUE)

[Package nlme version 3.1-137 Index]