quinModel {nlme}	R Documentation

Model function for the Quinidine data

Description

A model function for a model used with the Quinidine data. This function calls compiled C code.

Usage

quinModel(Subject, time, conc, dose, interval, lV, lKa, lCl)

Arguments

Subject 	 a factor identifying the patient on whom the data were collected.

time 	 a numeric vector giving the time (hr) at which the drug was administered or the blood sample drawn. This is measured from the time the patient entered the study.

conc 	 a numeric vector giving the serum quinidine concentration (mg/L).

dose 	 a numeric vector giving the dose of drug administered (mg). Although there were two different forms of quinidine administered, the doses were adjusted for differences in salt content by conversion to milligrams of quinidine base.

interval 	 a numeric vector giving the when the drug has been given at regular intervals for a sufficiently long period of time to assume steady state behavior, the interval is recorded.

lV 	 numeric. A vector of values of the natural log of the effective volume of distribution according to Subject and time.

lKa 	 numeric. A vector of values of the natural log of the absorption rate constant according to Subject and time.

lCl 	 numeric. A vector of values of the natural log of the clearance parameter according to Subject and time.

Details

See the details section of Quinidine for a description of the model function that quinModel evaluates.

Value

a numeric vector of predicted quinidine concentrations.

Author(s)

Jose Pinheiro jose.pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J. C. and Bates, D. M. (2000) Mixed-effects Models in S and S-PLUS, Springer. (section 8.2)

[Package nlme version 3.1-93 Index]