recalc {nlme}	R Documentation

Recalculate Condensed Linear Model Object

Description

This function is generic; method functions can be written to handle specific classes of objects. Classes which already have methods for this function include: corStruct, modelStruct, reStruct, and varFunc.

Usage


recalc(object, conLin, ...)

Arguments

object
any object which induces a recalculation of the condensed linear model object conLin.

conLin
a condensed linear model object, consisting of a list with components "Xy", corresponding to a regression matrix (X) combined with a response vector (y), and "logLik", corresponding to the log-likelihood of the underlying model.

...
some methods for this generic can take additional arguments.

Value

the recalculated condensed linear model object.

Note

This function is only used inside model fitting functions, such as lme and gls, that require recalculation of a condensed linear model object.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

recalc.corStruct, recalc.modelStruct, recalc.reStruct, recalc.varFunc

Examples


## see the method function documentation

[Package nlme version 3.1-137 Index]