recalc.modelStruct {nlme}	R Documentation

Recalculate for a modelStruct Object

Description

This method function recalculates the condensed linear model object using each element of object sequentially from last to first.

Usage


## S3 method for class 'modelStruct'
recalc(object, conLin, ...)

Arguments

object
an object inheriting from class "modelStruct", representing a list of model components, such as corStruct and varFunc objects.

conLin
an optional condensed linear model object, consisting of a list with components "Xy", corresponding to a regression matrix (X) combined with a response vector (y), and "logLik", corresponding to the log-likelihood of the underlying model. Defaults to attr(object, "conLin").

...
some methods for this generic require additional arguments. None are used in this method.

Value

the recalculated condensed linear model object.

Note

This method function is generally only used inside model fitting functions, such as lme and gls, that allow model components, such as correlated error terms and variance functions.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

recalc.corStruct, recalc.reStruct, recalc.varFunc

[Package nlme version 3.1-137 Index]