recalc.varFunc {nlme}	R Documentation

Recalculate for varFunc Object

Description

This method function pre-multiples the "Xy" component of conLin by a diagonal matrix with diagonal elements given by the weights corresponding to the variance structure represented by objecte and adds the log-likelihood contribution of object, given by logLik(object), to the "logLik" component of conLin.

Usage

## S3 method for class 'varFunc':
recalc(object, conLin, ...)

Arguments

object 	 an object inheriting from class varFunc, representing a variance function structure.

conLin 	 a condensed linear model object, consisting of a list with components "Xy", corresponding to a regression matrix (X) combined with a response vector (y), and "logLik", corresponding to the log-likelihood of the underlying model.

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

the recalculated condensed linear model object.

Note

This method function is only used inside model fitting functions, such as lme and gls, that allow heteroscedastic error terms.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

recalc, varWeights, logLik.varFunc

[Package nlme version 3.1-93 Index]