residuals.gls {nlme}	R Documentation

Extract gls Residuals

Description

The residuals for the linear model represented by object are extracted.

Usage

## S3 method for class 'gls':
residuals(object, type, ...)

Arguments

object 	 an object inheriting from class gls, representing a generalized least squares fitted linear model, or from class gnls, representing a generalized nonlinear least squares fitted linear model.

type 	 an optional character string specifying the type of residuals to be used. If "response", the "raw" residuals (observed - fitted) are used; else, if "pearson", the standardized residuals (raw residuals divided by the corresponding standard errors) are used; else, if "normalized", the normalized residuals (standardized residuals pre-multiplied by the inverse square-root factor of the estimated error correlation matrix) are used. Partial matching of arguments is used, so only the first character needs to be provided. Defaults to "pearson".

... 	 some methods for this generic function require additional arguments. None are used in this method.

Value

a vector with the residuals for the linear model represented by object.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

gls

Examples

fm1 <- gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), Ovary,
           correlation = corAR1(form = ~ 1 | Mare))
residuals(fm1)

[Package nlme version 3.1-93 Index]