residuals.glsStruct {nlme}	R Documentation

Calculate glsStruct Residuals

Description

The residuals for the linear model represented by object are extracted.

Usage

## S3 method for class 'glsStruct':
residuals(object, glsFit, ...)

Arguments

object 	 an object inheriting from class glsStruct, representing a list of linear model components, such as corStruct and varFunc objects.

glsFit 	 an optional list with components logLik (log-likelihood), beta (coefficients), sigma (standard deviation for error term), varBeta (coefficients' covariance matrix), fitted (fitted values), and residuals (residuals). Defaults to attr(object, "glsFit").

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

a vector with the residuals for the linear model represented by object.

Note

This method function is primarily used inside gls and residuals.gls.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

gls, glsStruct, residuals.gls, fitted.glsStruct

[Package nlme version 3.1-93 Index]