residuals.lmList {nlme}	R Documentation

Extract lmList Residuals

Description

The residuals are extracted from each lm component of object and arranged into a list with as many components as object, or combined into a single vector.

Usage


## S3 method for class 'lmList'
residuals(object, type, subset, asList, ...)

Arguments

object
an object inheriting from class "lmList", representing a list of lm objects with a common model.

subset
an optional character or integer vector naming the lm components of object from which the residuals are to be extracted. Default is NULL, in which case all components are used.

type
an optional character string specifying the type of residuals to be extracted. Options include "response" for the "raw" residuals (observed - fitted), "pearson" for the standardized residuals (raw residuals divided by the estimated residual standard error) using different standard errors for each lm fit, and "pooled.pearson" for the standardized residuals using a pooled estimate of the residual standard error. Partial matching of arguments is used, so only the first character needs to be provided. Defaults to "response".

asList
an optional logical value. If TRUE, the returned object is a list with the residuals split by groups; else the returned value is a vector. Defaults to FALSE.

...
some methods for this generic require additional arguments. None are used in this method.

Value

a list with components given by the residuals of each lm component of object, or a vector with the residuals for all lm components of object.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

lmList, fitted.lmList

Examples


fm1 <- lmList(distance ~ age | Subject, Orthodont)
residuals(fm1)

[Package nlme version 3.1-137 Index]