solve.pdMat {nlme}	R Documentation

Calculate Inverse of a Positive-Definite Matrix

Description

The positive-definite matrix represented by a is inverted and assigned to a.

Usage


## S3 method for class 'pdMat'
solve(a, b, ...)

Arguments

a
an object inheriting from class "pdMat", representing a positive definite matrix.

b
this argument is only included for consistency with the generic function and is not used in this method function.

...
some methods for this generic require additional arguments. None are used in this method.

Value

a pdMat object similar to a, but with coefficients corresponding to the inverse of the positive-definite matrix represented by a.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

pdMat

Examples


pd1 <- pdCompSymm(3 * diag(3) + 1)
solve(pd1)

[Package nlme version 3.1-137 Index]