solve.reStruct {nlme}	R Documentation

Apply Solve to an reStruct Object

Description

Solve is applied to each pdMat component of a, which results in inverting the positive-definite matrices they represent.

Usage


## S3 method for class 'reStruct'
solve(a, b, ...)

Arguments

a
an object inheriting from class "reStruct", representing a random effects structure and consisting of a list of pdMat objects.

b
this argument is only included for consistency with the generic function and is not used in this method function.

...
some methods for this generic require additional arguments. None are used in this method.

Value

an reStruct object similar to a, but with the pdMat components representing the inverses of the matrices represented by the components of a.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

solve.pdMat, reStruct

Examples


rs1 <- reStruct(list(A = pdSymm(diag(1:3), form = ~Score),
  B = pdDiag(2 * diag(4), form = ~Educ)))
solve(rs1)

[Package nlme version 3.1-137 Index]