summary.corStruct {nlme}	R Documentation

Summarize a corStruct Object

Description

This method function prepares object to be printed using the print.summary method, by changing its class and adding a structName attribute to it.

Usage

## S3 method for class 'corStruct':
summary(object, structName, ...)

Arguments

object 	 an object inheriting from class corStruct, representing a correlation structure.

structName 	 an optional character string defining the type of correlation structure associated with object, to be used in the print.summary method. Defaults to class(object)[1].

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

an object identical to object, but with its class changed to summary.corStruct and an additional attribute structName. The returned value inherits from the same classes as object.

Author(s)

Jose Pinheiro and Douglas Bates

See Also

corClasses, corNatural, Initialize.corStruct, summary

Examples

cs1 <- corAR1(0.2)
summary(cs1)

[Package nlme version 3.1-93 Index]