summary.gls {nlme}	R Documentation

Summarize a gls Object

Description

Additional information about the linear model fit represented by object is extracted and included as components of object.

Usage

## S3 method for class 'gls'
summary(object, verbose, ...)

Arguments

object
an object inheriting from class "gls", representing a generalized least squares fitted linear model.

verbose
an optional logical value used to control the amount of output when the object is printed. Defaults to FALSE.

...
some methods for this generic require additional arguments. None are used in this method.

Value

an object inheriting from class summary.gls with all components included in object (see glsObject for a full description of the components) plus the following components:

corBeta
approximate correlation matrix for the coefficients estimates

tTable
a data frame with columns Value, Std. Error, t-value, and p-value representing respectively the coefficients estimates, their approximate standard errors, the ratios between the estimates and their standard errors, and the associated p-value under a t approximation. Rows correspond to the different coefficients.

residuals
if more than five observations are used in the gls fit, a vector with the minimum, first quartile, median, third quartile, and maximum of the residuals distribution; else the residuals.

AIC
the Akaike Information Criterion corresponding to object.

BIC
the Bayesian Information Criterion corresponding to object.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

AIC, BIC, gls, summary

Examples

fm1 <- gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), Ovary,
           correlation = corAR1(form = ~ 1 | Mare))
summary(fm1)

[Package nlme version 3.1-127 Index]