summary.lmList {nlme}	R Documentation

Summarize an lmList Object

Description

The summary.lm method is applied to each lm component of object to produce summary information on the individual fits, which is organized into a list of summary statistics. The returned object is suitable for printing with the print.summary.lmList method.

Usage


## S3 method for class 'lmList'
summary(object, pool, ...)

Arguments

object
an object inheriting from class "lmList", representing a list of lm fitted objects.

pool
an optional logical value indicating whether a pooled estimate of the residual standard error should be used. Default is attr(object, "pool").

...
some methods for this generic require additional arguments. None are used in this method.

Value

a list with summary statistics obtained by applying summary.lm to the elements of object, inheriting from class summary.lmList. The components of value are:

call
a list containing an image of the lmList call that produced object.

coefficients
a three dimensional array with summary information on the lm coefficients. The first dimension corresponds to the names of the object components, the second dimension is given by "Value", "Std. Error", "t value", and "Pr(>|t|)", corresponding, respectively, to the coefficient estimates and their associated standard errors, t-values, and p-values. The third dimension is given by the coefficients names.

correlation
a three dimensional array with the correlations between the individual lm coefficient estimates. The first dimension corresponds to the names of the object components. The third dimension is given by the coefficients names. For each coefficient, the rows of the associated array give the correlations between that coefficient and the remaining coefficients, by lm component.

cov.unscaled
a three dimensional array with the unscaled variances/covariances for the individual lm coefficient estimates (giving the estimated variance/covariance for the coefficients, when multiplied by the estimated residual errors). The first dimension corresponds to the names of the object components. The third dimension is given by the coefficients names. For each coefficient, the rows of the associated array give the unscaled covariances between that coefficient and the remaining coefficients, by lm component.

df
an array with the number of degrees of freedom for the model and for residuals, for each lm component.

df.residual
the total number of degrees of freedom for residuals, corresponding to the sum of residuals df of all lm components.

fstatistics
an array with the F test statistics and corresponding degrees of freedom, for each lm component.

pool
the value of the pool argument to the function.

r.squared
a vector with the multiple R-squared statistics for each lm component.

residuals
a list with components given by the residuals from individual lm fits.

RSE
the pooled estimate of the residual standard error.

sigma
a vector with the residual standard error estimates for the individual lm fits.

terms
the terms object used in fitting the individual lm components.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

lmList, summary

Examples


fm1 <- lmList(distance ~ age | Subject, Orthodont)
summary(fm1)

[Package nlme version 3.1-137 Index]