summary.lme {nlme}	R Documentation

Summarize an lme Object

Description

Additional information about the linear mixed-effects fit represented by object is extracted and included as components of object. The returned object is suitable for printing with the print.summary.lme method.

Usage

## S3 method for class 'lme':
summary(object, adjustSigma, verbose, ...)

Arguments

object 	 an object inheriting from class lme, representing a fitted linear mixed-effects model.

adjustSigma 	 an optional logical value. If TRUE and the estimation method used to obtain object was maximum likelihood, the residual standard error is multiplied by sqrt(nobs/(nobs - npar)), converting it to a REML-like estimate. This argument is only used when a single fitted object is passed to the function. Default is TRUE.

verbose 	 an optional logical value used to control the amount of output in the print.summary.lme method. Defaults to FALSE.

... 	 some methods for this generic require additional arguments. None are used in this method.

Value

an object inheriting from class summary.lme with all components included in object (see lmeObject for a full description of the components) plus the following components:

corFixed 	 approximate correlation matrix for the fixed effects estimates

tTable 	 a data frame with columns Value, Std. Error, DF, t-value, and p-value representing respectively the fixed effects estimates, their approximate standard errors, the denominator degrees of freedom, the ratios between the estimates and their standard errors, and the associated p-value from a t distribution. Rows correspond to the different fixed effects.

residuals 	 if more than five observations are used in the lme fit, a vector with the minimum, first quartile, median, third quartile, and maximum of the innermost grouping level residuals distribution; else the innermost grouping level residuals.

AIC 	 the Akaike Information Criterion corresponding to object.

BIC 	 the Bayesian Information Criterion corresponding to object.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

AIC, BIC, lme, print.summary.lme

Examples

fm1 <- lme(distance ~ age, Orthodont, random = ~ age | Subject)
summary(fm1)

[Package nlme version 3.1-93 Index]