summary.nlsList {nlme}	R Documentation

Summarize an nlsList Object

Description

The summary function is applied to each nls component of object to produce summary information on the individual fits, which is organized into a list of summary statistics. The returned object is suitable for printing with the print.summary.nlsList method.

Usage


## S3 method for class 'nlsList'
summary(object, ...)

Arguments

object
an object inheriting from class "nlsList", representing a list of nls fitted objects.

...
optional arguments to the summary.lmList method. One such optional argument is pool, a logical value indicating whether a pooled estimate of the residual standard error should be used. Default is attr(object, "pool").

Value

a list with summary statistics obtained by applying summary to the elements of object, inheriting from class summary.nlsList. The components of value are:

call
a list containing an image of the nlsList call that produced object.

parameters
a three dimensional array with summary information on the nls coefficients. The first dimension corresponds to the names of the object components, the second dimension is given by "Value", "Std. Error", "t value", and "Pr(>|t|)", corresponding, respectively, to the coefficient estimates and their associated standard errors, t-values, and p-values. The third dimension is given by the coefficients names.

correlation
a three dimensional array with the correlations between the individual nls coefficient estimates. The first dimension corresponds to the names of the object components. The third dimension is given by the coefficients names. For each coefficient, the rows of the associated array give the correlations between that coefficient and the remaining coefficients, by nls component.

cov.unscaled
a three dimensional array with the unscaled variances/covariances for the individual lm coefficient estimates (giving the estimated variance/covariance for the coefficients, when multiplied by the estimated residual errors). The first dimension corresponds to the names of the object components. The third dimension is given by the coefficients names. For each coefficient, the rows of the associated array give the unscaled covariances between that coefficient and the remaining coefficients, by nls component.

df
an array with the number of degrees of freedom for the model and for residuals, for each nls component.

df.residual
the total number of degrees of freedom for residuals, corresponding to the sum of residuals df of all nls components.

pool
the value of the pool argument to the function.

RSE
the pooled estimate of the residual standard error.

sigma
a vector with the residual standard error estimates for the individual lm fits.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

nlsList, summary

Examples


fm1 <- nlsList(SSasymp, Loblolly)
summary(fm1)

[Package nlme version 3.1-137 Index]