update.modelStruct {nlme}	R Documentation

Update a modelStruct Object

Description

This method function updates each element of object, allowing the access to data.

Usage


## S3 method for class 'modelStruct'
update(object, data, ...)

Arguments

object
an object inheriting from class "modelStruct", representing a list of model components, such as corStruct and varFunc objects.

data
a data frame in which to evaluate the variables needed for updating the elements of object.

...
some methods for this generic require additional arguments. None are used in this method.

Value

an object similar to object (same class, length, and names), but with updated elements.

Note

This method function is primarily used within model fitting functions, such as lme and gls, that allow model components such as variance functions.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

reStruct

[Package nlme version 3.1-137 Index]