update.varFunc {nlme}	R Documentation

Update varFunc Object

Description

If the formula(object) includes a "." term, representing a fitted object, the variance covariate needs to be updated upon completion of an optimization cycle (in which the variance function weights are kept fixed). This method function allows a reevaluation of the variance covariate using the current fitted object and, optionally, other variables in the original data.

Usage


## S3 method for class 'varFunc'
update(object, data, ...)

Arguments

object
an object inheriting from class "varFunc", representing a variance function structure.

data
a list with a component named "." with the current version of the fitted object (from which fitted values, coefficients, and residuals can be extracted) and, if necessary, other variables used to evaluate the variance covariate(s).

...
some methods for this generic require additional arguments. None are used in this method.

Value

if formula(object) includes a "." term, an varFunc object similar to object, but with the variance covariate reevaluated at the current fitted object value; else object is returned unchanged.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

See Also

needUpdate, covariate<-.varFunc

[Package nlme version 3.1-137 Index]