varClasses {nlme}	R Documentation

Variance Function Classes

Description

Standard classes of variance function structures (varFunc) available in the nlme library. Covariates included in the variance function, denoted by variance covariates, may involve functions of the fitted model object, such as the fitted values and the residuals. Different coefficients may be assigned to the levels of a classification factor.

Value

Available standard classes ():

varExp 	 exponential of a variance covariate.

varPower 	 power of a variance covariate.

varConstPower 	 constant plus power of a variance covariate.

varIdent 	 constant variance(s), generally used to allow different variances according to the levels of a classification factor.

varFixed 	 fixed weights, determined by a variance covariate.

varComb 	 combination of variance functions.

Note

Users may define their own varFunc classes by specifying a constructor function and, at a minimum, methods for the functions coef, coef<-, and initialize. For examples of these functions, see the methods for class varPower.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

varComb, varConstPower, varExp, varFixed, varIdent, varPower, summary.varFunc

[Package nlme version 3.1-93 Index]