varFunc {nlme}	R Documentation

Variance Function Structure

Description

If object is a one-sided formula, it is used as the argument to varFixed and the resulting object is returned. Else, if object inherits from class varFunc, it is returned unchanged.

Usage

varFunc(object)

Arguments

object 	 either an one-sided formula specifying a variance covariate, or an object inheriting from class varFunc, representing a variance function structure.

Value

an object from class varFunc, representing a variance function structure.

Author(s)

Jose Pinheiro Jose.Pinheiro@pharma.novartis.com and Douglas Bates bates@stat.wisc.edu

See Also

summary.varFunc, varFixed, varWeights.varFunc, coef.varFunc

Examples

vf1 <- varFunc(~age)

[Package nlme version 3.1-93 Index]