varWeights.glsStruct {nlme}	R Documentation

Variance Weights for glsStruct Object

Description

If object includes a varStruct component, the inverse of the standard deviations of the variance function structure represented by the corresponding varFunc object are returned; else, a vector of ones of length equal to the number of observations in the data frame used to fit the associated linear model is returned.

Usage


## S3 method for class 'glsStruct'
varWeights(object)

Arguments

object
an object inheriting from class "glsStruct", representing a list of linear model components, such as corStruct and "varFunc" objects.

Value

if object includes a varStruct component, a vector with the corresponding variance weights; else, or a vector of ones.

Author(s)

José Pinheiro and Douglas Bates bates@stat.wisc.edu

References

Pinheiro, J.C., and Bates, D.M. (2000) "Mixed-Effects Models in S and S-PLUS", Springer.

See Also

varWeights

[Package nlme version 3.1-137 Index]